% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_horizon}
\alias{add_horizon}
\title{Add missing values for future dates}
\usage{
add_horizon(data, horizon, accumulate = 1L, obs_column = "confirm", by = NULL)
}
\arguments{
\item{data}{A \verb{<data.frame>} of confirmed cases (confirm) by date (date).
\code{confirm} must be numeric and \code{date} must be in date format. Optionally
this can also have a logical \code{accumulate} column which indicates whether
data should be added to the next data point. This is useful when modelling
e.g. weekly incidence data. See also the \code{\link[=fill_missing]{fill_missing()}} function which
helps add the \code{accumulate} column with the desired properties when dealing
with non-daily data. If any accumulation is done this happens after
truncation as specified by the \code{truncation} argument.}

\item{horizon}{Numeric, defaults to 7. Number of days into the future to
forecast.}

\item{accumulate}{The number of days to accumulate when generating posterior
prediction, e.g. 7 for weekly accumulated forecasts.}
}
\value{
A data.table with missing values for future dates
}
\description{
Add missing values for future dates
}
\keyword{internal}
