% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_for_truncation}
\alias{adjust_for_truncation}
\title{Adjust Case Counts for Truncation}
\usage{
adjust_for_truncation(
  cases,
  cum_freq,
  dates,
  confidence_adjustment = NULL,
  samples,
  max_upscale
)
}
\arguments{
\item{cases}{Numeric vector of cases}

\item{cum_freq}{Numeric vector of cumulative frequencies}

\item{dates}{Character vector of dates}

\item{confidence_adjustment}{Numeric vector of frequencies used to adjust confidence}

\item{samples}{Numeric, number of samples to take}

\item{max_upscale}{Numeric, maximum upscaling of cases allowed at each time point}
}
\value{
A \code{data.table} adjusted for truncation
}
\description{
Adjust Case Counts for Truncation
}
