% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{allocate_delays}
\alias{allocate_delays}
\title{Allocate Delays into Required Stan Format}
\usage{
allocate_delays(delay_var, no_delays)
}
\arguments{
\item{delay_var}{List of numeric delays}

\item{no_delays}{Numeric, number of delays}
}
\value{
A numeric array
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Allocate delays for stan. Used in \code{\link[=delay_opts]{delay_opts()}}.
}
\keyword{internal}
