% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{apply_default_tolerance}
\alias{apply_default_tolerance}
\title{Apply default tolerance to a <dist_spec> if it is unconstrained}
\usage{
apply_default_tolerance(dist, default_tolerance, tolerance_set)
}
\arguments{
\item{dist}{A <dist_spec>}

\item{default_tolerance}{Numeric; default tolerance to be used if an
unconstrained distribution is passed as \code{dist}. If \code{dist} is already
constrained by having a maximum or tolerance this is ignored.}

\item{tolerance_set}{Logical; whether the default tolerance has been set by
the user; if yes and \code{dist} is constrained a warning is issued}
}
\value{
A <dist_spec> with the default tolerance set if previously not
constrained
}
\description{
Apply default tolerance to a <dist_spec> if it is unconstrained
}
\keyword{internal}
