% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{apply_tolerance}
\alias{apply_tolerance}
\title{Applies a threshold to all nonparametric distributions in a <dist_spec>}
\usage{
apply_tolerance(x, tolerance)
}
\arguments{
\item{x}{A \verb{<dist_spec>}}

\item{tolerance}{Numeric; the desired tolerance level. Any part of the
cumulative distribution function beyond 1 minus this tolerance level is
removed.}
}
\value{
A \verb{<dist_spec>} where probability masses below the threshold level
have been removed
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated. Use \code{bound_dist()} instead.
}
\keyword{internal}
