% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{assert_param_spec}
\alias{assert_param_spec}
\title{Assert that an object is a parameter specification}
\usage{
assert_param_spec(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{An object to check.}

\item{name}{Name used to refer to \code{x} in the error message.}
}
\value{
Invisibly returns \code{x} if it is a parameter specification; otherwise
aborts.
}
\description{
Assert that an object is a parameter specification
}
\keyword{internal}
