% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{backcalc_settings}
\alias{backcalc_settings}
\title{Back Calculation Settings}
\usage{
backcalc_settings(backcalc = list())
}
\arguments{
\item{backcalc}{A list of settings to override the back calculation settings. See
\code{?backcalc_settings} for options and \code{backcalc_settings()} for current defaults.}
}
\value{
A list of back calculation settings
}
\description{
Defines a list specifying the optional arguments for the back calculation
of cases. Only used if \code{rt = NULL}. Custom settings can be supplied which override the
defaults. Used internally \code{estimate_infections}. The settings returned (all of which
are modifiable by the user) are:
\itemize{
\item \code{smoothing_window}: Numeric, defaults to 7 days. The mean smoothing window to apply
to mean shifted reports (used as a prior during back calculation). 7 days is the default
as this smooths day of the week effects but depending on the quality of the data and the
amount of information users wish to use as a prior (higher values equaling a less
informative prior).
}
}
\examples{
# default settings
backcalc_settings()
}
