% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{bound_dist}
\alias{bound_dist}
\title{Define bounds of a \verb{<dist_spec>}}
\usage{
bound_dist(x, max, tolerance)
}
\arguments{
\item{x}{A \verb{<dist_spec>}.}

\item{max}{Numeric, maximum value of the distribution. The distribution will
be truncated at this value. Default: \code{Inf}, i.e. no maximum.}

\item{tolerance}{Numeric; the desired tolerance level. Any part of the
cumulative distribution function beyond 1 minus this tolerance level is
removed.}
}
\value{
a \verb{<dist_spec>} with relevant attributes set that define its bounds
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This sets attributes for further processing
}
