% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{c.dist_spec}
\alias{c.dist_spec}
\title{Combines multiple delay distributions for further processing}
\usage{
\method{c}{dist_spec}(...)
}
\arguments{
\item{...}{The delay distributions (from calls to \code{\link[=dist_spec]{dist_spec()}}) to combine}
}
\value{
Combined delay distributions (with class \verb{<dist_spec>})
}
\description{
This combines the parameters so that they can be fed as multiple delay
distributions to \code{\link[=epinow]{epinow()}} or \code{\link[=estimate_infections]{estimate_infections()}}.
}
\author{
Sebastian Funk
}
