% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_prior_estimates}
\alias{calculate_prior_estimates}
\title{Calculate prior infections and growth}
\usage{
calculate_prior_estimates(cases, seeding_time)
}
\arguments{
\item{cases}{Numeric vector; the case counts from the input data.}

\item{seeding_time}{Integer; seeding time, usually obtained using
\code{\link[=get_seeding_time]{get_seeding_time()}}.}
}
\value{
A list containing \code{prior_infections} and \code{prior_growth}.
}
\description{
Calculates the prior infections and growth rate based on the
first week's data.
}
\keyword{internal}
