% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_combined_np_pmf_lengths}
\alias{check_combined_np_pmf_lengths}
\title{Check and warn if combined non-parametric delay distributions are longer
than the input data}
\usage{
check_combined_np_pmf_lengths(stan_args)
}
\arguments{
\item{stan_args}{List of stan argument data}
}
\value{
Called for its side effects
}
\description{
Check and warn if combined non-parametric delay distributions are longer
than the input data
}
\keyword{internal}
