% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_gp_fit}
\alias{check_gp_fit}
\title{Check the approximate GP of a fitted model}
\usage{
check_gp_fit(fit, stan_data)
}
\arguments{
\item{fit}{A fitted Stan model.}

\item{stan_data}{A list of Stan data as returned by \code{\link[=create_gp_data]{create_gp_data()}}.}
}
\value{
Invisibly, the result of \code{\link[=check_gp_lengthscale]{check_gp_lengthscale()}}.
}
\description{
Extracts the posterior lengthscale from a fitted model and passes it to
\code{\link[=check_gp_lengthscale]{check_gp_lengthscale()}}.
}
\keyword{internal}
