% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_np_delay_lengths}
\alias{check_np_delay_lengths}
\title{Check and warn if individual non-parametric delay PMFs are longer than the
input data}
\usage{
check_np_delay_lengths(stan_args, data_length)
}
\arguments{
\item{stan_args}{List of stan arguments including the data element with
delay information from \code{\link[=create_stan_delays]{create_stan_delays()}}
@param data_length Length of original data}
}
\value{
Called for its side effects
}
\description{
Check and warn if individual non-parametric delay PMFs are longer than the
input data
}
\keyword{internal}
