% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_reports_valid}
\alias{check_reports_valid}
\title{Validate data input}
\usage{
check_reports_valid(reports, for_estimate_secondary)
}
\arguments{
\item{reports}{A data frame with either:
\itemize{
\item two columns: date and confirm, if to be
used by estimate_infection() or estimate_truncation, or
\item three columns: date, primary, and secondary, if to be used by
estimate_secondary().
}}

\item{for_estimate_secondary}{Logical; whether the data is being passed by
estimate_secondary() or the other estimate_*() functions.}
}
\value{
Called for its side effects.
}
\description{
\code{check_reports_valid()} checks that the input data frame is a data frame,
and that it has the right column names and types. In particular, it checks
that the date column is a date, and that the other columns are integers. It
checks that the date column has no NA values.
}
\author{
James M. Azam
}
\keyword{internal}
