% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_single_np_pmf_lengths}
\alias{check_single_np_pmf_lengths}
\title{Check and warn if individual non-parametric PMFs are longer than the input
data}
\usage{
check_single_np_pmf_lengths(..., data)
}
\arguments{
\item{...}{Delay distributions}

\item{data}{A \verb{<data.frame>} of disease reports (confirm) by date (date).
\code{confirm} must be numeric and \code{date} must be in date format. Optionally,
\code{data} can also have a logical \code{accumulate} column which indicates whether
data should be added to the next data point. This is useful when modelling
e.g. weekly incidence data. See also the \code{\link[=fill_missing]{fill_missing()}} function which
helps add the \code{accumulate} column with the desired properties when dealing
with non-daily data. If any accumulation is done this happens after
truncation as specified by the \code{truncation} argument. If all entries
of \code{confirm} are missing (\code{NA}) the returned estimates will represent the
prior distributions.}
}
\value{
Called for its side effects
}
\description{
Check and warn if individual non-parametric PMFs are longer than the input
data
}
\keyword{internal}
