% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_sparse_pmf_tail}
\alias{check_sparse_pmf_tail}
\title{Check that PMF tail is not sparse}
\usage{
check_sparse_pmf_tail(pmf, span = 5, tol = 1e-06)
}
\arguments{
\item{pmf}{A probability mass function}

\item{span}{The number of consecutive indices in the tail to check}

\item{tol}{The tolerance for the tail values}
}
\value{
Called for its side effects.
}
\description{
Checks the tail of a PMF to see if more than \code{span} consecutive
indices in the PMF are smaller than \code{tol}. If so, a warning is thrown.
}
\keyword{internal}
