% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_stan_delay}
\alias{check_stan_delay}
\title{Validate probability distribution for passing to stan}
\usage{
check_stan_delay(dist)
}
\arguments{
\item{dist}{A \code{dist_spec} object.`}
}
\value{
Called for its side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
\code{check_stan_delay()} checks that the supplied data is a \verb{<dist_spec>},
that it is a supported distribution, and that is has a finite maximum.
}
\keyword{internal}
