% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_stan_dist}
\alias{check_stan_dist}
\title{Validate probability distribution for passing to stan}
\usage{
check_stan_dist(dist)
}
\arguments{
\item{dist}{A \code{dist_spec} object.`}
}
\value{
Called for its side effects.
}
\description{
\code{check_stan_dist()} checks that the supplied data is a \verb{<dist_spec>},
and that it can be further processed by stan.
}
\author{
Sebastian Funk
}
\keyword{internal}
