% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{check_stanr_available}
\alias{check_stanr_available}
\title{Check that the stanr backends can be used}
\usage{
check_stanr_available()
}
\value{
Invisibly returns \code{TRUE} if \code{stanr} is available.
}
\description{
Errors informatively if the \code{stanr} package is not installed.
}
\keyword{internal}
