% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_nowcasts}
\alias{clean_nowcasts}
\title{Clean Nowcasts for a Supplied Date}
\usage{
clean_nowcasts(date = NULL, nowcast_dir = ".")
}
\arguments{
\item{date}{Date object. Defaults to today's date}

\item{nowcast_dir}{Character string giving the filepath to the nowcast results directory. Defaults
to the current directory.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
This function removes nowcasts in the format produced by \code{EpiNow2} from a target
directory for the date supplied.
}
