% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clear_epinow2_cache}
\alias{clear_epinow2_cache}
\title{Clear EpiNow2 Cache}
\usage{
clear_epinow2_cache()
}
\value{
invisible
}
\description{
This function clears anything that may exist in the EpiNow2 Cache. Most
Commonly if you are using the CmdStan backend, you'll have items in your
cache.
}
\section{Warning}{

This will removed cached models and requires that \code{rappdirs} is installed
}

