% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{collapse}
\alias{collapse}
\title{Collapse nonparametric distributions in a <dist_spec>}
\usage{
collapse(x)
}
\arguments{
\item{x}{A \verb{<dist_spec>}}
}
\value{
A \verb{<dist_spec>} where consecutive nonparametric distributions
have been convolved
}
\description{
This convolves any consecutive nonparametric distributions contained
in the <dist_spec>.
}
\author{
Sebastian Funk
}
