% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{construct_output}
\alias{construct_output}
\title{Construct Output}
\usage{
construct_output(
  estimates,
  forecast = NULL,
  estimated_reported_cases,
  plots = NULL,
  summary = NULL,
  samples = TRUE
)
}
\arguments{
\item{estimates}{List of data frames as output by \code{estimate_infections}}

\item{forecast}{A list of data frames as output by \code{forecast_infections}}

\item{estimated_reported_cases}{A list of dataframes as produced by
\code{estimates_by_report_date}.}

\item{plots}{A list of plots as produced by \code{report_plots}}

\item{summary}{A list of summary output as produced by \code{report_summary}}

\item{samples}{Logical, defaults to TRUE. Should samples be saved}
}
\value{
A list of output as returned by \code{epinow}
}
\description{
Construct Output
}
