% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_logmean}
\alias{convert_to_logmean}
\title{Convert mean and sd to log mean for a log normal distribution}
\usage{
convert_to_logmean(mean, sd)
}
\arguments{
\item{mean}{Numeric, mean of a distribution}

\item{sd}{Numeric, standard deviation of a distribution}
}
\value{
The log mean of a lognormal distribution
}
\description{
Convert mean and sd to log mean for a log normal distribution
}
\examples{

convert_to_logmean(2, 1)
}
