% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_logsd}
\alias{convert_to_logsd}
\title{Convert mean and sd to log standard deviation for a log normal distribution}
\usage{
convert_to_logsd(mean, sd)
}
\arguments{
\item{mean}{Numeric, mean of a distribution}

\item{sd}{Numeric, standard deviation of a distribution}
}
\value{
The log standard deviation of a lognormal distribution
}
\description{
Convert mean and sd to log standard deviation for a log normal distribution
}
\examples{

convert_to_logsd(2, 1)
}
