% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{convert_to_natural}
\alias{convert_to_natural}
\title{Internal function for converting parameters to natural parameters.}
\usage{
convert_to_natural(x, distribution)
}
\arguments{
\item{x}{A numerical named parameter vector}

\item{distribution}{Character; the distribution to use.}
}
\value{
A list with two elements, \code{params_mean} and \code{params_sd}, containign
mean and sd of natural parameters.
}
\description{
This is used for preprocessing before generating a \code{dist_spec} object
from a given set of parametrs and distribution
}
\author{
Sebastian Funk
}
