% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_generation_times}
\alias{covid_generation_times}
\title{Covid Generation Time Estimates}
\format{
A \code{data.table} of mean and standard deviation estimates
}
\usage{
covid_generation_times
}
\description{
Default Covid generation time estimates. See
here for details: https://github.com/epiforecasts/EpiNow/blob/master/data-raw/ganyani-generation-time.R
}
\keyword{datasets}
