% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_generation_times_summary}
\alias{covid_generation_times_summary}
\title{Covid Generation Time Estimates - Summary}
\format{
A \code{data.table} of summarising the distribution
}
\usage{
covid_generation_times_summary
}
\description{
Default Covid generation time estimates (summary). See
here for details: https://github.com/epiforecasts/EpiNow/blob/master/data-raw/ganyani-generation-time.R
}
\keyword{datasets}
