% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_incubation_period}
\alias{covid_incubation_period}
\title{Covid Incubation Period}
\format{
A vector giviing the probability for each day
}
\usage{
covid_incubation_period
}
\description{
Default Covid incubation period estimates. See
here for details: https://github.com/epiforecasts/EpiNow/blob/master/data-raw/incubation-period.R
}
\keyword{datasets}
