% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_serial_intervals}
\alias{covid_serial_intervals}
\title{Covid Generation Serial Intervals}
\format{
A vector giviing the probability for each day
}
\usage{
covid_serial_intervals
}
\description{
Default Covid serial interval estimates. See
here for details: https://github.com/epiforecasts/EpiNow/blob/master/data-raw/nishiura-serial-interval.R
}
\keyword{datasets}
