% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_backcalc_data}
\alias{create_backcalc_data}
\title{Create Back Calculation Data}
\usage{
create_backcalc_data(backcalc = backcalc_opts())
}
\arguments{
\item{backcalc}{A list of options as generated by \code{backcalc_opts()} to
define the back calculation. Defaults to \code{backcalc_opts()}.}
}
\value{
A list of settings defining the Gaussian process
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Takes the output of \code{backcalc_opts()} and converts it into a list understood
by \code{stan}.
}
\examples{
create_backcalc_data(backcalc = backcalc_opts())
}
\seealso{
backcalc_opts
}
\author{
Sam Abbott
}
