% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_forecast_data}
\alias{create_forecast_data}
\title{Create forecast settings}
\usage{
create_forecast_data(forecast = forecast_opts(), data)
}
\arguments{
\item{forecast}{A list of options as generated by \code{\link[=forecast_opts]{forecast_opts()}} defining
the forecast opitions. Defaults to \code{\link[=forecast_opts]{forecast_opts()}}. If NULL then no
forecasting will be done.}

\item{data}{A \verb{<data.frame>} of confirmed cases (confirm) by date (date).
\code{confirm} must be numeric and \code{date} must be in date format. Optionally
this can also have a logical \code{accumulate} column which indicates whether
data should be added to the next data point. This is useful when modelling
e.g. weekly incidence data. See also the \code{\link[=fill_missing]{fill_missing()}} function which
helps add the \code{accumulate} column with the desired properties when dealing
with non-daily data. If any accumulation is done this happens after
truncation as specified by the \code{truncation} argument.}
}
\value{
A list of settings ready to be passed to stan defining
the Observation Model
}
\description{
Create forecast settings
}
\keyword{internal}
