% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_gp_data}
\alias{create_gp_data}
\title{Create Gaussian Process Data}
\usage{
create_gp_data(gp = gp_opts(), data)
}
\arguments{
\item{gp}{A list of options as generated by \code{\link[=gp_opts]{gp_opts()}} to define the
Gaussian process. Defaults to \code{\link[=gp_opts]{gp_opts()}}. Set to \code{NULL} to disable the
Gaussian process.}

\item{data}{A list containing the following numeric values:
\code{t}, \code{seeding_time}, \code{horizon}, \code{future_fixed}, \code{fixed_from},
\code{stationary} and \code{estimate_r}.}
}
\value{
A list of settings defining the Gaussian process
}
\description{
Takes the output of \code{\link[=gp_opts]{gp_opts()}} and converts it into a list understood by
stan.
}
\examples{
\dontrun{
# define input data required
data <- list(
  t = 30,
  seeding_time = 7,
  horizon = 7,
  future_fixed = 0,
  fixed_from = 0,
  stationary = 0,
  estimate_r = 1
)

# default gaussian process data
create_gp_data(data = data)

# settings when no gaussian process is desired
create_gp_data(NULL, data)

# custom lengthscale
create_gp_data(gp_opts(ls = LogNormal(mean = 14, sd = 7)), data)
}
}
\seealso{
\code{\link[=gp_opts]{gp_opts()}}
}
\keyword{internal}
