% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{create_logging_sampler_values}
\alias{create_logging_sampler_values}
\title{Calculate number of post-warmup iterations and samples based on the backend}
\usage{
create_logging_sampler_values(args)
}
\arguments{
\item{args}{List of stan arguments.}
}
\description{
Calculate number of post-warmup iterations and samples based on the backend
}
\keyword{internal}
