% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_obs_model}
\alias{create_obs_model}
\title{Create Observation Model Settings}
\usage{
create_obs_model(obs_model = list())
}
\arguments{
\item{obs_model}{A list of settings to override the package default observation
model settings. See the documentation for \code{obs_model_settings} for details of these
settings and \code{obs_model_settings()} for the current defaults.}
}
\value{
A list of settings defining the Observation Model
}
\description{
Create Observation Model Settings
}
\examples{
# default observation model data
create_obs_model()

# Poisson observation model
create_obs_model(gp = NULL, data)

# Applying a observation scaling to the data
create_obs_model(obs_model = list(scale = list(mean = 0.4, sd = 0.01)))
}
\seealso{
gp_settings
}
