% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_rt_data}
\alias{create_rt_data}
\title{Create Time-varying Reproduction Number Data}
\usage{
create_rt_data(rt = rt_opts(), delay = 0, horizon = 0, data = NULL)
}
\arguments{
\item{rt}{A list of options as generated by \code{\link[=rt_opts]{rt_opts()}} defining Rt
estimation. Defaults to \code{\link[=rt_opts]{rt_opts()}}. To generate new infections using
the non-mechanistic model instead of the renewal equation model, use
\code{rt = NULL}. The non-mechanistic model internally uses the setting
\code{rt = rt_opts(use_rt = FALSE)}.}

\item{delay}{Numeric mean delay}

\item{horizon}{Numeric, forecast horizon.}

\item{data}{A \code{data.table} of case data (optional). Used for validation
checks.}
}
\value{
A list of settings defining the time-varying reproduction number
}
\description{
Takes the output from \code{\link[=rt_opts]{rt_opts()}} and converts it into a list understood by
stan.
}
\examples{
\dontrun{
# default Rt data
create_rt_data()

# settings when no Rt is desired
create_rt_data(rt = NULL)
}
}
\seealso{
\code{\link[=rt_opts]{rt_opts()}}
}
\keyword{internal}
