% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_shifted_cases}
\alias{create_shifted_cases}
\title{Create Delay Shifted Cases}
\usage{
create_shifted_cases(data, shift, smoothing_window, horizon)
}
\arguments{
\item{data}{A \verb{<data.frame>} of disease reports (confirm) by date (date).
\code{confirm} must be non-negative numeric and \code{date} must be in date format
with no missing (\code{NA}) entries; data that does not meet these requirements
is rejected with an informative error before model fitting starts.
Optionally, \code{data} can also have a logical \code{accumulate} column which
indicates whether data should be added to the next data point. This is
useful when modelling e.g. weekly incidence data. See also the
\code{\link[=fill_missing]{fill_missing()}} function which helps add the \code{accumulate} column with the
desired properties when dealing with non-daily data. If any accumulation is
done this happens after truncation as specified by the \code{truncation}
argument. If all entries of \code{confirm} are missing (\code{NA}) the returned
estimates will represent the prior distributions.}

\item{shift}{Numeric, mean delay shift to apply.}

\item{smoothing_window}{Numeric, the rolling average smoothing window
to apply. Must be odd in order to be defined as a centred average.}
}
\value{
A \verb{<data.frame>} for shifted reported cases
}
\description{
This functions creates a data frame of reported cases that has been smoothed
using a centred partial rolling average (with a period set by
\code{smoothing_window}) and shifted back in time by some delay. It is used by
\code{\link[=estimate_infections]{estimate_infections()}} to generate the mean shifted prior on which the back
calculation method (see \code{\link[=backcalc_opts]{backcalc_opts()}}) is based.
}
\details{
The function first shifts all the data back in time by \code{shift} days (thus
discarding the first \code{shift} days of data) and then applies a centred
rolling mean of length \code{smoothing_window} to the shifted data except for
the final period. The final period (the forecast horizon plus half the
smoothing window) is instead replaced by a log-linear model fit (with 1
added to the data for fitting to avoid zeroes and later subtracted again),
projected to the end of the forecast horizon. The initial part of the data
(corresponding to the length of the smoothing window) is then removed, and
any non-integer resulting values rounded up.
}
\examples{
\dontrun{
shift <- 7
horizon <- 7
smoothing_window <- 14
## add NAs for horizon
cases <- add_horizon(example_confirmed[1:30], horizon)
## add zeroes initially
cases <- data.table::rbindlist(list(
  data.table::data.table(
    date = seq(
      min(cases$date) - 10,
      min(cases$date) - 1,
      by = "days"
    ),
    confirm = 0, breakpoint = 0
  ),
  cases
))
create_shifted_cases(cases, shift, smoothing_window, horizon)
}
}
\keyword{internal}
