% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_standardised_stan_args}
\alias{create_standardised_stan_args}
\title{Create Standardizsed Stan Arguments}
\usage{
create_standardised_stan_args(stan_args_in)
}
\arguments{
\item{stan_args_in}{a named list. Arguments}
}
\value{
a list of standardized values
}
\description{
This function handles some of the additional parsing of the \code{stan_args}
argument to set features like the backend used, algorithm applied, and
and previously fit models (or new models)
}
\keyword{internal}
