% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{deconvolution_opts}
\alias{deconvolution_opts}
\title{Deconvolution Model Specification}
\usage{
deconvolution_opts(
  prior = c("reports", "none", "infections"),
  prior_window = 14,
  rt_window = 1
)
}
\arguments{
\item{prior}{A character string defaulting to "reports". Defines the prior
to use when deconvolving. Currently implemented options are to use smoothed
mean delay shifted reported cases ("reports"), to use the estimated
infections from the previous time step seeded for the first time step using
mean shifted reported cases ("infections"), or no prior ("none"). Using no
prior will result in poor real time performance. No prior and using
infections are only supported when a Gaussian process is present. If
observed data is not reliable then a sensible first step is to explore
increasing the \code{prior_window} with a sensible second step being to no longer
use reported cases as a prior (i.e. set \code{prior = "none"}).}

\item{prior_window}{Integer, defaults to 14 days. The mean centred smoothing
window to apply to mean shifted reports (used as a prior during
deconvolution). 7 days is minimum recommended setting as this smooths day of
the week effects but depending on the quality of the data and the amount of
information users wish to use as a prior (higher values equalling a less
informative prior).}

\item{rt_window}{Integer, defaults to 1. The size of the centred rolling
average to use when estimating Rt. This must be odd so that the central
estimate is included.}
}
\value{
A \verb{<deconvolution_opts>} object that can be passed to the \code{model}
argument of \code{\link[=estimate_infections]{estimate_infections()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Specifies that infections should be estimated using the deconvolution
(non-mechanistic) model. This model directly estimates infections on the log
scale without using the renewal equation, and therefore does not require a
generation time distribution.
}
\examples{
# default settings
deconvolution_opts()

# with custom prior window
deconvolution_opts(prior_window = 21)

# with no prior (requires Gaussian process)
deconvolution_opts(prior = "none")
}
\seealso{
\code{\link[=renewal_opts]{renewal_opts()}} \code{\link[=estimate_infections]{estimate_infections()}}
}
