% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{default_fill_missing_obs}
\alias{default_fill_missing_obs}
\title{Temporary function to support the transition to full support of missing
data.}
\usage{
default_fill_missing_obs(data, obs, obs_column)
}
\arguments{
\item{data}{Data frame with a \code{date} column. The other columns depend on the
model that the data are to be used, e.g. \code{\link[=estimate_infections]{estimate_infections()}} or
\code{\link[=estimate_secondary]{estimate_secondary()}}. See the documentation there for the expected
format. The data must not already have an \code{accumulate} function, otherwise
the function will fail with an error.}

\item{obs}{A list of options as generated by \code{\link[=obs_opts]{obs_opts()}} defining the
observation model. Defaults to \code{\link[=obs_opts]{obs_opts()}}.}

\item{obs_column}{Character (default: "confirm"). If given, only the column
specified here will be used for checking missingness. This is useful if
using a data set that has multiple columns of hwich one of them
corresponds to observations that are to be processed here.}
}
\value{
data set with missing dates filled in as na values
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
