% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{delay_dist}
\alias{delay_dist}
\title{Delay distribution.}
\usage{
delay_dist(
  mean,
  sd = 0,
  mean_sd = 0,
  sd_sd = 0,
  dist = c("lognormal", "gamma"),
  max = NULL,
  fixed = FALSE
)
}
\arguments{
\item{mean}{Numeric. If the only non-zero summary parameter
then this is the fixed interval of the delay distribugion. If the \code{sd} is
non-zero then this is the mean of the distribution given by \code{dist}.
If this is not given a vector of empty vectors is returned.}

\item{sd}{Numeric, defaults to 0. Sets the standard deviation of the delay
distribution.}

\item{mean_sd}{Numeric, defaults to 0. Sets the standard deviation of the
uncertainty around the mean of the delay distribution.}

\item{sd_sd}{Numeric, defaults to 0. Sets the standard deviation of the
uncertainty around the sd of the delay distribution.}

\item{dist}{Character, defaults to "lognormal". The (discretised) distribution
to be used. If sd == 0 then the delay is fixed and a delta function will be
used whatever the choice here.}

\item{max}{Numeric, maximum value of the delay distribution}

\item{fixed}{Logical, defaults to \code{FALSE}. Should delays be treated
as coming from fixed (vs uncertain) distributions. Making this simplification
drastically reduces compute requirements.}
}
\value{
A list of delay distribution options to be used downstream
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Defines the parameters of a delay distribution
}
\author{
Sebastian Funk
}
