% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{dist_id_to_name}
\alias{dist_id_to_name}
\title{Get distribution name from primarycensored Stan dist_id}
\usage{
dist_id_to_name(dist_id)
}
\arguments{
\item{dist_id}{Integer Stan distribution ID from primarycensored
(1 = lognormal, 2 = gamma, 3 = weibull, 4 = exponential).}
}
\value{
A character string distribution name.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Maps a primarycensored Stan distribution ID back to an EpiNow2
distribution name.
}
\keyword{internal}
