% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{distspec-reexports}
\alias{distspec-reexports}
\alias{Gamma}
\alias{LogNormal}
\alias{Normal}
\alias{Fixed}
\alias{Exp}
\alias{Weibull}
\alias{Dirichlet}
\alias{NonParametric}
\alias{discretise}
\alias{discretize}
\alias{get_pmf}
\alias{convert_to_logmean}
\alias{convert_to_logsd}
\alias{fix_parameters}
\alias{get_distribution}
\alias{get_parameters}
\alias{is_constrained}
\alias{bound_dist}
\alias{collapse}
\alias{new_dist_spec}
\title{Distribution functions re-exported from distspec}
\usage{
Gamma(...)

LogNormal(...)

Normal(...)

Fixed(...)

Exp(...)

Weibull(...)

Dirichlet(...)

NonParametric(...)

discretise(...)

discretize(...)

get_pmf(...)

convert_to_logmean(...)

convert_to_logsd(...)

fix_parameters(...)

get_distribution(...)

get_parameters(...)

is_constrained(...)

bound_dist(...)

collapse(...)

new_dist_spec(...)
}
\arguments{
\item{...}{Passed to the \pkg{distspec} function of the same name.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
The probability distribution interface has moved to the \pkg{distspec}
package. These functions are re-exported so that the \verb{EpiNow2::} form keeps
resolving, and will be removed in a future release; use \pkg{distspec}
directly (the bare names are attached when EpiNow2 is loaded).
}
\keyword{internal}
