% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{.dist_spec}
\alias{.dist_spec}
\title{Specify a distribution.}
\usage{
.dist_spec(
  distribution = c("lognormal", "normal", "gamma", "fixed"),
  params_mean = numeric(0),
  params_sd = numeric(0),
  max = Inf,
  pmf = numeric(0)
)
}
\arguments{
\item{distribution}{Character, defaults to "lognormal". The (discretised)
distribution to be used. Can be "lognormal", "gamma", "normal" or "fixed".
The corresponding parameters (defined in \code{\link[=natural_params]{natural_params()}}) are passed
as \code{params_mean},  and their uncertainty as \code{params_sd}.}

\item{params_mean}{Numeric. Central values of the parameters of the
distribution as defined in [natural_params().}

\item{params_sd}{Numeric. Standard deviations of the parameters of the
distribution as defined in [natural_params().}

\item{max}{Numeric, maximum value of the distribution. The distribution will
be truncated at this value. Default: \code{Inf}, i.e. no maximum.}

\item{pmf}{Numeric, a vector of values that represent the (nonparametric)
probability mass function of the delay (starting with 0); defaults to an
empty vector corresponding to a parametric specification of the distribution
(using \code{params_mean}, and \code{params_sd}.}
}
\value{
A \code{dist_spec} object defining the distribution.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Defines the parameters of a supported distribution for use in onward
modelling. Multiple distribution families are supported - see the
documentation for \code{family} for details. Alternatively, a nonparametric
distribution can be specified using the \code{pmf} argument.
This function provides distribution
functionality in \code{\link[=delay_opts]{delay_opts()}}, \code{\link[=generation_time_opts]{generation_time_opts()}}, and
\code{\link[=trunc_opts]{trunc_opts()}}.
}
\examples{
\dontrun{
# A fixed lognormal distribution with mean 5 and sd 1.
.dist_spec(
  params_mean = c(meanlog = 5, sdlog = 1), max = 20,
  distribution = "lognormal"
)

# An uncertain gamma distribution with mean 3 and sd 2
.dist_spec(
  params_mean = c(alpha = 3, beta = 0.5),
  params_sd = c(alpha = 0.75, beta = 0.25),
  distribution = "gamma"
)
}
}
\author{
Sebastian Funk

Sam Abbott
}
\keyword{internal}
