% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist_prototype.R
\name{.extract_dist_spec}
\alias{.extract_dist_spec}
\title{Extract dist_spec from primarycensored fit (internal)}
\usage{
.extract_dist_spec(fit, dist, max_value)
}
\arguments{
\item{fit}{CmdStanMCMC object from primarycensored}

\item{dist}{Distribution type}

\item{max_value}{Maximum value for dist_spec}
}
\value{
A dist_spec object
}
\description{
Extract dist_spec from primarycensored fit (internal)
}
\keyword{internal}
