% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist.R
\name{.fit_with_cmdstanr_pcd}
\alias{.fit_with_cmdstanr_pcd}
\title{Fit with cmdstanr}
\usage{
.fit_with_cmdstanr_pcd(stan_data, samples, chains, cores, verbose, ...)
}
\description{
Fit with cmdstanr
}
\keyword{internal}
