% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist.R
\name{.fit_with_rstan_pcd}
\alias{.fit_with_rstan_pcd}
\title{Fit with rstan using pre-compiled estimate_dist model}
\usage{
.fit_with_rstan_pcd(stan_data, samples, chains, cores, verbose, ...)
}
\description{
Fit with rstan using pre-compiled estimate_dist model
}
\keyword{internal}
