% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist.R
\name{.get_dist_id}
\alias{.get_dist_id}
\title{Map distribution name to primarycensored ID}
\usage{
.get_dist_id(dist)
}
\arguments{
\item{dist}{Character distribution name}
}
\value{
Integer distribution ID
}
\description{
Map distribution name to primarycensored ID
}
\keyword{internal}
