% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist.R
\name{.get_param_names}
\alias{.get_param_names}
\title{Map distribution name to parameter names}
\usage{
.get_param_names(dist)
}
\arguments{
\item{dist}{Character distribution name}
}
\value{
Character vector of parameter names
}
\description{
Map distribution name to parameter names
}
\keyword{internal}
