% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist_mle.R
\name{.get_start_params}
\alias{.get_start_params}
\title{Get starting parameter values for MLE}
\usage{
.get_start_params(delay_data, dist)
}
\arguments{
\item{delay_data}{Data frame with delay intervals}

\item{dist}{Distribution name (fitdistrplus format)}
}
\value{
Named list of starting parameters
}
\description{
Get starting parameter values for MLE
}
\keyword{internal}
