% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist_prototype.R
\name{.linelist_to_pcd_data}
\alias{.linelist_to_pcd_data}
\title{Convert linelist to primarycensored data (internal)}
\usage{
.linelist_to_pcd_data(data, dist_id, primary_id, verbose)
}
\arguments{
\item{data}{Input data (vector or data.frame)}

\item{dist_id}{Distribution ID for Stan}

\item{primary_id}{Primary distribution ID for Stan}

\item{verbose}{Verbose output?}
}
\value{
List for primarycensored
}
\description{
Convert linelist to primarycensored data (internal)
}
\keyword{internal}
