% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist.R
\name{.numeric_to_linelist}
\alias{.numeric_to_linelist}
\title{Convert numeric delay vector to linelist format}
\usage{
.numeric_to_linelist(data, verbose = FALSE)
}
\arguments{
\item{data}{Numeric vector of delay values}

\item{verbose}{Logical, print progress messages?}
}
\value{
A data frame with columns: \code{delay_lwr}, \code{delay_upr},
\code{pwindow}, \code{relative_obs_time}, \code{n}.
}
\description{
Convert numeric delay vector to linelist format
}
\keyword{internal}
