% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist_mle.R
\name{.prepare_delay_data_for_mle}
\alias{.prepare_delay_data_for_mle}
\title{Prepare delay data for MLE fitting}
\usage{
.prepare_delay_data_for_mle(data, verbose = FALSE)
}
\arguments{
\item{data}{Input data (vector or data frame)}

\item{verbose}{Logical, print messages?}
}
\value{
Data frame with delay_lower, delay_upper, n
}
\description{
Prepare delay data for MLE fitting
}
\keyword{internal}
