% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist.R
\name{.prepare_delay_intervals}
\alias{.prepare_delay_intervals}
\title{Prepare delay data as intervals}
\usage{
.prepare_delay_intervals(data, verbose = FALSE)
}
\description{
Prepare delay data as intervals
}
\keyword{internal}
