% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist_prototype.R
\name{.vector_to_pcd_data}
\alias{.vector_to_pcd_data}
\title{Convert vector to primarycensored data (internal)}
\usage{
.vector_to_pcd_data(values, dist_id, primary_id, verbose)
}
\arguments{
\item{dist_id}{Distribution ID for Stan}

\item{primary_id}{Primary distribution ID for Stan}

\item{verbose}{Verbose output?}
}
\value{
List for primarycensored
}
\description{
Convert vector to primarycensored data (internal)
}
\keyword{internal}
