% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_measures_pipeline.R
\name{epi_measures_pipeline}
\alias{epi_measures_pipeline}
\title{Estimate time-varying measures and forecast}
\usage{
epi_measures_pipeline(
  nowcast = NULL,
  generation_times = NULL,
  min_est_date = NULL,
  gt_samples = 1,
  rt_samples = 5,
  rt_windows = 7,
  rate_window = 7,
  rt_prior = NULL,
  forecast_model = NULL,
  horizon = 0,
  verbose = TRUE
)
}
\arguments{
\item{nowcast}{A nowcast as produced by \code{nowcast_pipeline}}

\item{generation_times}{A matrix with columns representing samples and rows representing the probability of the generation timebeing on
that day.}

\item{min_est_date}{Date to begin estimation.}

\item{gt_samples}{Numeric, the number of samples to take from the generaiton times supplied}

\item{rt_samples}{Numeric, the number of samples to take from the estimated R distribution for each time point.}

\item{rt_windows}{Numeric vector, windows over which to estimate time-varying R. The best performing window will be
selected per serial interval sample by default (based on which window best forecasts current cases).}

\item{rate_window}{Numeric, the window to use to estimate the rate of spread.}

\item{rt_prior}{A list defining the reproduction number prior containing the mean (\code{mean_prior}) and standard deviation (\code{std_prior})}

\item{forecast_model}{An uninitialised bsts model passed to \code{EpiSoon::forecast_rt} to be used for forecasting
future Rt values. An example of the required structure is: \code{function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}}.}

\item{horizon}{Numeric, defaults to 0. The horizon over which to forecast Rts and cases.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should progress messages be shown.}
}
\value{
A list of data frames containing reproduction number estimates, case forecasts, rate of growth estimates
both summarised and raw.
}
\description{
Estimate time-varying measures and forecast
}
\examples{
## Construct example distributions
## reporting delay dist
delay_dist <- EpiNow::lognorm_dist_def(mean = 3, 
                                      mean_sd = 1,
                                      sd = 3,
                                      sd_sd = 1,
                                      max_value = 30,
                                      samples = 1)

## incubation delay dist
incubation_dist <- delay_dist

## Uses example case vector from EpiSoon
cases <- data.table::setDT(EpiSoon::example_obs_cases)
cases <- cases[, `:=`(confirm = as.integer(cases), import_status = "local")]

## Basic nowcast
nowcast <- nowcast_pipeline(reported_cases = cases, 
                            target_date = max(cases$date),
                            delay_defs = delay_dist,
                            incubation_defs = incubation_dist)
 
## Estimate parameters                           
estimates <- epi_measures_pipeline(nowcast[type \%in\% "infection_upscaled"], 
                       generation_times = EpiNow::covid_generation_times,
                       rt_prior = list(mean_prior = 2.6, std_prior = 2))   
                       
estimates                                                             
}
