% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{epinow2_rstan_model}
\alias{epinow2_rstan_model}
\title{Load an EpiNow2 rstan model.}
\usage{
epinow2_rstan_model(model = "estimate_infections")
}
\arguments{
\item{model}{A character string indicating the model to use. Needs to be
amongst the compiled models shipped with "EpiNow2" (see the \code{stan} directory
for a list). Defaults to "estimate_infections".}
}
\value{
An \code{rstan} model.
}
\description{
The models are pre-compiled upon package install and is returned here.
}
\keyword{internal}
