% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{==.dist_spec}
\alias{==.dist_spec}
\alias{!=.dist_spec}
\title{Compares two delay distributions}
\usage{
\method{==}{dist_spec}(e1, e2)

\method{!=}{dist_spec}(e1, e2)
}
\arguments{
\item{e1}{The first delay distribution (of type <dist_spec>) to
combine.}

\item{e2}{The second delay distribution (of type <dist_spec>) to
combine.}
}
\value{
TRUE or FALSE
}
\description{
Compares two delay distributions
}
\examples{
Fixed(1) == Normal(1, 0.5)
}
