% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_littler.R
\name{estimate_doubling_time}
\alias{estimate_doubling_time}
\title{Estimate the doubling time}
\usage{
estimate_doubling_time(r)
}
\arguments{
\item{r}{An estimate of the rate of change (r)}
}
\value{
A vector of numeric values
}
\description{
Estimate the doubling time
}
\examples{

estimate_doubling_time(0.2)
}
