% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{estimate_early_dynamics}
\alias{estimate_early_dynamics}
\title{Calculate prior infections and fit early growth}
\usage{
estimate_early_dynamics(cases, seeding_time)
}
\arguments{
\item{cases}{Numeric vector; the case counts from the input data.}

\item{seeding_time}{Integer; seeding time, usually obtained using
\code{\link[=get_seeding_time]{get_seeding_time()}}.}
}
\value{
A list containing \code{initial_infections_estimate} and
\code{initial_growth_estimate}.
}
\description{
Calculates the prior infections and growth rate based on the
first week's data.
}
\keyword{internal}
