% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_littler.R
\name{estimate_little_r}
\alias{estimate_little_r}
\title{Estimate r}
\usage{
estimate_little_r(sample, min_time = NULL, max_time = NULL)
}
\arguments{
\item{sample}{A datatable containing a numeric cases variable.}

\item{min_time}{Numeric, minimum time to use to fit the model.}

\item{max_time}{Numeric, maximum time to use to fit the model.}
}
\value{
A datatable containing an estimate of r, its standard deviation and a
measure of the goodness of fit.
}
\description{
Estimate r
}
\examples{

cases <- data.table::setDT(EpiSoon::example_obs_cases)[, 
                          cases := as.integer(cases)]

estimate_little_r(cases)
}
