% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R0.R
\name{estimate_R0}
\alias{estimate_R0}
\title{Estimate the time varying R0 - using EpiEstim}
\usage{
estimate_R0(
  cases = NULL,
  generation_times = NULL,
  rt_prior = NULL,
  windows = NULL,
  gt_samples = 100,
  rt_samples = 100,
  min_est_date = NULL,
  forecast_model = NULL,
  horizon = 0
)
}
\arguments{
\item{cases}{A dataframe containing a list of local cases with the following variables: \code{date}, \code{cases}, and \code{import_status}}

\item{generation_times}{A matrix with columns representing samples and rows representing the probability of the generation timebeing on
that day.}

\item{rt_prior}{A list defining the reproduction number prior containing the mean (\code{mean_prior}) and standard deviation (\code{std_prior})}

\item{windows}{Numeric vector, windows over which to estimate time-varying R. The best performing window will be
selected per serial interval sample by default (based on which window best forecasts current cases).}

\item{gt_samples}{Numeric, the number of samples to take from the generaiton times supplied}

\item{rt_samples}{Numeric, the number of samples to take from the estimated R distribution for each time point.}

\item{min_est_date}{Date to begin estimation.}

\item{forecast_model}{An uninitialised bsts model passed to \code{EpiSoon::forecast_rt} to be used for forecasting
future Rt values. An example of the required structure is: \code{function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}}.}

\item{horizon}{Numeric, defaults to 0. The horizon over which to forecast Rts and cases.}
}
\value{
A list of \code{data.table}'s containing the date and summarised R estimate and optionally a case forecast
}
\description{
Estimate the time varying R0 - using EpiEstim
}
\examples{

## Nowcast Rts                  
estimates <- estimate_R0(cases = EpiSoon::example_obs_cases, 
                         generation_times = as.matrix(EpiNow::covid_generation_times[,2]), 
                         rt_prior = list(mean_prior = 2.6, std_prior = 2),
                         windows = c(1, 3, 7), rt_samples = 10, gt_samples = 1)
                         
                         
estimates$rts
 
\dontrun{ 
## Nowcast Rts, forecast Rts and the forecast cases
estimates <- estimate_R0(cases = EpiSoon::example_obs_cases, 
                         generation_times = as.matrix(EpiNow::covid_generation_times[,1]), 
                         rt_prior = list(mean_prior = 2.6, std_prior = 2),
                         windows = c(1, 3, 7), rt_samples = 10, gt_samples = 20,
                         min_est_date =  as.Date("2020-02-18"),
                         forecast_model = function(...){
                              EpiSoon::fable_model(model = fable::ETS(y ~ trend("A")), ...)
                              },
                         horizon = 14)
                                           
## Rt estimates and forecasts
estimates$rts



## Case forecasts
estimates$cases
}
}
