% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_littler.R
\name{estimate_r_in_window}
\alias{estimate_r_in_window}
\title{Estimate r in a set time window}
\usage{
estimate_r_in_window(
  onsets = NULL,
  min_time = NULL,
  max_time = NULL,
  bootstrap_samples = 1000
)
}
\arguments{
\item{onsets}{A list of samples datasets nested within the dataset sampled from.}

\item{min_time}{Numeric, the minimum time to fit the model to.}

\item{max_time}{Numeric, the maximum time to fit the model to.}

\item{bootstrap_samples}{Numeric, defaults to 1000. The number of samples to take when
bootstrapping little r to account for model uncertainty.}
}
\value{
A list of 3 dataframes containing estimates for little r, doubling time and
model goodness of fit.
}
\description{
Estimate r in a set time window
}
