% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{estimates_by_report_date}
\alias{estimates_by_report_date}
\title{Estimate Cases by Report Date}
\usage{
estimates_by_report_date(
  estimates,
  forecast,
  delays,
  CrIs = c(0.2, 0.5, 0.9),
  target_folder = NULL,
  samples = TRUE
)
}
\arguments{
\item{estimates}{List of data frames as output by \code{estimate_infections}}

\item{forecast}{A list of data frames as output by \code{forecast_infections}}

\item{delays}{A call to \code{delay_opts} defining delay distributions and options. See the documentation of \code{delay_opts}
and the examples below for details.}

\item{CrIs}{Numeric vector of credible intervals to calculate.}

\item{target_folder}{Character string specifying where to save results (will create if not present).}

\item{samples}{Logical, defaults to TRUE. Should samples be saved}
}
\value{
A list of samples and summarised estimates of estimated cases by date of report
}
\description{
\lifecycle{questioning}
Either extracts or converts reported cases from an input datatable. For ouput from
\code{estimate_infections} this is a simple filtering step but for output from \code{forecast_infection}
this is currently an approximate convolution. This step is likely to be updated/deprecated
in new releases as \code{forecast_infections} evolves to be based on \code{stan} functionality.
}
