% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_estimate_infections}
\alias{example_estimate_infections}
\title{Example estimate_infections output}
\format{
An \code{estimate_infections} object with the results
}
\usage{
example_estimate_infections
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
An example output of \code{estimate_infections} containing 200 samples.
See here for details:
https://github.com/epiforecasts/EpiNow2/blob/main/data-raw/estimate_infections.R # nolint
}
\keyword{datasets}
