% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_generation_time}
\alias{example_generation_time}
\title{Example generation time}
\format{
A \code{dist_spec} object summarising the distribution
}
\usage{
example_generation_time
}
\description{
An example of a generation time estimate. See here for details:
https://github.com/epiforecasts/EpiNow2/blob/main/data-raw/generation-time.R
}
\keyword{datasets}
