% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_regional_epinow}
\alias{example_regional_epinow}
\title{Example regional_epinow output}
\format{
A list with the results
}
\usage{
example_regional_epinow
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
An example output of \code{regional_epinow} containing 200 samples per region.
See here for details:
https://github.com/epiforecasts/EpiNow2/blob/main/data-raw/estimate_infections.R # nolint
}
\keyword{datasets}
