% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_latent_state}
\alias{extract_latent_state}
\title{Extract Samples for a Latent State from a Stan model}
\usage{
extract_latent_state(param, samples, dates)
}
\arguments{
\item{param}{Character string indicating the latent state to extract}

\item{samples}{Extracted stan model (using \code{\link[rstan:stanfit-method-extract]{rstan::extract()}})}

\item{dates}{A vector identifying the dimensionality of the latent state to
extract. Generally this will be a date.}
}
\value{
A \verb{<data.frame>} containing the parameter name, date, sample id and
sample value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Extracts a time-varying latent state from a list of stan output and returns
it as a \verb{<data.table>}.
}
\keyword{internal}
