% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_parameter}
\alias{extract_parameter}
\title{Extract Samples from a Parameter with a Single Dimension}
\usage{
extract_parameter(param, samples)
}
\arguments{
\item{param}{Character string indicating the latent state to extract}

\item{samples}{Extracted stan model (using \code{\link[rstan:stanfit-method-extract]{rstan::extract()}})}
}
\value{
A \verb{<data.frame>} containing the parameter name, sample id and sample
value, or NULL if the parameter doesn't exist in the samples
}
\description{
Extract Samples from a Parameter with a Single Dimension
}
\keyword{internal}
