% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract samples from all parameters}
\usage{
extract_parameters(samples, args = NULL)
}
\arguments{
\item{samples}{Extracted stan model (using \code{\link[rstan:stanfit-method-extract]{rstan::extract()}})}

\item{args}{Stan data list containing param_id_* and params_variable_lookup
for parameter naming. If NULL, generic names are used.}
}
\value{
A \verb{<data.table>} containing the parameter name, sample id and sample
value, or NULL if parameters don't exist in the samples
}
\description{
Extract samples from all parameters
}
\keyword{internal}
