% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{extract_params}
\alias{extract_params}
\title{Internal function for extracting given parameter names of a distribution
from the environment. Called by \code{new_dist_spec}}
\usage{
extract_params(params, distribution)
}
\arguments{
\item{params}{Given parameters (obtained using \code{as.list(environment())})}

\item{distribution}{Character; the distribution to use.}
}
\value{
A character vector of parameters and their values.
}
\description{
Internal function for extracting given parameter names of a distribution
from the environment. Called by \code{new_dist_spec}
}
\keyword{internal}
