% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract all samples from a stan fit}
\usage{
extract_samples(stan_fit, pars = NULL, include = TRUE)
}
\arguments{
\item{stan_fit}{A \verb{<stanfit>} or \verb{<CmdStanMCMC>} object as returned by
\code{\link[=fit_model]{fit_model()}}.}

\item{pars}{Any selection of parameters to extract}

\item{include}{whether the parameters specified in \code{pars} should be included
(\code{TRUE}, the default) or excluded (\code{FALSE})}
}
\value{
List of data.tables with samples
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use \code{\link[=get_samples]{get_samples()}} on the fitted model object (e.g. the result of
\code{\link[=estimate_infections]{estimate_infections()}}) instead, with \code{format = "list"} to obtain the
same list of arrays that this function returns.

If the \code{object} argument is a \verb{<stanfit>} object, it simply returns the
result of \code{\link[rstan:extract]{rstan::extract()}}. If it is a \verb{<CmdStanMCMC>} it returns samples
in the same format as \code{\link[rstan:extract]{rstan::extract()}} does for \verb{<stanfit>} objects.
}
\seealso{
\code{\link[=get_samples]{get_samples()}} for the recommended replacement
}
