% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_static_parameter}
\alias{extract_static_parameter}
\title{Extract Samples from a Parameter with a Single Dimension}
\usage{
extract_static_parameter(param, samples)
}
\arguments{
\item{param}{Character string indicating the parameter to extract}

\item{samples}{Extracted stan model (using \code{rstan::extract})}
}
\value{
A data frame containing the parameter name, sample id and sample value
}
\description{
Extract Samples from a Parameter with a Single Dimension
}
