% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fill_missing_obs}
\alias{fill_missing_obs}
\title{Temporary function to support the transition to full support of missing
data.}
\usage{
fill_missing_obs(data, obs, cols)
}
\arguments{
\item{data}{A \verb{<data.frame>} of confirmed cases (confirm) by date (date).
\code{confirm} must be numeric and \code{date} must be in date format. Optionally
this can also have a logical \code{accumulate} column which indicates whether
data should be added to the next data point. This is useful when modelling
e.g. weekly incidence data. See also the \code{\link[=fill_missing]{fill_missing()}} data which helps
add the \code{accumulate} column with the desired properties when dealing with
non-daily data.}

\item{obs}{A list of options as generated by \code{\link[=obs_opts]{obs_opts()}} defining the
observation model. Defaults to \code{\link[=obs_opts]{obs_opts()}}.}
}
\value{
data set with missing dates filled in as na values
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
