% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_estimate_model}
\alias{fit_estimate_model}
\title{Build Stan arguments and fit an estimation model}
\usage{
fit_estimate_model(
  stan,
  data,
  init,
  model,
  id = model,
  verbose = FALSE,
  time_points = NULL
)
}
\arguments{
\item{stan}{A list of stan options as generated by \code{\link[=stan_opts]{stan_opts()}}. Defaults
to \code{\link[=stan_opts]{stan_opts()}}. Can be used to override \code{data}, \code{init}, and \code{verbose}
settings if desired.}

\item{data}{A list of stan data as created by \code{\link[=create_stan_data]{create_stan_data()}}}

\item{init}{Initial conditions passed to \code{{rstan}}. Defaults to "random"
(initial values randomly drawn between -2 and 2) but can also be a
function (as supplied by \code{\link[=create_initial_conditions]{create_initial_conditions()}}).}

\item{model}{Character, name of the model for which arguments are
to be created.}

\item{id}{Character, identifier used in logging. Defaults to \code{model}.}

\item{verbose}{Logical, defaults to \code{FALSE}. Should verbose progress
messages be returned.}

\item{time_points}{Integer length of the observed time period passed to
\code{\link[=check_truncation_length]{check_truncation_length()}}. If \code{NULL} (the default) the check is
skipped.}
}
\value{
A fitted model object as returned by \code{\link[=fit_model]{fit_model()}}.
}
\description{
Shared fitting step for the \verb{estimate_*()} functions. Builds the Stan
arguments with \code{\link[=create_stan_args]{create_stan_args()}}, warns if the truncation distribution
is longer than the observed time period, and fits the model with
\code{\link[=fit_model]{fit_model()}}.
}
\keyword{internal}
