% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{fit_model}
\alias{fit_model}
\title{Fit a model using the chosen backend.}
\usage{
fit_model(args, id = "stan")
}
\arguments{
\item{args}{List of stan arguments.}

\item{id}{A character string used to assign logging information on error.
Used by \code{\link[=regional_epinow]{regional_epinow()}} to assign errors to regions. Alter the default to
run with error catching.}
}
\description{
Internal function for dispatch to fitting with NUTS or VB.
}
\keyword{internal}
