% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_infections.R
\name{fit_model_with_laplace}
\alias{fit_model_with_laplace}
\title{Fit a Stan Model using the Laplace Algorithm}
\usage{
fit_model_with_laplace(args, id = "stan")
}
\arguments{
\item{args}{List of stan arguments.}

\item{id}{A character string used to assign logging information on error.
Used by \code{\link[=regional_epinow]{regional_epinow()}} to assign errors to regions. Alter the default to
run with error catching.}
}
\value{
A stan model object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
