% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fix_dist}
\alias{fix_dist}
\title{Remove uncertainty in the parameters of a \verb{<dist_spec>}}
\usage{
fix_dist(x, strategy = c("mean", "sample"))
}
\arguments{
\item{x}{A \verb{<dist_spec>}}

\item{strategy}{Character; either "mean" (use the mean estimates of the
mean and standard deviation) or "sample" (randomly sample mean and
standard deviation from uncertainty given in the \verb{<dist_spec>}}
}
\value{
A \verb{<dist_spec>} object without uncertainty
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function has been renamed to \code{\link[=fix_parameters]{fix_parameters()}} as a more appropriate
name.
}
\keyword{internal}
