% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{format_quantiles_scoringutils}
\alias{format_quantiles_scoringutils}
\title{Format quantiles for scoringutils}
\usage{
format_quantiles_scoringutils(samples, quantiles, forecast_date)
}
\arguments{
\item{samples}{Data.table with date and value columns}

\item{quantiles}{Numeric vector of quantile levels}

\item{forecast_date}{Date when the forecast was made}
}
\value{
Data.table with columns: forecast_date, date, horizon,
quantile_level, predicted
}
\description{
Helper function to format posterior samples into quantiles in the structure
expected by \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
}
\keyword{internal}
