% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{format_sample_predictions}
\alias{format_sample_predictions}
\title{Format sample predictions}
\usage{
format_sample_predictions(samples, forecast_date)
}
\arguments{
\item{samples}{Data.table with date, sample, and value columns}

\item{forecast_date}{Date when the forecast was made}
}
\value{
Data.table with columns: forecast_date, date, horizon, sample,
predicted
}
\description{
Helper function to format posterior samples into the structure expected by
\code{\link[scoringutils:as_forecast_sample]{scoringutils::as_forecast_sample()}}.
}
\keyword{internal}
