% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{generate_dist_spec}
\alias{generate_dist_spec}
\title{Internal function for generating a \code{dist_spec} given parameters and a
distribution.}
\usage{
generate_dist_spec(params, distribution)
}
\arguments{
\item{params}{Given parameters (obtained using \code{as.list(environment())})}

\item{distribution}{Character; the distribution to use.}
}
\value{
A \code{dist_spec} of the given specification.
}
\description{
This will convert all parameters to natural parameters before generating
a \code{dist_spec}. If they have uncertainty this will be done using sampling.
}
\author{
Sebastian Funk
}
