% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_tools.R
\name{generate_pseudo_linelist}
\alias{generate_pseudo_linelist}
\title{Generate a sample linelist from the observed linelist and sampled linelists}
\usage{
generate_pseudo_linelist(
  count_linelist = NULL,
  observed_linelist = NULL,
  merge_actual_onsets = TRUE
)
}
\arguments{
\item{count_linelist}{Dataframe with two variables: date_report and daily_linelist. As generated by  \code{linelist_from_case_counts}.}

\item{observed_linelist}{Dataframe with two variables: date_report and daily_observed_linelist. As generated by `split_linelist_by_day``}

\item{merge_actual_onsets}{Logical, defaults to \code{TRUE}. Should linelist onset dates be used where available?}
}
\value{
Dataframe with   two variables: date_report and date_onset
}
\description{
Generate a sample linelist from the observed linelist and sampled linelists
}
