% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{get_dist_def}
\alias{get_dist_def}
\title{Get a Parameters that Define a Discrete Distribution}
\usage{
get_dist_def(
  values,
  verbose = FALSE,
  samples = 1,
  bootstraps = 1,
  bootstrap_samples = 250
)
}
\arguments{
\item{values}{Numeric vector of integer values.}

\item{verbose}{Logical, defaults to \code{FALSE}. Should progress messages be printed}

\item{samples}{Numeric, number of samples to take overall from the bootstrapped posteriors.}

\item{bootstraps}{Numeric, defaults to 1. The number of bootstrap samples (with replacement)
of the delay distribution to take.}

\item{bootstrap_samples}{Numeric, defaults to 100. The number of samples to take in each boostrap.
When the sample size of the supplied delay distribution is less than 100 this is used instead.}
}
\value{
A data.table of distributions and the parameters that define them.
}
\description{
Get a Parameters that Define a Discrete Distribution
}
\examples{

## Example with exponential and a small smaple
delays <- rexp(20, 1)

get_dist_def(delays, samples = 10, verbose = TRUE)


\dontrun{
## Example with gamma and a larger sample
delays <- rgamma(50, 4, 1)

out <- get_dist_def(delays, samples = 2, bootstraps = 2)

## Inspect
out

## Inspect one parameter
out$params[[1]]


## Load into skeleton and sample with truncation
EpiNow::dist_skel(10, model = out$model[[1]],
                  params = out$params[[1]],
                  max_value = out$max_value[[1]])
}
}
\author{
Sebastian Funk \href{mailto:sebastian.funk@lshtm.ac.uk}{sebastian.funk@lshtm.ac.uk}
}
