% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{get_dist_spec_id}
\alias{get_dist_spec_id}
\title{Perform checks for \verb{<dist_spec>} \code{get_...} functions}
\usage{
get_dist_spec_id(x, id)
}
\arguments{
\item{x}{A \verb{<dist_spec>}.}

\item{id}{Integer; the id of the distribution to get parameters of (if x is
a composite distribution). If \code{x} is a single distribution this is ignored
and can be left as \code{NULL}.}
}
\value{
The id to use.
}
\description{
Perform checks for \verb{<dist_spec>} \code{get_...} functions
}
\author{
Sebastian Funk
}
\keyword{internal}
