% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{get_distribution}
\alias{get_distribution}
\title{Get the distribution of a \code{\link[=dist_spec]{dist_spec()}}}
\usage{
get_distribution(x, id = NULL)
}
\arguments{
\item{x}{A \verb{<dist_spec>}.}

\item{id}{Integer; the id of the distribution to get parameters of (if x is
a composite distribution). If \code{x} is a single distribution this is ignored
and can be left as \code{NULL}.}
}
\value{
A character string naming the distribution (or "nonparametric")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
dist <- Gamma(shape = 3, rate = 2, max = 10)
get_distribution(dist)
}
