% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{get_element}
\alias{get_element}
\title{Extracts an element of a \verb{<dist_spec>}}
\usage{
get_element(x, id, element)
}
\arguments{
\item{x}{A \verb{<dist_spec>}.}

\item{id}{Integer; the id of the distribution to get parameters of (if x is
a composite distribution). If \code{x} is a single distribution this is ignored
and can be left as \code{NULL}.}

\item{element}{The element, i.e. "parameters", "pmf" or "distribution.}
}
\value{
The id to use.
}
\description{
Extracts an element of a \verb{<dist_spec>}
}
\keyword{internal}
