% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_generation_time}
\alias{get_generation_time}
\title{Get a Literature Distribution for the Generation Time}
\usage{
get_generation_time(disease, source, max_value = 14, fixed = FALSE)
}
\arguments{
\item{disease}{A character string indicating the disease of interest.}

\item{source}{A character string indicating the source of interest.}

\item{max_value}{Numeric, the maximum value to allow. Defaults to 14 days.}

\item{fixed}{Logical, defaults to \code{FALSE}. Should distributions be supplied
as fixed values (vs with uncertainty)?}
}
\value{
A list defining a distribution
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Extracts a literature distribution from \code{generation_times}.
}
\examples{
get_generation_time(disease = "SARS-CoV-2", source = "ganyani")
}
\author{
Sam Abbott
}
