% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_incubation_period}
\alias{get_incubation_period}
\title{Get a Literature Distribution for the Incubation Period}
\usage{
get_incubation_period(disease, source, max_value = 15)
}
\arguments{
\item{disease}{A character string indicating the disease of interest.}

\item{source}{A character string indicating the source of interest.}

\item{max_value}{Numeric, the maximum value to allow. Defaults to 15 days.}
}
\value{
A list defining a distribution
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Extracts a literature distribution from \code{incubation_periods}
}
\examples{
get_incubation_period(disease = "SARS-CoV-2", source = "lauer")
}
