% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_local_import_case_counts}
\alias{get_local_import_case_counts}
\title{Combine total and imported case counts}
\usage{
get_local_import_case_counts(total_cases, linelist = NULL, cases_from = NULL)
}
\arguments{
\item{total_cases}{Dataframe with following variables: \code{date} and \code{cases}.}

\item{linelist}{Dataframe with at least the following variables: \code{date_confirm}, \code{import_status}}

\item{cases_from}{A character string containing a date in the format \code{"yyyy-mm-dd"}. Applies a
filter to returned cases.}
}
\value{
A tibble containing cases by date locally and imported
}
\description{
Combine total and imported case counts
}
