% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_raw_result}
\alias{get_raw_result}
\title{Get a Single Raw Result}
\usage{
get_raw_result(file, region, date, result_dir)
}
\arguments{
\item{file}{Character string giving the result files name.}

\item{region}{Character string giving the region of interest.}

\item{date}{Target date (in the format \verb{"yyyy-mm-dd}).}

\item{result_dir}{Character string giving the location of the target directory}
}
\value{
An R object read in from the targeted .rds file
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
}
