% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_regions_with_most_reports}
\alias{get_regions_with_most_reports}
\title{Get Regions with Most Reported Cases}
\usage{
get_regions_with_most_reports(reported_cases, time_window = 7, no_regions = 6)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date), and region (\code{region}).}

\item{time_window}{Numeric, number of days to include from latest date in data.
Defaults to 7 days.}

\item{no_regions}{Numeric, number of regions to return. Defaults to 6.}
}
\value{
A character vector of regions with the highest reported cases
}
\description{
\lifecycle{stable}
Extract a vector of regions with the most reported cases in a set time window.
}
