% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_half_range}
\alias{gp_half_range}
\title{Half-range of the Gaussian process time points}
\usage{
gp_half_range(n)
}
\arguments{
\item{n}{Length of the Gaussian process.}
}
\value{
The half-range (in days) of \code{n} equally spaced daily time points,
with a minimum of 0.5.
}
\description{
Half-range of the Gaussian process time points
}
\keyword{internal}
