% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_ls_quantiles}
\alias{gp_ls_quantiles}
\title{Quantiles of the lengthscale prior}
\usage{
gp_ls_quantiles(ls, probs = c(0.05, 0.95))
}
\arguments{
\item{ls}{A \verb{<dist_spec>} giving the lengthscale prior.}

\item{probs}{Numeric vector of probabilities.}
}
\value{
A numeric vector of quantiles.
}
\description{
Quantiles of a lengthscale prior truncated to be positive and below its
maximum.
}
\keyword{internal}
