% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_noise_terms}
\alias{gp_noise_terms}
\title{Number of time points modelled by the Gaussian process}
\usage{
gp_noise_terms(data)
}
\arguments{
\item{data}{A list containing the following numeric values:
\code{t}, \code{seeding_time}, \code{horizon}, \code{future_fixed}, \code{fixed_from},
\code{stationary} and \code{estimate_r}.}
}
\value{
An integer giving the length of the Gaussian process.
}
\description{
Mirrors \code{setup_noise()} in the Stan model.
}
\keyword{internal}
