% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_settings}
\alias{gp_settings}
\title{Approximate Gaussian Process Settings}
\usage{
gp_settings(gp = list(), time = NA)
}
\arguments{
\item{gp}{A list of settings to override the defaults. Defaults to an empty list.}

\item{time}{The maximum observed time, defaults to NA.}
}
\value{
A list of settings defining the gaussian process
}
\description{
Defines a list specifying the structure of the approximate Gaussian
process. Custom settings can be supplied which override the defaults. Used internally
by \code{create_gp_data}, \code{create_stan_data}, and \code{estimate_infections}. The settings
returned (all of which are modifiable by the user) are:
\itemize{
\item \code{ls_mean} and \code{ls_sd}: The mean and standard deviation of the log normal length scale with
default values of 21 days and 7 days respectively.
\item \code{ls_min} and \code{ls_max}: The minimum and maximum values of the length scale with default values
of 3 and the maximum length of the data (input by the user) or 9 weeks if no maximum given.
\item \code{alpha_sd}: The standard deviation of the magnitude parameter of the kernel. This defaults to 0.1 and
should be approximately the expected standard deviation of the logged Rt.
\item \code{kernel}: The type of kernel required, currently supporting the squared exponential kernel ("se")
and the 3 over 2 Matern kernel ("matern", with \code{matern_type = 3/2} (no other form of Matern
kernels are currently supported)). Defaulting to the Matern 3 over 2 kernel.
as discontinuities are expected in Rt and infections.
\item \code{basis_prop} and \code{boundary_scale}: The proportion of basis functions to time points and
the boundary scale of the approximate gaussian process. The proportion of basis functions
defaults to 0.3 (and much be greater than 0), and the boundary scale defaults to 2.
Decreasing either of these values should increase run times at the cost of the accuracy of
Gaussian process approximation. In general smaller posterior length scales require a
higher proportion of basis functions and the user should only rarely alter the boundary scale.
These settings are an area of active research. See https://arxiv.org/abs/2004.11408 for further
details.
\item \code{stationary}: Should the Gaussian process be estimated with a stationary global mean or be second
order and so depend on the previous value. A stationary Gaussian process may be more
tractable but will revert to the global average when data is sparse i.e for near real
time estimates. The default is FALSE. This feature is experimental.
}
}
\examples{
# default settings
gp_settings()

# add a custom length scale
gp_settings(gp = list(ls_mean = 4))
}
