% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incubation_period}
\alias{incubation_period}
\title{Literature Estimates of Incubation Periods}
\format{
A \code{data.table} of summarising the distribution
}
\usage{
incubation_period
}
\description{
Incubation period estimates. See
here for details: https://github.com/epiforecasts/EpiNow2/blob/master/data-raw/incubation-period.R
}
\keyword{datasets}
