% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incubation_periods}
\alias{incubation_periods}
\title{Literature Estimates of Incubation Periods}
\format{
A \code{data.table} of summarising the distribution
}
\usage{
incubation_periods
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
Incubation period estimates. See here for details:
https://github.com/epiforecasts/EpiNow2/blob/master/data-raw/incubation-period.R
}
\keyword{datasets}
