% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{initialise_with_pathfinder}
\alias{initialise_with_pathfinder}
\title{Initialise Sampling Using the Pathfinder Algorithm}
\usage{
initialise_with_pathfinder(stan, data, init, init_method, fixed_param, verbose)
}
\arguments{
\item{stan}{A list of stan options as generated by \code{\link[=stan_opts]{stan_opts()}}. Defaults
to \code{\link[=stan_opts]{stan_opts()}}. Can be used to override \code{data}, \code{init}, and \code{verbose}
settings if desired.}

\item{data}{A list of stan data as created by \code{\link[=create_stan_data]{create_stan_data()}}}

\item{init}{Initial conditions passed to \code{{rstan}}. Defaults to "random"
(initial values randomly drawn between -2 and 2) but can also be a
function (as supplied by \code{\link[=create_initial_conditions]{create_initial_conditions()}}). Overridden when
\code{stan$init_method} is "pathfinder" (see \code{\link[=stan_sampling_opts]{stan_sampling_opts()}}), unless
\code{fixed_param} is \code{TRUE}.}

\item{init_method}{A character string, either "random" or "pathfinder".}

\item{fixed_param}{Logical, defaults to \code{FALSE}. Should arguments be
created to sample from fixed parameters (used by simulation functions).}

\item{verbose}{Logical, defaults to \code{FALSE}. Should verbose progress
messages be returned.}
}
\value{
\code{init}, or the pathfinder algorithm's fitted object
}
\description{
Returns \code{init} unchanged unless \code{init_method} is "pathfinder" and sampling
(rather than fixed-parameter sampling or an approximate method) is being
used, in which case the pathfinder algorithm's estimate of the posterior
is returned instead, for use as \code{init} when initialising the NUTS sampler.
}
\keyword{internal}
