% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{is_process_spec}
\alias{is_process_spec}
\title{Test whether an object is a time-varying process specification}
\usage{
is_process_spec(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical, \code{TRUE} if \code{x} is a \verb{<process_spec>}.
}
\description{
Test whether an object is a time-varying process specification
}
\keyword{internal}
