% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{is_state_spec}
\alias{is_state_spec}
\title{Test whether an object is a time-varying state specification}
\usage{
is_state_spec(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical, \code{TRUE} if \code{x} is a \verb{<state_spec>}.
}
\description{
Test whether an object is a time-varying state specification
}
\keyword{internal}
