% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_tools.R
\name{linelist_from_case_counts}
\alias{linelist_from_case_counts}
\title{Sample a linelist from case counts and a reporting delay distribution}
\usage{
linelist_from_case_counts(cases = NULL)
}
\arguments{
\item{cases}{Dataframe with two variables: confirm (numeric) and date_report (date).}
}
\value{
A linelist grouped by day as a \code{data.table} with two variables: date_report, and daily_observed_linelist
}
\description{
Sample a linelist from case counts and a reporting delay distribution
}
