% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{load_nowcast_result}
\alias{load_nowcast_result}
\title{Load nowcast results}
\usage{
load_nowcast_result(
  file = NULL,
  region = NULL,
  date = target_date,
  result_dir = results_dir
)
}
\arguments{
\item{file}{Character string giving the result files name.}

\item{region}{Character string giving the region of interest.}

\item{date}{Target date (in the format \verb{"yyyy-mm-dd}).}

\item{result_dir}{Character string giving the location of the target directory}
}
\value{
An R object read in from the targeted .rds file
}
\description{
Load nowcast results
}
