% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{lower_bounds}
\alias{lower_bounds}
\title{Get the lower bounds of the parameters of a distribution}
\usage{
lower_bounds(distribution)
}
\arguments{
\item{distribution}{Character; the distribution to use.}
}
\value{
A numeric vector, the lower bounds.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This is used to avoid sampling parameter values that have no support.
}
\examples{
\dontrun{
lower_bounds("lognormal")
}
}
\keyword{internal}
