% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_param}
\alias{make_param}
\title{Internal function to create a parameter list}
\usage{
make_param(name, dist = NULL, lower_bound = -Inf)
}
\arguments{
\item{name}{Character, name of the parameter}

\item{dist}{\verb{<dist_spec>}, the distribution of the parameter}

\item{lower_bound}{Numeric, lower bound for the parameter; default: -Inf}
}
\value{
A list with the parameter details, classed as "EpiNow2.param"
}
\description{
Internal function to create a parameter list
}
\keyword{internal}
