% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_output_arguments}
\alias{match_output_arguments}
\title{Match Input Output Arguments with Supported Options}
\usage{
match_output_arguments(
  input_args = c(),
  supported_args = c("fit", "estimates", "partial", "samples", "plots"),
  logger = "EpiNow2"
)
}
\arguments{
\item{input_args}{A character vector of input arguments (can be partial).}

\item{supported_args}{A character vector of supported output arguments.}

\item{logger}{A character vector indicating the logger to target messages at.}
}
\value{
A logical vector of output arguments
}
\description{
Match Input Output Arguments with Supported Options
}
\examples{
# select nothing
match_output_arguments()

# select just plots
match_output_arguments("plots")

# select plots and samples
match_output_arguments(c("plots", "samples"))

# lazily select arguments
match_output_arguments("p")
}
