% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{natural_params}
\alias{natural_params}
\title{Get the names of the natural parameters of a distribution}
\usage{
natural_params(distribution)
}
\arguments{
\item{distribution}{Character; the distribution to use.}
}
\value{
A character vector, the natural parameters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
These are the parameters used in the stan models. All other parameter
representations are converted to these using \code{\link[=convert_to_natural]{convert_to_natural()}} before
being passed to the stan models.
}
\examples{
\dontrun{
natural_params("gamma")
}
}
\keyword{internal}
