% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{ndist}
\alias{ndist}
\title{Calculate the number of distributions in a \verb{<dist_spec>}}
\usage{
ndist(x)
}
\arguments{
\item{x}{A \verb{<dist_spec>} object.}
}
\value{
The number of distributions.
}
\description{
Calculate the number of distributions in a \verb{<dist_spec>}
}
\keyword{internal}
