% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{new_dist_spec}
\alias{new_dist_spec}
\title{Internal function for generating a \code{dist_spec} given parameters and a
distribution.}
\usage{
new_dist_spec(param, distribution)
}
\arguments{
\item{distribution}{Character; the distribution to use.}

\item{params}{Parameters of the distribution (including \code{max})}
}
\value{
A \code{dist_spec} of the given specification.
}
\description{
This will convert all parameters to natural parameters before generating
a \code{dist_spec}. If they have uncertainty this will be done using sampling.
}
\author{
Sebastian Funk
}
\keyword{internal}
