% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{new_epinowfit}
\alias{new_epinowfit}
\title{Construct a fitted model object}
\usage{
new_epinowfit(..., class)
}
\arguments{
\item{...}{Named components of the object, in the order they should
appear.}

\item{class}{Character, the model-specific class.}
}
\value{
A list of class \code{c(class, "epinowfit", "list")}.
}
\description{
Combines the components of a fitted model into a list with class
\code{c(class, "epinowfit", "list")}.
}
\keyword{internal}
