% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{new_process_spec}
\alias{new_process_spec}
\title{Construct a process specification}
\usage{
new_process_spec(type, mean, init, settings = list())
}
\arguments{
\item{type}{Character, the process type (\code{"gp"} or \code{"rw"}).}

\item{mean, init}{A \verb{<dist_spec>}; exactly one must be supplied.}

\item{settings}{A list of additional process settings (e.g. a \verb{<gp_opts>}
object for \code{"gp"}, or the step standard deviation prior for \code{"rw"}).}
}
\value{
A \verb{<process_spec>} object.
}
\description{
Construct a process specification
}
\keyword{internal}
