% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{obs_model_settings}
\alias{obs_model_settings}
\title{Observation Model Settings}
\usage{
obs_model_settings(obs_model = list())
}
\arguments{
\item{obs_model}{A list of settings to override the defaults.
Defaults to an empty list.}
}
\value{
A list of observation model settings.
}
\description{
Defines a list specifying the structure of the observation
model. Custom settings can be supplied which override the defaults. Used internally
by \code{create_obs_model}, \code{create_stan_data}, and \code{estimate_infections}. The settings
returned (all of which are modifiable by the user) are:
\itemize{
\item \code{family}: Observation model family. Options are Negative binomial ("negbin"),
the default, and Poisson.
\item \code{weight}: Weight to give the observed data in the log density. Numeric, defaults
to 1.
\item \code{week_effect}: Logical defaulting to \code{TRUE}. Should a day of the week effect be used.
\item \code{scale} List, defaulting to an empty list. Should an scaling factor be applied to map
latent infections (convolved to date of report). If none empty a mean (\code{mean}) and standard
deviation (\code{sd}) needs to be supplied definining the normally distributed scalng factor.
}
}
\examples{
# default settings
obs_model_settings()

# Turn off day of the week effect
obs_model_settings(obs_model = list(week_effect = TRUE))


# Scale reported data
obs_model_settings(obs_model = list(scale = list(mean = 0.2, sd = 0.02)))
}
