% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{package_model}
\alias{package_model}
\title{Load and compile the nowcasting model}
\usage{
package_model(
  model = c("estimate_infections", "simulate_infections", "estimate_secondary",
    "simulate_secondary", "estimate_truncation", "dist_fit"),
  dir = system.file("stan", package = "EpiNow2"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{model}{A character string indicating the model to use. One of
"estimate_infections" (default), "simulate_infections", "estimate_secondary",
"simulate_secondary", "estimate_truncation" or "dist_fit".}

\item{dir}{A character string specifying the path to any stan
files to include in the model. If missing the package default is used.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose
messages be shown.}

\item{...}{Additional arguments passed to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}
}
\value{
A \code{cmdstanr} model.
}
\description{
The function has been adapted from a similar function in the epinowcast
package (Copyright holder: epinowcast authors, under MIT License).
}
