% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pad_reported_cases}
\alias{pad_reported_cases}
\title{Pads reported cases with daily initial zeros}
\usage{
pad_reported_cases(reported_cases, n, with = NA_real_)
}
\arguments{
\item{reported_cases}{A <data.frame> of reported cases.}

\item{n}{The number of days to pad the reported cases by.}

\item{with}{What to pad with}
}
\value{
A data.table of reported cases.
}
\description{
Pads reported cases with daily initial zeros
}
\keyword{internal}
