% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_secondary.R
\name{plot.estimate_secondary}
\alias{plot.estimate_secondary}
\title{Plot method for estimate_secondary}
\usage{
\method{plot}{estimate_secondary}(x, primary = FALSE, ...)
}
\arguments{
\item{x}{A list of output as produced by \code{estimate_secondary}}

\item{primary}{Logical, defaults to \code{FALSE}. Should \code{primary} reports also
be plot?}

\item{...}{Pass additional arguments to plot function. Not currently in use.}
}
\value{
\code{ggplot2} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
\code{plot} method for class "estimate_secondary".
}
\seealso{
plot estimate_secondary
}
