% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_truncation.R
\name{plot.estimate_truncation}
\alias{plot.estimate_truncation}
\title{Plot method for estimate_truncation}
\usage{
\method{plot}{estimate_truncation}(x)
}
\arguments{
\item{x}{A list of output as produced by \code{estimate_truncation}}
}
\value{
\code{ggplot2} object
}
\description{
\lifecycle{experimental}
\code{plot} method for class "estimate_truncation". Returns
a plot facetted over each dataset used in fitting with the latest
observations as columns, the data observed at the time (and so truncated)
as dots and the truncation adjusted estimates as a ribbon.
}
\seealso{
plot estimate_truncation
}
