% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.forecast_infections}
\alias{plot.forecast_infections}
\title{Plot method for forecast_infections}
\usage{
\method{plot}{forecast_infections}(x, type = "summary", CrIs = c(0.2, 0.5, 0.9), ...)
}
\arguments{
\item{x}{A list of output as produced by \code{forecast_infections}}

\item{type}{A character vector indicating the name of the plot to return.
Defaults to  "summary" with supported options being "infections", "reports",
"R", "growth_rate", "summary", "all". If "all" is supplied all plots are
generated.}

\item{CrIs}{Numeric vector of credible intervals to calculate.}

\item{...}{Pass additional arguments to report_plots}
}
\value{
List of plots as produced by \code{\link[=report_plots]{report_plots()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
\code{plot} method for class \verb{<forecast_infections>}.
}
\seealso{
\code{\link[=report_plots]{report_plots()}} \code{\link[=forecast_infections]{forecast_infections()}}
}
