% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_secondary.R
\name{plot.forecast_secondary}
\alias{plot.forecast_secondary}
\title{Plot method for forecast_secondary objects}
\usage{
\method{plot}{forecast_secondary}(x, primary = FALSE, from = NULL, to = NULL, new_obs = NULL, ...)
}
\arguments{
\item{x}{A list of output as produced by \code{estimate_secondary}}

\item{primary}{Logical, defaults to \code{FALSE}. Should \code{primary} reports also
be plot?}

\item{from}{Date object indicating when to plot from.}

\item{to}{Date object indicating when to plot up to.}

\item{new_obs}{A \verb{<data.frame>} containing the columns \code{date} and \code{secondary}
which replace the secondary observations stored in the \code{estimate_secondary}
output.}

\item{...}{Pass additional arguments to plot function. Not currently in use.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Plot method for forecast secondary observations.
}
