% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_confidence}
\alias{plot_confidence}
\title{Plot a Time Series with Confidence.}
\usage{
plot_confidence(
  data,
  outer_alpha = 0.1,
  inner_alpha = 0.2,
  plot_median = TRUE,
  legend = "none"
)
}
\arguments{
\item{data}{Dataframe containing the follwoing variables: \code{date}, \code{median}, \code{type}, \code{bottom},
\code{top}, \code{lower}, \code{upper},  and \code{confidence}}

\item{outer_alpha}{Numeric, outer alpha level.}

\item{inner_alpha}{Numeric, inner alpha level.}

\item{plot_median}{Logical, defaults to \code{FALSE}. Should the median be plotted.}

\item{legend}{Character string defaults to "none". Should a legend be displayed.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a Time Series with Confidence.
}
