% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Plot Estimates}
\usage{
plot_estimates(
  estimate,
  reported,
  ylab = "Cases",
  hline,
  bar_width = 1.5,
  obs_as_col = TRUE
)
}
\arguments{
\item{estimate}{A data.table of estimates containing the following variables: date, type
(must contain "estimate", "estimate based on partial data" and optionally "forecast"),}

\item{reported}{A data.table of reported cases with the following variables: date, confirm.}

\item{ylab}{Character string, defaulting to "Cases". Title for the plot y axis.}

\item{hline}{Numeric, if supplied gives the horizontal intercept for a indicator line.}

\item{bar_width}{Numeric, defaults to 1.5. The width of the estimate bar. Users will likely wish to tune
this as their plot varies in size.}

\item{obs_as_col}{Logical, defaults to \code{TRUE}. Should observed data, if supplied, be plotted using columns or
as points (linked using a line).}
}
\value{
A \code{ggplot2} object
}
\description{
Plot Estimates
}
\examples{
\dontrun{
# first run the example in ?epinow

## Plot infections
plot_estimates(estimate = out$estimates$summarised[variable == "infections"][date <= as.Date("2020-03-04")],
               reported = cases, 
               ylab = "Cases")

## Plot reported cases estimated via Rt
plot_estimates(estimate = out$estimates$summarised[variable == "reported_cases_rt"],
               reported = cases, 
               ylab = "Cases")
               
## Plot Rt estimates
plot_estimates(estimate = out$estimates$summarised[variable == "R"],
               ylab = "Effective Reproduction No.",
               hline = 1)

}


}
