% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Add a Forecast to a Plot}
\usage{
plot_forecast(plot = NULL, forecast = NULL)
}
\arguments{
\item{plot}{\code{ggplot2} plot.}

\item{forecast}{Dataframe containing a forecast with the following variables: \code{bottom},
\code{top}, \code{lower}, and \code{upper}.}
}
\value{
A \code{ggplot2} plot
}
\description{
Add a Forecast to a Plot
}
