% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_grid}
\alias{plot_grid}
\title{Plot a Grid of Plots}
\usage{
plot_grid(
  regions = NULL,
  plot_object = "bigr_eff_plot.rds",
  results_dir = "results",
  target_date = NULL,
  ...
)
}
\arguments{
\item{regions}{A character string containing the list of regions to extract results for
(must all have results for the same target date).}

\item{plot_object}{A character string indicating the plot object to use as the
base for the grid.}

\item{results_dir}{A character string indicating the location of the results directory to extract results
from.}

\item{target_date}{A character string indicating the target date to extract results for. All regions must have results
for this date.}

\item{...}{Additional arguments to pass to \code{patchwork::plot_layout}}
}
\value{
A \code{ggplot2} object combining multiple plots
}
\description{
Plot a Grid of Plots
}
