% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pipeline}
\alias{plot_pipeline}
\title{Plot Pipeline Results}
\usage{
plot_pipeline(
  target_date = NULL,
  target_folder = NULL,
  min_plot_date = NULL,
  report_forecast = FALSE
)
}
\arguments{
\item{target_date}{Character string, in the form "2020-01-01". Date to cast.}

\item{target_folder}{Character string, name of the folder in which to save the results.}

\item{min_plot_date}{Character string, in the form "2020-01-01". Minimum date at which to start
plotting estimates.}

\item{report_forecast}{Logical, defaults to \code{FALSE}. Should the forecast be reported.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot Pipeline Results
}
