% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{+.dist_spec}
\alias{+.dist_spec}
\title{Creates a delay distribution as the sum of two other delay distributions}
\usage{
\method{`+`}{dist_spec}(e1, e2)
}
\arguments{
\item{e1}{The fist delay distributions (from a calls to \code{\link[=dist_spec]{dist_spec()}}) to
combine}

\item{e2}{The second delay distributions (from a calls to \code{\link[=dist_spec]{dist_spec()}})
to combine}
}
\value{
Delay distributions representing the sum of the two delays
(with class \code{\link[=dist_spec]{dist_spec()}}`)
}
\description{
This is done via convolution with \code{stats::convolve()}.
}
\author{
Sebastian Funk
}
