% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{print.dist_spec}
\alias{print.dist_spec}
\title{Prints the parameters of one or more delay distributions}
\usage{
\method{print}{dist_spec}(x, ...)
}
\arguments{
\item{x}{The \code{\link[=dist_spec]{dist_spec()}} to use}

\item{...}{Not used}
}
\value{
invisible
}
\description{
This displays the parameters of the uncertain and probability mass
functions of fixed delay distributions combined in the passed \code{\link[=dist_spec]{dist_spec()}}.
}
\author{
Sebastian Funk
}
