% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{print.epinowfit}
\alias{print.epinowfit}
\title{Print information about and object that has resulted from a model fit.}
\usage{
\method{print}{epinowfit}(x, ...)
}
\arguments{
\item{x}{The object containing fit results.}

\item{...}{Ignored}
}
\value{
Invisible
}
\description{
Print information about and object that has resulted from a model fit.
}
