% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_epinow.R
\name{process_region}
\alias{process_region}
\title{Process regional estimate}
\usage{
process_region(
  out,
  target_region,
  timing,
  return_output,
  complete_logger = "EpiNow2.epinow"
)
}
\arguments{
\item{out}{List of output returned by \code{epinow}}

\item{target_region}{Character string indicating the region being evaluated}

\item{timing}{Output from \code{Sys.time}}

\item{return_output}{Logical, defaults to TRUE. Should output be returned. This must either be true or a
\code{target_folder} must be specified in order to enable output saving to disk.}

\item{complete_logger}{Character string indicating the logger to output
the completion of estimation to.}
}
\value{
A list of processed output
}
\description{
Process regional estimate
}
