% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_epinow.R
\name{process_regions}
\alias{process_regions}
\title{Process all Region Estimates}
\usage{
process_regions(regional_out, regions)
}
\arguments{
\item{regional_out}{A list of output from multiple runs of
\code{\link[=regional_epinow]{regional_epinow()}}}

\item{regions}{A character vector identifying the regions that have been run}
}
\value{
A list of all regional estimates and successful regional estimates
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Internal function that processes the output from multiple \code{\link[=epinow]{epinow()}} runs,
adds summary logging information.
}
\seealso{
\code{\link[=regional_epinow]{regional_epinow()}} \code{\link[=epinow]{epinow()}}
}
\keyword{internal}
