% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pull_max_var}
\alias{pull_max_var}
\title{Extract a the Maximum Value of a Variable Based on a Filter}
\usage{
pull_max_var(df, max_var = NULL, sel_var = NULL, type_selected = NULL)
}
\arguments{
\item{df}{Datatable with the following variables: \code{type} and \code{si_dist}}

\item{max_var}{Character string containing the variable name to pull out the maximum R estimate for.}

\item{sel_var}{Character string indicating the variable to extract}

\item{type_selected}{The nowcast type to extract.}
}
\value{
A character string containing the maximum variable
}
\description{
Extract a the Maximum Value of a Variable Based on a Filter
}
\examples{

df <- data.table::data.table(type = c("nowcast", "other"),
                             var = c(1:10),
                             sel = "test")
                             
pull_max_var(df, max_var = "var", sel_var = "var", type_selected = "nowcast")
}
