% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{reconstruct_delay}
\alias{reconstruct_delay}
\title{Reconstruct a dist_spec from stored stan data}
\usage{
reconstruct_delay(args, delay_name)
}
\arguments{
\item{args}{The args list from a fitted model (containing stan data)}

\item{delay_name}{The name of the delay (e.g., "generation_time")}
}
\value{
A dist_spec object, or NULL if the delay doesn't exist
}
\description{
Reconstruct a dist_spec from stored stan data
}
\keyword{internal}
