% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Gamma}
\alias{LogNormal}
\alias{Normal}
\alias{Fixed}
\alias{Exp}
\alias{Weibull}
\alias{Dirichlet}
\alias{NonParametric}
\alias{discretise}
\alias{discretize}
\alias{get_pmf}
\alias{convert_to_logmean}
\alias{convert_to_logsd}
\alias{fix_parameters}
\alias{get_distribution}
\alias{get_parameters}
\alias{is_constrained}
\alias{bound_dist}
\alias{collapse}
\alias{new_dist_spec}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{distspec}{\code{\link[distspec]{bound_dist}}, \code{\link[distspec]{collapse}}, \code{\link[distspec]{convert_to_logmean}}, \code{\link[distspec]{convert_to_logsd}}, \code{\link[distspec]{Dirichlet}}, \code{\link[distspec]{discretise}}, \code{\link[distspec:discretise]{discretize}}, \code{\link[distspec:Exponential]{Exp}}, \code{\link[distspec]{fix_parameters}}, \code{\link[distspec]{Fixed}}, \code{\link[distspec]{Gamma}}, \code{\link[distspec]{get_distribution}}, \code{\link[distspec]{get_parameters}}, \code{\link[distspec]{get_pmf}}, \code{\link[distspec]{is_constrained}}, \code{\link[distspec]{LogNormal}}, \code{\link[distspec]{new_dist_spec}}, \code{\link[distspec]{NonParametric}}, \code{\link[distspec]{Normal}}, \code{\link[distspec]{Weibull}}}
}}

