% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_rt_pipeline.R
\name{regional_rt_pipeline}
\alias{regional_rt_pipeline}
\title{Regional Realtime Pipeline}
\usage{
regional_rt_pipeline(
  cases = NULL,
  linelist = NULL,
  delay_defs = NULL,
  incubation_defs = NULL,
  target_folder = "results",
  target_date = NULL,
  merge_onsets = FALSE,
  case_limit = 40,
  onset_modifier = NULL,
  dt_threads = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{cases}{A dataframe of cases (\code{confirm}) by date of confirmation (\code{date}), import status (\code{import_status}; ("imp)), and region (\code{region}).}

\item{linelist}{A dataframe of of cases (by row) containing the following variables:
\code{import_status} (values "local" and "imported"), \code{date_onset}, \code{date_confirm}, \code{report_delay}, and \code{region}. If a national linelist is not available a proxy linelist may be
used but in this case \code{merge_onsets} should be set to \code{FALSE}.}

\item{delay_defs}{A data.table that defines the delay distributions (model, parameters and maximum delay for each model).
See \code{get_delay_dist} for an example of the structure.}

\item{incubation_defs}{A data.table that defines the incubation distributions (model, parameters and maximum delay for each model).
See \code{get_delay_dist} for an example of the structure.}

\item{target_folder}{Character string indicating the folder into which to save results.
Also used to extract previously generated results.}

\item{target_date}{Character string, in the form "2020-01-01". Date to cast.}

\item{merge_onsets}{Logical defaults to \code{FALSE}. Should available onset data be used. Typically if \code{regional_delay} is}

\item{case_limit}{Numeric, the minimum number of cases in a region required for that region to be evaluated. Defaults to 10.
set to \code{FALSE} this should also be \code{FALSE}}

\item{onset_modifier}{data.frame containing a \code{date} variable and a function \code{modifier} variable. This is used
to modify estimated cases by onset date. \code{modifier} must be a function that returns a proportion when called
(enables inclusion of uncertainty) and takes the following arguments: \code{n} (samples to return) and \code{status} ("local" or "import").}

\item{dt_threads}{Numeric, the number of data.table threads to use. Set internally to avoid issue when running in parallel.
Defaults to 1 thread.}

\item{verbose}{Logical, defaults to \code{FALSE}. Should progress messages be shown for each reigon?}

\item{...}{Pass additional arguments to \code{rt_pipeline}}
}
\description{
Runs a pipeline by region.
}
\examples{
 \dontrun{
## Save everything to a temporary directory 
## Change this to inspect locally
target_dir <- tempdir() 

## Construct example distributions
## reporting delay dist
delay_dist <- suppressWarnings(
               EpiNow::get_dist_def(rexp(25, 1/10), 
                                    samples = 5, bootstraps = 1))

## Uses example case vector from EpiSoon
cases <- data.table::setDT(EpiSoon::example_obs_cases)
cases <- cases[, `:=`(confirm = as.integer(cases), import_status = "local")][,
                  cases := NULL]

cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))
  
## Run basic nowcasting pipeline
regional_rt_pipeline(cases = cases,
            delay_defs = delay_dist,
            target_folder = target_dir)
}
}
