% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{renewal_opts}
\alias{renewal_opts}
\title{Renewal Model Specification}
\usage{
renewal_opts(rt = rt_opts(), generation_time = gt_opts())
}
\arguments{
\item{rt}{A call to \code{\link[=rt_opts]{rt_opts()}} defining the time-varying reproduction
number. Defaults to \code{\link[=rt_opts]{rt_opts()}}.}

\item{generation_time}{A call to \code{\link[=gt_opts]{gt_opts()}} (or its alias
\code{\link[=generation_time_opts]{generation_time_opts()}}) defining the generation time distribution.
Defaults to \code{\link[=gt_opts]{gt_opts()}}.}
}
\value{
A \verb{<renewal_opts>} object that can be passed to the \code{model}
argument of \code{\link[=estimate_infections]{estimate_infections()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Specifies that infections should be estimated using the renewal equation
model. This is the mechanistic model that uses the generation time
distribution to relate the reproduction number to infections.
}
\examples{
# default settings
renewal_opts()

# with custom Rt prior
renewal_opts(rt = rt_opts(prior = LogNormal(mean = 2, sd = 1)))

# with custom generation time
generation_time <- Gamma(
  shape = Normal(1.3, 0.3),
  rate = Normal(0.37, 0.09),
  max = 14
)
renewal_opts(generation_time = gt_opts(generation_time))
}
\seealso{
\code{\link[=deconvolution_opts]{deconvolution_opts()}} \code{\link[=estimate_infections]{estimate_infections()}} \code{\link[=rt_opts]{rt_opts()}}
\code{\link[=gt_opts]{gt_opts()}}
}
