% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_littler}
\alias{report_littler}
\title{Report Rate of Growth Estimates}
\usage{
report_littler(target_folder)
}
\arguments{
\item{target_folder}{Character string indicating the folder into which to save results.
Also used to extract previously generated results.}
}
\description{
Report Rate of Growth Estimates
}
