% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_reff}
\alias{report_reff}
\title{Report Effective Reproduction Number Estimates}
\usage{
report_reff(target_folder)
}
\arguments{
\item{target_folder}{Character string indicating the folder into which to save results.
Also used to extract previously generated results.}
}
\description{
Report Effective Reproduction Number Estimates
}
