% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_regional_rt.R
\name{report_regional_rt}
\alias{report_regional_rt}
\title{Regional Rt}
\usage{
report_regional_rt(
  cases = NULL,
  target_folder = "results",
  target_date = NULL,
  case_limit = 20,
  cores = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{cases}{A dataframe of cases (\code{confirm}) by date of confirmation (\code{date}), import status (\code{import_status}; ("imp)), and region (\code{region}).}

\item{case_limit}{Numeric, the minimum number of cases in a region required for that region to be evaluated. Defaults to 10.
set to \code{FALSE} this should also be \code{FALSE}}

\item{cores}{Numeric, defaults to 2. The number of cores to use when fitting the stan model.}

\item{verbose}{Logical, defaults to \code{FALSE}. Should progress messages be shown for each reigon?}

\item{...}{Pass additional arguments to \code{rt_pipeline}}

\item{merge_onsets}{Logical defaults to \code{FALSE}. Should available onset data be used. Typically if \code{regional_delay} is}
}
\description{
Estimates Rt by region
}
\examples{
 \dontrun{
## Save everything to a temporary directory 
## Change this to inspect locally
target_dir <- tempdir() 

## Construct example distributions
## reporting delay dist
delay_dist <- suppressWarnings(
               EpiNow2::get_dist_def(rexp(25, 1/10), 
                                    samples = 5, bootstraps = 1))

## Uses example case vector from EpiSoon
cases <- data.table::setDT(EpiSoon::example_obs_cases)
cases <- cases[, `:=`(confirm = as.integer(cases))][,
                  cases := NULL]

cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))
  
## Run basic nowcasting pipeline
report_regional_rt(cases = cases,
            target_folder = target_dir)
}
}
