% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_rt.R
\name{report_rt}
\alias{report_rt}
\title{Real-time Pipeline}
\usage{
report_rt(
  reported_cases,
  family = "negbin",
  generation_time = generation_time,
  incubation_period = incubation_period,
  reporting_delay = reporting_delay,
  rt_prior = rt_prior,
  model = model,
  cores = 4,
  chains = 4,
  estimate_rt = TRUE,
  return_fit = FALSE,
  forecast_model,
  horizon = 0,
  report_forecast = FALSE,
  min_forecast_cases = 200,
  target_folder = NULL,
  target_date = NULL,
  verbose = FALSE
)
}
\arguments{
\item{report_forecast}{Logical, defaults to \code{FALSE}. Should the forecast be reported.}

\item{min_forecast_cases}{Numeric, defaults to 200. The minimum number of cases required in the last 7 days
of data in order for a forecast to be run. This prevents spurious forecasts based on highly uncertain Rt estimates.}

\item{target_folder}{Character string indicating the folder into which to save results.
Also used to extract previously generated results.}

\item{target_date}{Character string, in the form "2020-01-01". Date to cast.}

\item{verbose}{Logical, defaults to \code{FALSE}. Should internal nowcasting progress messages be returned.}

\item{linelist}{A dataframe of of cases (by row) containing the following variables:
\code{import_status} (values "local" and "imported"), \code{date_onset}, \code{date_confirm}, \code{report_delay}.}

\item{delay_cutoff_date}{Character string, in the form "2020-01-01". Cutoff date to use
to estimate the delay distribution.}

\item{earliest_allowed_onset}{A character string in the form of a date ("2020-01-01") indiciating the earliest
allowed onset.}

\item{approx_threshold}{Numeric, defaults to 10,000. Threshold of cases below which explicit sampling of onsets
always occurs.}

\item{generation_times}{A matrix with columns representing samples and rows representing the probability of the serial intervel being on
that day. Defaults to \code{EpiNow2::covid_generation_times}.}

\item{rt_samples}{Numeric, the number of samples to take from the estimated R distribution for each time point.}

\item{dt_threads}{Numeric, the number of data.table threads to use. Set internally to avoid issue when running in parallel.
Defaults to 1 thread.}
}
\description{
Combine fitting a delay distribution, constructing a set of
complete sampled linelists, nowcast cases by onset date, and estimate
the time-varying effective reproduction number and rate of spread.
}
\examples{
\dontrun{
## Save everything to a temporary directory 
## Change this to inspect locally
target_dir <- tempdir()

## Construct example distributions
## reporting delay dist
delay_dist <- suppressWarnings(
               EpiNow2::get_dist_def(rexp(25, 1/10), 
                                    samples = 5, bootstraps = 1))

## Uses example case vector from EpiSoon
cases <- data.table::setDT(EpiSoon::example_obs_cases)
cases <- cases[, `:=`(confirm = as.integer(cases), import_status = "local")][,
                  cases := NULL]

## Run basic nowcasting pipeline
rt_pipeline(cases = cases,
            delay_defs = delay_dist,
            target_date = max(cases$date),
            target_folder = target_dir)
            
            
## Run with forecasting and approximate delay sampling

## Requires additional packages:
library(EpiSoon)
library(forecastHybrid)

## Runs the estimation pipeline as before but this time with a 14 day Rt and case forecast
## Uses the {forecastHybrid} package to produce an unweighted
## ensemble using the last 3 weeks of data
rt_pipeline(cases = cases,
            target_folder = target_dir,
            horizon = 14, 
            report_forecast = TRUE
            forecast_model = function(y, ...){EpiSoon::forecastHybrid_model(
            y = y[max(1, length(y) - 21):length(y)],
            model_params = list(models = "aefz", weights = "equal"),
            forecast_params = list(PI.combination = "mean"), ...)})

}
}
