% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{resolve_gp_component}
\alias{resolve_gp_component}
\title{Resolve one Gaussian process component's state data}
\usage{
resolve_gp_component(comp, name, hyper_params, base_id, pos)
}
\arguments{
\item{comp}{The component (\verb{list(type = "gp", settings = <gp_opts>)}).}

\item{name}{The time-varying parameter's name, used in error messages.}

\item{hyper_params}{The parameter list to append the magnitude and
lengthscale to.}

\item{base_id}{The parameter id of the last existing parameter.}

\item{pos}{This component's index within the GP type group.}
}
\value{
A list of this component's contribution to \code{create_state_data()}'s
component table: \code{comp_type}, \code{comp_pos}, \code{hyper_params}, \code{gp_alpha_id},
\code{gp_rho_id}, \code{gp_kernel}, \code{gp_nu}, \code{gp_basis_prop}, \code{gp_boundary_scale}.
}
\description{
Resolve one Gaussian process component's state data
}
\keyword{internal}
