% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{rstan_opts}
\alias{rstan_opts}
\title{Rstan Options}
\usage{
rstan_opts(object = NULL, samples = 2000, method = "sampling", ...)
}
\arguments{
\item{object}{Stan model object. By default uses the compiled package
default.}

\item{samples}{Numeric, default 2000. Overall number of posterior samples.
When using multiple chains iterations per chain is samples / chains.}

\item{method}{A character string, defaulting to sampling. Currently supports
\code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} ("sampling") or \code{\link[rstan:stanmodel-method-vb]{rstan::vb()}} ("vb").}

\item{...}{Additional parameters to pass  underlying option functions.}
}
\value{
A list of arguments to pass to the appropriate rstan functions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated; specify options in \code{\link[=stan_opts]{stan_opts()}} instead.
}
\seealso{
\code{\link[=rstan_sampling_opts]{rstan_sampling_opts()}} \code{\link[=rstan_vb_opts]{rstan_vb_opts()}}
}
