% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{rstan_vb_opts}
\alias{rstan_vb_opts}
\title{Rstan Variational Bayes Options}
\usage{
rstan_vb_opts(samples = 2000, trials = 10, iter = 10000, ...)
}
\arguments{
\item{samples}{Numeric, default 2000. Overall number of approximate posterior
samples.}

\item{trials}{Numeric, defaults to 10. Number of attempts to use
rstan::vb()] before failing.}

\item{iter}{Numeric, defaulting to 10000. Number of iterations to use in
\code{\link[rstan:stanmodel-method-vb]{rstan::vb()}}.}

\item{...}{Additional parameters to pass to \code{\link[rstan:stanmodel-method-vb]{rstan::vb()}} or
\code{\link[cmdstanr:model-method-variational]{cmdstanr::variational()}}, depending on the chosen backend.}
}
\value{
A list of arguments to pass to \code{\link[rstan:stanmodel-method-vb]{rstan::vb()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated; use \code{\link[=stan_vb_opts]{stan_vb_opts()}} instead.
}
