% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{rt_settings}
\alias{rt_settings}
\title{Time-Varying Reproduction Number Settings}
\usage{
rt_settings(rt = list())
}
\arguments{
\item{rt}{A list of settings to override the defaults. Defaults to an empty list.}
}
\value{
A list of settings defining the time-varying reproduction number
}
\description{
Defines a list specifying the optional arguments for the time-varying
reproduction number. Custom settings can be supplied which override the defaults. Used internally
by \code{create_rt_data}, \code{create_stan_data}, and \code{estimate_infections}. The settings
returned (all of which are modifiable by the user) are:
\itemize{
\item \code{prior}: The mean and standard deviation of the log normal Rt prior. Defaults to
mean of 1 and standard deviation of 1.
\item \code{use_rt} Should Rt be used to generate infections and hence reported cases. Defaults
to \code{TRUE}.
\item \code{rw}: Numeric step size of the random walk. Defaults to 0. To specify a weekly random
walk set \code{rw = 7}. For more custom break point settings consider passing in a \code{breakpoints}
variable as outlined in the next section.
\item \code{use_breakpoints}: Should break points be used if present as a \code{breakpoint} variable in
the input data. Break points should be defined as 1 if present and otherwise 0. By default
breakpoints are fit jointly with a global non-parametric effect and so represent a conservative
estimate of break point changes (alter this by setting \code{gp = NULL}).
}
}
\examples{
# default settings
rt_settings()

# add a custom length scale
rt_settings(rt = list(prior = list(mean = 2, sd = 1)))

# add a weekly random walk
rt_settings(rt = list(rw = 7))
}
