% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_tools.R
\name{sample_delay}
\alias{sample_delay}
\title{Sample Onset Dates for Cases missing them}
\usage{
sample_delay(linelist = NULL, delay_fn = NULL, earliest_allowed_onset = NULL)
}
\arguments{
\item{linelist}{Dataframe with two variables: date_report and date_onset. As generated by  \code{generate_pseudo_linelist}.}

\item{delay_fn}{A sampling funtion that takes a single numeric argument and returns a vector of
numeric samples this long.}

\item{earliest_allowed_onset}{A character string in the form of a date ("2020-01-01") indiciating the earliest
allowed onset.}
}
\value{
Dataframe with no missing data and two variables: date_report and date_onset
}
\description{
Sample Onset Dates for Cases missing them
}
