% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{save_estimate_infections}
\alias{save_estimate_infections}
\title{Save Estimated Infections}
\usage{
save_estimate_infections(
  estimates,
  target_folder = NULL,
  samples = TRUE,
  return_fit = TRUE
)
}
\arguments{
\item{estimates}{List of data frames as output by \code{estimate_infections}}

\item{samples}{Logical, defaults to TRUE. Should samples be saved}

\item{return_fit}{Logical, defaults to FALSE. Should the fitted stan model be returned.}
}
\description{
Save Estimated Infections
}
