% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{save_forecast_infections}
\alias{save_forecast_infections}
\title{Save Forecast Infections}
\usage{
save_forecast_infections(forecast, target_folder = NULL, samples = TRUE)
}
\arguments{
\item{forecast}{A list of data frames as output by \code{forecast_infections}}

\item{target_folder}{Character string specifying where to save results (will create if not present).}

\item{samples}{Logical, defaults to TRUE. Should samples be saved}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Saves the output from \code{forecast_infections} to a target directory.
}
\seealso{
forecast_infections
}
