% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{sd.dist_spec}
\alias{sd.dist_spec}
\title{Returns the standard deviation of one or more delay distribution}
\usage{
\method{sd}{dist_spec}(x, ...)
}
\arguments{
\item{x}{The \code{\link[=dist_spec]{dist_spec()}} to use}

\item{...}{Not used}
}
\value{
A vector of standard deviations.
}
\description{
This works out the standard deviation of all the (parametric /
nonparametric) delay distributions combined in the passed \code{\link[=dist_spec]{dist_spec()}}.
}
\examples{
# A fixed lognormal distribution with sd 5 and sd 1.
lognormal <- dist_spec(
 sd = 5, sd = 1, max = 20, distribution = "lognormal"
)
sd(lognormal)

# An uncertain gamma distribution with sd 3 and sd 2
gamma <- dist_spec(
 sd = 3, sd = 2, sd_sd = 0.5, sd_sd = 0.5, max = 20,
 distribution = "gamma"
)
sd(gamma)

# The sd of the sum of two distributions
sd(lognormal + gamma)
}
\author{
Sebastian Funk
}
