% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{select_plots}
\alias{select_plots}
\title{Internal helper function to select plots from those created by
\code{\link[=report_plots]{report_plots()}}}
\usage{
select_plots(
  plots,
  type = c("summary", "infections", "reports", "R", "growth_rate", "all")
)
}
\arguments{
\item{plots}{A list of plots as produced by \code{\link[=report_plots]{report_plots()}}.}

\item{type}{A character vector indicating the name of the plot to return.
Defaults to  "summary" with supported options being "infections", "reports",
"R", "growth_rate", "summary", "all". If "all" is supplied all plots are
generated.}
}
\value{
Selected plots by type
}
\description{
Internal helper function to select plots from those created by
\code{\link[=report_plots]{report_plots()}}
}
\keyword{internal}
