% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_default_logging}
\alias{setup_default_logging}
\title{Setup Default Logging}
\usage{
setup_default_logging(
  logs = tempdir(),
  mirror_epinow = FALSE,
  mirror_epinow_fit = FALSE,
  target_date = NULL
)
}
\arguments{
\item{logs}{Character path indicating the target folder in which to store log
information. Defaults to the temporary directory.}

\item{mirror_epinow}{Logical, defaults to FALSE. Should internal logging be
returned from \code{epinow} to the console.}

\item{mirror_epinow_fit}{Logical, defaults to FALSE. Should internal logging
be returned from internal fitting functions to the console.}

\item{target_date}{Date, defaults to maximum found in the data if not specified.}
}
\description{
Setup Default Logging
}
\examples{
setup_default_logging()
}
