% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setup_forecast}
\alias{setup_forecast}
\title{Ensure forecast is properly set up}
\usage{
setup_forecast(forecast, horizon = NULL)
}
\arguments{
\item{horizon}{Deprecated; use \code{forecast} instead to specify the predictive
horizon}
}
\description{
Ensure forecast is properly set up
}
\keyword{internal}
