% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setup_logging}
\alias{setup_logging}
\title{Setup Logging}
\usage{
setup_logging(threshold = "INFO", file = NULL, mirror_to_console = TRUE)
}
\arguments{
\item{threshold}{Character string indicating the logging level see (?futile.logger
for details of the available options). Defaults to "INFO".}

\item{file}{Character string indicating the path to save logs to. By default logs will be
written to the console.}

\item{mirror_to_console}{Logical, defaults to \code{TRUE}. If saving logs to a file should they
also be duplicated in the console.}
}
\value{
Nothing
}
\description{
Setup Logging
}
\examples{

# Set up error only logs (info logs are enabled by default.)
setup_logging("ERROR")
}
