% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_target_folder}
\alias{setup_target_folder}
\title{Setup Target Folder for Saving}
\usage{
setup_target_folder(target_folder = NULL, target_date)
}
\arguments{
\item{target_folder}{Character string specifying where to save results (will
create if not present).}

\item{target_date}{Date, defaults to maximum found in the data if not
specified.}
}
\value{
A list containing the path to the dated folder and the latest folder
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Sets up a folders for saving results
}
