% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stable_convolve}
\alias{stable_convolve}
\title{Numerically stable convolution function for two pmf vectors}
\usage{
stable_convolve(a, b)
}
\arguments{
\item{a}{Numeric vector, the first sequence.}

\item{b}{Numeric vector, the second sequence.}
}
\value{
A numeric vector representing the convolution of \code{a} and \code{b}.
}
\description{
Unlike \code{\link[stats:convolve]{stats::convolve()}}, this function does not use the FFT algorithm,
which can generate negative numbers when below machine precision.
}
\keyword{internal}
