% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{stan_vb_opts}
\alias{stan_vb_opts}
\title{Stan Variational Bayes Options}
\usage{
stan_vb_opts(samples = 2000, trials = 10, iter = 10000, ...)
}
\arguments{
\item{samples}{Numeric, default 2000. Overall number of approximate posterior
samples.}

\item{trials}{Numeric, defaults to 10. Number of attempts to use
rstan::vb()] before failing.}

\item{iter}{Numeric, defaulting to 10000. Number of iterations to use in
\code{\link[rstan:stanmodel-method-vb]{rstan::vb()}}.}

\item{...}{Additional parameters to pass to \code{\link[rstan:stanmodel-method-vb]{rstan::vb()}} or
\code{\link[cmdstanr:model-method-variational]{cmdstanr::variational()}}, depending on the chosen backend.}
}
\value{
A list of arguments to pass to \code{\link[rstan:stanmodel-method-vb]{rstan::vb()}} or
\code{\link[cmdstanr:model-method-variational]{cmdstanr::variational()}}, depending on the chosen backend.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Defines a list specifying the arguments passed to \code{\link[rstan:stanmodel-method-vb]{rstan::vb()}} or
\code{\link[cmdstanr:model-method-variational]{cmdstanr::variational()}}. Custom settings can be supplied which override the
defaults.
}
\examples{
stan_vb_opts(samples = 1000)
}
