% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state_kernel_cov}
\alias{state_kernel_cov}
\title{Gaussian process kernel covariance for prior-predictive state plots}
\usage{
state_kernel_cov(n, alpha, rho, kernel, matern_order)
}
\arguments{
\item{n}{Number of time points.}

\item{alpha}{Gaussian process magnitude.}

\item{rho}{Gaussian process lengthscale.}

\item{kernel}{Kernel type (one of "se", "matern", "ou", "periodic").}

\item{matern_order}{Matern order (used when \code{kernel} is "matern").}
}
\value{
An \code{n} by \code{n} covariance matrix.
}
\description{
Gaussian process kernel covariance for prior-predictive state plots
}
\keyword{internal}
