% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow.R
\name{[[.epinow}
\alias{[[.epinow}
\title{Extract elements from epinow objects with bracket notation}
\usage{
\method{[[}{epinow}(x, i)
}
\arguments{
\item{x}{An \code{epinow} object}

\item{i}{The name or index of the element to extract}
}
\value{
The requested element with a deprecation warning for deprecated
elements
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Provides backward compatibility for bracket-based access to deprecated
elements. See \link{$.epinow} for details on the deprecation.
}
\keyword{internal}
