% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_cast}
\alias{summarise_cast}
\title{Summarise a nowcast}
\usage{
summarise_cast(nowcast)
}
\arguments{
\item{nowcast}{A dataframe as produced by \code{nowcast_pipeline}}
}
\value{
A summarised dataframe
}
\description{
Summarise a nowcast
}
