% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_key_measures}
\alias{summarise_key_measures}
\title{Summarise rt and cases}
\usage{
summarise_key_measures(
  regional_results = NULL,
  results_dir = NULL,
  summary_dir = NULL,
  type = "region",
  date = "latest"
)
}
\arguments{
\item{regional_results}{A list of dataframes as produced by
\code{get_regional_results}}

\item{results_dir}{Character string indicating the directory from which to
extract results.}

\item{summary_dir}{Character string the directory into which to save results
as a csv.}

\item{type}{Character string, the region identifier to apply (defaults to
region).}

\item{date}{A Character string (in the format "yyyy-mm-dd") indicating the
date to extract data for. Defaults to "latest" which finds the latest
results available.}
}
\value{
A list of summarised Rt, cases by date of infection and cases by
date of report
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
Produces summarised \verb{<data.frame>}s of output across regions.
Used internally by \code{regional_summary}.
}
\seealso{
regional_summary
}
