% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise Realtime Results}
\usage{
summarise_results(
  regions,
  summaries,
  results_dir,
  target_date,
  region_scale = "Region"
)
}
\arguments{
\item{regions}{An character string containing the list of regions to extract results for
(must all have results for the same target date).}

\item{summaries}{A list of summary data frames as output by \code{epinow}}

\item{results_dir}{An optional character string indicating the location of the results directory to extract results
from.}

\item{target_date}{A character string indicating the target date to extract results for. All regions must have results
for this date.}

\item{region_scale}{A character string indicating the name to give the regions being summarised.}
}
\value{
A list of summary data
}
\description{
Summarise Realtime Results
}
\examples{
\dontrun{
# see ?regional_epinow for code to generate regional results to use with this example.

region_sum_tab <- summarise_results(regions = regions,
                                    summaries = purrr::map(out, ~ .$summary))
                  
region_sum_tab
}

}
