% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_to_csv}
\alias{summarise_to_csv}
\title{Summarise rt and cases as a csv}
\usage{
summarise_to_csv(
  results_dir = NULL,
  summary_dir = NULL,
  type = "country",
  date = NULL
)
}
\arguments{
\item{results_dir}{Character string indicating the directory from which to extract results}

\item{summary_dir}{Character string the directory into which to save results}

\item{type}{Character string, the region identifier to apply}

\item{date}{A Character string (in the format "yyyy-mm-dd") indicating the date to extract
data for. Defaults to "latest" which finds the latest results available.}
}
\description{
Summarise rt and cases as a csv
}
