% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_var}
\alias{summarise_var}
\title{Summarise a variable}
\usage{
summarise_var(var, by_date = TRUE)
}
\arguments{
\item{nowcast}{A dataframe as produced by \code{nowcast_pipeline}}
}
\value{
A summarised dataframe
}
\description{
Summarise a variable
}
