% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary.estimate_dist}
\alias{summary.estimate_dist}
\title{Summarise results from estimate_dist}
\usage{
\method{summary}{estimate_dist}(object, CrIs = c(0.2, 0.5, 0.9), ...)
}
\arguments{
\item{object}{A fitted model object from \code{estimate_dist()}}

\item{CrIs}{Numeric vector of credible intervals to calculate.}

\item{...}{Additional arguments (currently unused)}
}
\value{
A \code{<data.table>} with summary statistics for the delay
distribution parameters.
}
\description{
\code{r lifecycle::badge("experimental")}
Returns parameter summary statistics for the fitted delay
distribution model.
}
