% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary.estimate_truncation}
\alias{summary.estimate_truncation}
\title{Summarise results from estimate_truncation}
\usage{
\method{summary}{estimate_truncation}(object, CrIs = c(0.2, 0.5, 0.9), ...)
}
\arguments{
\item{object}{A fitted model object from \code{estimate_truncation()}}

\item{CrIs}{Numeric vector of credible intervals to calculate.}

\item{...}{Additional arguments (currently unused)}
}
\value{
A \verb{<data.table>} with summary statistics for the truncation
distribution parameters.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns parameter summary statistics for the fitted truncation model.
}
