% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{test_data_complete}
\alias{test_data_complete}
\title{Check if data has either explicit NA values or implicit missing dates.}
\usage{
test_data_complete(data, cols_to_check)
}
\arguments{
\item{data}{The data to be checked}

\item{cols_to_check}{A character vector of the columns to check}
}
\value{
\code{TRUE} if data is complete, else if data has implicit or explicit
missingness, \code{FALSE}.
}
\description{
Check if data has either explicit NA values or implicit missing dates.
}
\keyword{internal}
