% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{to_stan.dist_spec}
\alias{to_stan.dist_spec}
\title{Converts a \code{dist_spec} object for use in stan}
\usage{
\method{to_stan}{dist_spec}(x, name = NULL)
}
\arguments{
\item{x}{The \code{\link[=dist_spec]{dist_spec()}} to use}

\item{name}{A name to give to the returned objects. IF given This will be
prepended to any of the variable names in the returned list, separated
by an underscore (default: NULL, in which case nothing will be prepended).}
}
\value{
A list to be used as stan data.
}
\description{
Converts a \code{dist_spec} object for use in stan
}
\author{
Sebastian Funk
}
