% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{to_stan}
\alias{to_stan}
\title{Converts R objects to ones that can be passed to stan as data.}
\usage{
to_stan(x, ...)
}
\arguments{
\item{x}{An R object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Converts R objects to ones that can be passed to stan as data.
}
