% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{tune_inv_gamma}
\alias{tune_inv_gamma}
\title{Tune an Inverse Gamma to Achieve the Target Truncation}
\usage{
tune_inv_gamma(lower = 2, upper = 21)
}
\arguments{
\item{lower}{Numeric, defaults to 2. Lower truncation bound.}

\item{upper}{Numeric, defaults to 21. Upper truncation bound.}
}
\value{
A list of alpha and beta values that describe a inverse gamma
distribution that achieves the target truncation.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}
Allows an inverse gamma distribution to be. tuned so that less than 0.01 of its
probability mass function falls outside of the specified
bounds. This is required when using an inverse gamma prior, for example for a
Gaussian process. As no inverse gamma priors are currently in use and this function
has some stability issues it may be deprecated at a later date.
}
\examples{

tune_inv_gamma(lower = 2, upper = 21)
}
