% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{update_defaults}
\alias{update_defaults}
\title{Update Default Settings}
\usage{
update_defaults(defaults = list(), optional = list())
}
\arguments{
\item{defaults}{A list of default settings}

\item{optional}{A list of optional settings to override defaults}
}
\value{
A list
}
\description{
\lifecycle{stable}
Used internally to handle updating default settings in a list.
}
