% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{allocate_delays}
\alias{allocate_delays}
\title{Allocate Delays into Required Stan Format}
\usage{
allocate_delays(delay_var, no_delays)
}
\arguments{
\item{delay_var}{List of numeric delays}

\item{no_delays}{Numeric, number of delays}
}
\value{
A numeric array
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
Allocate delays for stan. Used in \code{delay_opts}.
}
