% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{allocate_empty}
\alias{allocate_empty}
\title{Allocate Empty Parameters to a List}
\usage{
allocate_empty(data, params, n = 0)
}
\arguments{
\item{data}{A list of parameters}

\item{params}{A character vector of parameters to allocate to
empty if missing.}

\item{n}{Numeric, number of samples to assign an empty array}
}
\value{
A list of parameters some allocated to be empty
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
Allocate missing parameters to be empty two dimensional arrays. Used
internally by \code{simulate_infections.}
}
\examples{
data <- list(x = 1, y = 2, z = 30)
EpiNow2:::allocate_empty(data, params = c("x", "t"))
}
