% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{allocate_empty}
\alias{allocate_empty}
\title{Allocate Empty Parameters to a List}
\usage{
allocate_empty(data, params)
}
\arguments{
\item{data}{A list of parameters}

\item{params}{A character vector of params to allocate to
empty if missing.}
}
\value{
A list of parameters some allocated to be emtpy
}
\description{
Allocate Empty Parameters to a List
}
\examples{
data <- list(x = 1, y = 2, z = 30)
EpiNow2:::allocate_empty(data, params = c("x", "t"))
}
