% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{apply_default_cdf_max}
\alias{apply_default_cdf_max}
\title{Apply default CDF level to a <dist_spec> if it is unconstrained}
\usage{
apply_default_cdf_max(dist, default_cdf_max, cdf_max_set)
}
\arguments{
\item{dist}{A <dist_spec>}

\item{default_cdf_max}{Numeric; default CDF level to keep the distribution up
to if an unconstrained distribution is passed as \code{dist}. If \code{dist} is
already constrained by having a maximum or CDF level this is ignored. Note
that this can only be done for <dist_spec> objects with fixed parameters.}

\item{cdf_max_set}{Logical; whether the default CDF level has been set by
the user; if yes and \code{dist} is constrained a warning is issued}
}
\value{
A <dist_spec> with the default CDF level set if previously not
constrained
}
\description{
Apply default CDF level to a <dist_spec> if it is unconstrained
}
\keyword{internal}
