% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{bootstrapped_dist_fit}
\alias{bootstrapped_dist_fit}
\title{Fit a Subsampled Boostrap to Integer Values and Summarise Distribution Parameters}
\usage{
bootstrapped_dist_fit(
  values,
  verbose = FALSE,
  samples = 2000,
  bootstraps = 10,
  bootstrap_samples = 250,
  dist = "lognormal"
)
}
\arguments{
\item{values}{Numeric vector of integer values.}

\item{verbose}{Logical, defaults to \code{FALSE}. Should progress messages be printed}

\item{samples}{Numeric, number of samples to take overall from the bootstrapped posteriors.}

\item{bootstraps}{Numeric, defaults to 1. The number of bootstrap samples (with replacement)
of the delay distribution to take.}

\item{bootstrap_samples}{Numeric, defaults to 100. The number of samples to take in each boostrap.
When the sample size of the supplied delay distribution is less than 100 this is used instead.}
}
\value{
A list summarising the bootstrapped distribution
}
\description{
Fit a Subsampled Boostrap to Integer Values and Summarise Distribution Parameters
}
\examples{

\dontrun{
# lognormal
delays <- rlnorm(50, log(5), 1)

out <- bootstrapped_dist_fit(delays, samples = 2, bootstraps = 10, 
                             dist = "lognormal")

## Inspect
out

}
}
