% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_gp_ls_prior}
\alias{check_gp_ls_prior}
\title{Check the approximate GP settings can be chosen from a lengthscale prior}
\usage{
check_gp_ls_prior(ls)
}
\arguments{
\item{ls}{A \verb{<dist_spec>} giving the lengthscale prior.}
}
\value{
Invisibly, the name of the prior distribution. Called for its side
effect of erroring for unsupported priors.
}
\description{
The settings can be chosen from fixed, lognormal, gamma and normal priors
without uncertain parameters.
}
\keyword{internal}
