% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_na_setting_against_data}
\alias{check_na_setting_against_data}
\title{Check the na settings in obs_opts() with the data and throw messages
where necessary}
\usage{
check_na_setting_against_data(data, obs)
}
\arguments{
\item{data}{The raw data}

\item{obs}{A call to \code{\link[=obs_opts]{obs_opts()}}}
}
\value{
Called for its side effects
}
\description{
Check the na settings in obs_opts() with the data and throw messages
where necessary
}
\details{
This function checks the data to see if it is complete or not and then
checks if the user specified na to be treated as missing. If the latter is
false, it does nothing. If TRUE, it informs the user about how the implicit
or explicit missingness is treated. This function is necessary because
the data argument and observation model do not interact internally.
}
\keyword{internal}
