% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_reports_valid}
\alias{check_reports_valid}
\title{Validate data input}
\usage{
check_reports_valid(reports, model)
}
\arguments{
\item{reports}{A data frame with either:
\itemize{
\item a minimum of two columns: date and confirm, if to be
used by \code{\link[=estimate_infection]{estimate_infection()}} or \code{\link[=estimate_truncation]{estimate_truncation()}}, or
\item a minimum of three columns: date, primary, and secondary, if to be used by
\code{\link[=estimate_secondary]{estimate_secondary()}}.
}}

\item{model}{The EpiNow2 model to be used. Either
"estimate_infections", "estimate_truncation", or "estimate_secondary".
This is used to determine which checks to perform on the data input.}
}
\value{
Called for its side effects.
}
\description{
\code{check_reports_valid()} checks that the supplied data is a \verb{<data.frame>},
and that it has the right column names and types. In particular, it checks
that the date column is in date format and does not contain NA's, and that
the other columns are numeric.
}
\author{
James M. Azam
}
\keyword{internal}
