% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{combine_tv_and_static_params}
\alias{combine_tv_and_static_params}
\title{Combine time-varying and static parameters}
\usage{
combine_tv_and_static_params(time_varying_list, raw_samples, args)
}
\arguments{
\item{time_varying_list}{Named list of time-varying parameter data.tables}

\item{raw_samples}{Raw samples from extract_samples()}

\item{args}{Model arguments containing delay and parameter specifications}
}
\value{
A \code{data.table} combining all parameters with variable column
}
\description{
Internal helper that combines time-varying parameters (which
get their variable name from list keys) with static parameters (which
already have a variable column from extractors).
}
\keyword{internal}
