% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_logmean}
\alias{convert_to_logmean}
\title{Convert mean and sd to log mean for a log normal distribution}
\usage{
convert_to_logmean(mean, sd)
}
\arguments{
\item{mean}{Numeric, mean of a distribution}

\item{sd}{Numeric, standard deviation of a distribution}
}
\value{
The log mean of a lognormal distribution
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Convert from mean and standard deviation to the log mean of the
lognormal distribution. Useful for defining distributions supported by
\code{estimate_infections}, \code{epinow}, and \code{regional_epinow}.
}
\examples{

convert_to_logmean(2, 1)
}
