% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_logsd}
\alias{convert_to_logsd}
\title{Convert mean and sd to log standard deviation for a log normal distribution}
\usage{
convert_to_logsd(mean, sd)
}
\arguments{
\item{mean}{Numeric, mean of a distribution}

\item{sd}{Numeric, standard deviation of a distribution}
}
\value{
The log standard deviation of a lognormal distribution
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
Convert from mean and standard deviation to the log standard deviation of the
lognormal distribution. Useful for defining distributions supported by
\code{estimate_infections}, \code{epinow}, and \code{regional_epinow}.
}
\examples{

convert_to_logsd(2, 1)
}
