% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{copy_results_to_latest}
\alias{copy_results_to_latest}
\title{Copy Results From Dated Folder to Latest}
\usage{
copy_results_to_latest(target_folder = NULL, latest_folder = NULL)
}
\arguments{
\item{target_folder}{Character string specifying where to save results (will create if not present).}

\item{latest_folder}{Character string containing the path to the latest target folder.
As produced by \code{setup_target_folder}.}
}
\value{
No return value, called for side effects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}
Copies output from the dated folder to a latest folder. May be undergo changes in later releases.
}
