% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{country_map}
\alias{country_map}
\title{Generate a country map for a single variable.}
\usage{
country_map(
  data = NULL,
  country = NULL,
  variable = NULL,
  variable_label = NULL,
  trans = "identity",
  fill_labels = NULL,
  scale_fill = NULL,
  ...
)
}
\arguments{
\item{data}{Dataframe containing variables to be mapped. Must contain a \code{region_code} variable.}

\item{country}{Character string indicating the name of the country to be mapped.}

\item{variable}{A character string indicating the variable to map data for. This must be supplied.}

\item{variable_label}{A character string indicating the variable label to use. If not supplied then the underlying
variable name is used.}

\item{trans}{A character string specifying the transform to use on the specified metric. Defaults to no
transform ("identity"). Other options include log scaling ("log") and log base 10 scaling
("log10"). For a complete list of options see \code{ggplot2::continous_scale}.}

\item{fill_labels}{A function to use to allocate legend labels. An example (used below) is \code{scales::percent},
which can be used for percentage data.}

\item{scale_fill}{Function to use for scaling the fill. Defaults to a custom \code{ggplot2::scale_fill_manual}}

\item{...}{Additional arguments passed to the \code{scale_fill} function}
}
\value{
A \code{ggplot2} object containing a country map.
}
\description{
This general purpose function can be used to generate a country map for a single variable. It has few defaults but
the data supplied must contain a \code{region_code} variable for linking to mapping data. This function requires
the installation of the \code{rnaturalearth} package.
}
\examples{

country_map
}
