% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_clean_reported_cases}
\alias{create_clean_reported_cases}
\title{Create Clean Reported Cases}
\usage{
create_clean_reported_cases(reported_cases, horizon)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date). confirm must be integer and date must be
in date format.}

\item{horizon}{Numeric, defaults to 7. Number of days into the future to forecast.}
}
\value{
A cleaned data frame of reported cases
}
\description{
Create Clean Reported Cases
}
