% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_complete_cases}
\alias{create_complete_cases}
\title{Create complete cases}
\usage{
create_complete_cases(cases)
}
\arguments{
\item{cases}{data frame with a column "confirm" that may contain NA values}
}
\value{
A data frame without NA values, with two columns: confirm (number)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Creates a complete data set without NA values and appropriate indices
}
\keyword{internal}
