% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_delay_inits}
\alias{create_delay_inits}
\title{Create initial conditions for delays}
\usage{
create_delay_inits(data)
}
\arguments{
\item{data}{A list of data as produced by \code{create_stan_data.}}
}
\value{
A list of initial conditions for delays
}
\description{
Create initial conditions for delays
}
\author{
Sebastian Funk
}
\keyword{internal}
