% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_future_rt}
\alias{create_future_rt}
\title{Construct the Required Future Rt assumption}
\usage{
create_future_rt(future_rt = "latest", delay = 0)
}
\arguments{
\item{future_rt}{A character string or integer. This argument indicates how to set future Rt values. Supported
options are to project using the Rt model ("project"), to use the latest estimate based on partial data ("latest"),
to use the latest estimate based on data that is over 50\% complete ("estimate"). If an integer is supplied then the Rt estimate
from this many days into the future (or past if negative) past will be used forwards in time.}

\item{delay}{Numeric mean delay}
}
\value{
A list containing a logical called fixed and an integer called from
}
\description{
Construct the Required Future Rt assumption
}
