% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_gp_data}
\alias{create_gp_data}
\title{Create Gaussian Process Stan Data}
\usage{
create_gp_data(gp = list(), data)
}
\arguments{
\item{gp}{A list of settings to override the package default Gaussian
process settings. See the documentation for \code{gp_settings} for details of these
settings and \code{gp_settings()} for the current defaults.}

\item{data}{A list containing the following numeric values: \code{t}, \code{seeding_time},
\code{horizon}.}
}
\value{
A list of settings defining the Gaussian process
}
\description{
Create Gaussian Process Stan Data
}
\examples{
# define input data required
data <- list(
    t = 30,
    seeding_time = 7,
    horizon = 7)

# default gaussian process data     
create_gp_data(data = data)

# settings when no gaussian process is desired
create_gp_data(gp = NULL, data)

# custom lengthscale
create_gp_data(gp = list(ls_mean = 14), data)
}
\seealso{
gp_settings
}
