% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_initial_conditions}
\alias{create_initial_conditions}
\title{Create Initial Conditions Generating Function}
\usage{
create_initial_conditions(stan_data, params)
}
\arguments{
\item{stan_data}{A list of data as produced by \code{\link[=create_stan_data]{create_stan_data()}}.}

\item{params}{A list of \verb{<EpiNow2.params>} as created by \code{\link[=make_param]{make_param()}}}
}
\value{
An initial condition generating function
}
\description{
Uses the output of \code{\link[=create_stan_data]{create_stan_data()}} to create a function which can be
used to sample from the prior distributions (or as close as possible) for
parameters. Used in order to initialise each stan chain within a range of
plausible values.
}
\details{
\code{R_mean} is seeded from the initial-Rt prior carried in \code{stan_data} by
\code{\link[=make_init_priors]{make_init_priors()}}, so chains start near the configured reproduction
number; this is a stopgap until derived-prior parameters are initialised
through the shared path (#1481). The distribution code follows
\code{\link[=pack_init_prior]{pack_init_prior()}} (0: lognormal, 1: gamma, 2: normal).
}
\keyword{internal}
