% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_initial_conditions}
\alias{create_initial_conditions}
\title{Create Initial Conditions Generating Function}
\usage{
create_initial_conditions(stan_data, params, rt_prior = NULL)
}
\arguments{
\item{stan_data}{A list of data as produced by \code{\link[=create_stan_data]{create_stan_data()}}.}

\item{params}{A list of \verb{<EpiNow2.params>} as created by \code{\link[=make_param]{make_param()}}}

\item{rt_prior}{A \verb{<dist_spec>} giving the prior on the initial
reproduction number, used to seed \code{R_mean}. \code{NULL} when the reproduction
number is not estimated.}
}
\value{
An initial condition generating function
}
\description{
Uses the output of \code{\link[=create_stan_data]{create_stan_data()}} to create a function which can be
used to sample from the prior distributions (or as close as possible) for
parameters. Used in order to initialise each stan chain within a range of
plausible values.
}
\keyword{internal}
