% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_initial_conditions}
\alias{create_initial_conditions}
\title{Create Initial Conditions Generating Function}
\usage{
create_initial_conditions(data)
}
\arguments{
\item{data}{A list of data as produced by \code{create_stan_data.}}
}
\value{
An initial condition generating function
}
\description{
Create Initial Conditions Generating Function
}
