% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_initial_conditions}
\alias{create_initial_conditions}
\title{Create Initial Conditions Generating Function}
\usage{
create_initial_conditions(data)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=create_stan_data]{create_stan_data()}}.}
}
\value{
An initial condition generating function
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Uses the output of \code{\link[=create_stan_data]{create_stan_data()}} to create a function which can be
used to sample from the prior distributions (or as close as possible) for
parameters. Used in order to initialise each \code{{stan}} chain within a range of
plausible values.
}
