% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_obs_model}
\alias{create_obs_model}
\title{Create Observation Model Settings}
\usage{
create_obs_model(obs = obs_opts())
}
\arguments{
\item{obs}{A list of options as generated by \code{obs_opts} defining the
observation model. Defaults to \code{obs_opts()}.}
}
\value{
A list of settings ready to be passed to stan defining
the Observation Model
}
\description{
\lifecycle{stable}
Takes the output of \code{obs_opts} and convrts it into a list understood by \code{stan}.
}
\examples{
# default observation model data
create_obs_model()

# Poisson observation model
create_obs_model(obs_opts(family = "poisson"))

# Applying a observation scaling to the data
create_obs_model(obs_opts(scale = list(mean = 0.4, sd = 0.01)))
}
\seealso{
obs_opts
}
