% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_rt_data}
\alias{create_rt_data}
\title{Create Time-varying Reproduction Number Data}
\usage{
create_rt_data(rt = list(), breakpoints = NULL)
}
\arguments{
\item{rt}{A list of settings to override the package default time-varying reproduction
number settings. See the documentation for \code{?rt_settings} for details of these settings
and \code{rt_settings()} for the current defaults. Set to \code{NULL} to switch to using
back calculation rather than generating infections using Rt.}

\item{breakpoints}{An integer vector (binary) indicating the location of breakpoints.}
}
\value{
A list of settings defining the time-varying reproduction number
}
\description{
Create Time-varying Reproduction Number Data
}
\examples{
# default Rt data     
create_rt_data()

# settings when no Rt is desired
create_rt_data(rt = NULL)

# using breakpoints
create_rt_data(rt = list(use_breakpoints = TRUE), breakpoints = rep(1, 10))
}
\seealso{
rt_settings
}
