% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_rt_data}
\alias{create_rt_data}
\title{Create Time-varying Reproduction Number Data}
\usage{
create_rt_data(rt = rt_opts(), breakpoints = NULL, delay = 0, horizon = 0)
}
\arguments{
\item{rt}{A list of options as generated by \code{rt_opts} defining Rt estimation.
Defaults to \code{rt_opts()}. Set to \code{NULL} to switch to using  back calculation
rather than generating infections using Rt.}

\item{breakpoints}{An integer vector (binary) indicating the location of breakpoints.}

\item{delay}{Numeric mean delay}

\item{horizon}{Numeric, forecast horizon.}
}
\value{
A list of settings defining the time-varying reproduction number
}
\description{
\lifecycle{stable}
Takes the output from \code{rt_opts} and converts it into a list understood by
\code{stan}.
}
\examples{
# default Rt data     
create_rt_data()

# settings when no Rt is desired
create_rt_data(rt = NULL)

# using breakpoints
create_rt_data(rt_opts(use_breakpoints = TRUE), breakpoints = rep(1, 10))
}
\seealso{
rt_settings
}
