% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_shifted_cases}
\alias{create_shifted_cases}
\title{Create a Data Frame of Mean Delay Shifted Cases}
\usage{
create_shifted_cases(reported_cases, shift, smoothing_window, horizon)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date). confirm must be integer and date must be
in date format.}

\item{shift}{Numeric, mean delay shift to apply.}

\item{smoothing_window}{Numeric, the rolling average smoothing window
to apply.}

\item{horizon}{Numeric, defaults to 7. Number of days into the future to forecast.}
}
\value{
A dataframe for shifted reported cases
}
\description{
Create a Data Frame of Mean Delay Shifted Cases
}
\examples{
create_shifted_cases(example_confirmed, 7, 14, 7)
}
