% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_shifted_cases}
\alias{create_shifted_cases}
\title{Create a Data Frame of Mean Delay Shifted Cases}
\usage{
create_shifted_cases(
  reported_cases,
  mean_shift,
  prior_smoothing_window,
  horizon
)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date). confirm must be integer and date must be
in date format.}

\item{mean_shift}{Numeric, mean delay shift}

\item{prior_smoothing_window}{Numeric defaults to 7. The number of days over which to take a rolling average
for the prior based on reported cases.}

\item{horizon}{Numeric, defaults to 7. Number of days into the future to forecast.}
}
\value{
A dataframe for shifted reported cases
}
\description{
Create a Data Frame of Mean Delay Shifted Cases
}
