% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_stan_args}
\alias{create_stan_args}
\title{Create a List of Stan Arguments}
\usage{
create_stan_args(
  stan_args = list(),
  data = NULL,
  init = "random",
  samples = 1000,
  verbose = FALSE
)
}
\arguments{
\item{stan_args}{A list of settings to override the package default stan settings. See the documentation
for \code{create_stan_args} for details and \code{create_stan_args()} for current defaults.}

\item{data}{A list of stan data as created by \code{create_stan_data}}

\item{init}{Initial conditions passed to \code{rstan}. Defaults to "random" but can also be a function (
as supplied by \code{create_intitial_conditions}).}

\item{samples}{Numeric, defaults to 1000. The overall number of posterior samples to return (Note: not the
number of samples per chain as is the default in stan).}

\item{verbose}{Logical, defaults to \code{FALSE}. Should verbose progress messages be returned.}
}
\value{
A list of stan arguments
}
\description{
Generates a list of arguments as required by \code{rstan::sampling} (when \verb{stan_args$method = "sampling})
or \code{rstan::vb} (when \verb{stan_args$method = "vb}). See \code{create_stan_args()} for the defaults and the relevant \code{rstan}
functions for additional options. Defaults can be overwritten by passing a list containing the new setting to
\code{stan_args}.
}
\examples{
# default settings
create_stan_args()

# increasing warmup
create_stan_args(stan_args = list(warmup = 1000))
}
