% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Create Stan Data Required for estimate_infections}
\usage{
create_stan_data(
  reported_cases,
  shifted_reported_cases,
  horizon,
  no_delays,
  mean_shift,
  generation_time,
  rt,
  gp,
  obs,
  delays
)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date). confirm must be integer and date must be
in date format.}

\item{shifted_reported_cases}{A dataframe of delay shifted reported cases}

\item{horizon}{Numeric, forecast horizon.}

\item{no_delays}{Numeric, number of delays}

\item{mean_shift}{Numeric, mean delay shift}

\item{generation_time}{A list containing the mean, standard deviation of the mean (mean_sd),
standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
generation time (assuming a gamma distribution).}

\item{rt}{A list of options as generated by \code{rt_opts} defining Rt estimation.
Defaults to \code{rt_opts()}. Set to \code{NULL} to switch to using  back calculation
rather than generating infections using Rt.}

\item{gp}{A list of options as generated by \code{gp_opts} to define the
Gaussian process. Defaults to \code{gp_opts()}.Set to NULL to disable the
Gaussian process.}

\item{obs}{A list of options as generated by \code{obs_opts} defining the
observation model. Defaults to \code{obs_opts()}.}

\item{delays}{A list of delays (i.e incubation period/reporting delay) between infection and report.
Each list entry must also be a list containing the mean, standard deviation of the mean (mean_sd),
standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
that delay (assuming a lognormal distribution with all parameters excepting the max allowed value
on the log scale). To use no delays set this to \code{list()}.}
}
\value{
A list of stan data
}
\description{
Create Stan Data Required for estimate_infections
}
