% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Create Stan Data Required for estimate_infections}
\usage{
create_stan_data(
  reported_cases,
  shifted_reported_cases,
  horizon,
  no_delays,
  mean_shift,
  generation_time,
  rt_prior,
  estimate_rt,
  estimate_week_eff,
  stationary,
  fixed,
  break_no,
  fixed_future_rt,
  gp,
  family
)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date). confirm must be integer and date must be
in date format.}

\item{shifted_reported_cases}{A dataframe of delay shifted reported cases}

\item{horizon}{Numeric, defaults to 7. Number of days into the future to forecast.}

\item{no_delays}{Numeric, number of delays}

\item{mean_shift}{Numeric, mean delay shift}

\item{generation_time}{A list containing the mean, standard deviation of the mean (mean_sd),
standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
generation time (assuming a gamma distribution).}

\item{rt_prior}{A list contain the mean and standard deviation (sd) of the gamma distributed prior for
Rt. By default this is assumed to be mean 1 with a standard deviation of 1.}

\item{estimate_rt}{Logical, defaults TRUE. Should Rt be estimated when imputing infections.}

\item{estimate_week_eff}{Logical, defaults TRUE. Should weekly reporting effects be estimated.}

\item{stationary}{Logical, defaults to FALSE. Should Rt be estimated with a global mean. When estimating Rt
this should substantially improve run times but will revert to the global average for real time and forecasted estimates.
This setting is most appropriate when estimating historic Rt or when combined with breakpoints.}

\item{fixed}{Logical, defaults to FALSE. If TRUE then a Gaussian process is not used and Rt is assumed to be constant over time
(apart from any manually included breakpoints). If \code{estimate_rt} is FALSE then this reduces the backcalculation to a simple mean shift.
This option can be used to produce a null model estimate, to produce a single Rt estimate for a short timeseries or as part of a wider
analysis on the impact of interventions.}

\item{break_no}{Numeric, number of breakpoints}

\item{fixed_future_rt}{Logical, defaults to FALSE. IF TRUE then the estimated Rt from the last time point with data is used for all
future time points without data.}

\item{gp}{List controlling the Gaussian process approximation. Must contain
the \code{basis_prop} (number of basis functions based on scaling the time points) which defaults to 0.3 and must be
between 0 and 1 (increasing this increases the accuracy of the approximation and the cost of additional compute.
Must also contain the \code{boundary_scale} (multiplied by half the range of the input time series). Increasing this
increases the accuracy of the approximation at the cost of additional compute.
See here: https://arxiv.org/abs/2004.11408 for more information on setting these parameters.
Can optionally also contain the  \code{lengthscale_mean} and \code{lengthscale_sd}. If these are specified this will override
the defaults of 0 and 2 (normal distributed truncated at zero).}

\item{family}{A character string indicating the reporting model to use. Defaults to negative
binomial ("negbin") with poisson ("poisson") also supported.}
}
\value{
A list of stan data
}
\description{
Create Stan Data Required for estimate_infections
}
