% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Create Stan Data Required for estimate_infections}
\usage{
create_stan_data(
  reported_cases,
  shifted_reported_cases,
  horizon,
  no_delays,
  mean_shift,
  generation_time,
  rt,
  gp,
  obs_model,
  delays
)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date). confirm must be integer and date must be
in date format.}

\item{shifted_reported_cases}{A dataframe of delay shifted reported cases}

\item{horizon}{Numeric, defaults to 7. Number of days into the future to forecast.}

\item{no_delays}{Numeric, number of delays}

\item{mean_shift}{Numeric, mean delay shift}

\item{generation_time}{A list containing the mean, standard deviation of the mean (mean_sd),
standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
generation time (assuming a gamma distribution).}

\item{rt}{Logical, defaults to \code{TRUE}. Is Rt being estimated? This controls Rt
specific Gaussian process settings that are not supported for back calculation.}

\item{gp}{A list of settings to override the package default Gaussian
process settings. See the documentation for \code{gp_settings} for details of these
settings and \code{gp_settings()} for the current defaults.}

\item{obs_model}{A list of settings to override the package default observation
model settings. See the documentation for \code{obs_model_settings} for details of these
settings and \code{obs_model_settings()} for the current defaults.}

\item{delays}{A list of delays (i.e incubation period/reporting delay) between infection and report.
Each list entry must also be a list containing the mean, standard deviation of the mean (mean_sd),
standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
that delay (assuming a lognormal distribution with all parameters excepting the max allowed value
on the log scale). To use no delays set this to \code{list()}.}

\item{break_no}{Numeric, number of breakpoints}

\item{estimate_rt}{Logical, should Rt be estimated.}
}
\value{
A list of stan data
}
\description{
Create Stan Data Required for estimate_infections
}
