% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_stan_delays}
\alias{create_stan_delays}
\title{Create delay variables for stan}
\usage{
create_stan_delays(..., ot)
}
\arguments{
\item{...}{Named delay distributions specified using \code{dist_spec()}.
The names are assigned to IDs}

\item{ot}{Integer, number of observations (needed if weighing any priors)
with the number of observations}
}
\value{
A list of variables as expected by the stan model
}
\description{
Create delay variables for stan
}
\author{
Sebastian Funk
}
