% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_stan_params}
\alias{create_stan_params}
\title{Create parameters for stan}
\usage{
create_stan_params(params)
}
\arguments{
\item{params}{A list of \verb{<EpiNow2.params>} as created by \code{\link[=make_param]{make_param()}}}
}
\value{
A list of variables as expected by the stan model
}
\description{
Create parameters for stan
}
\keyword{internal}
