% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_stan_params}
\alias{create_stan_params}
\title{Create parameters for stan}
\usage{
create_stan_params(..., lower_bounds = NULL)
}
\arguments{
\item{...}{Named delay distributions. The names are assigned to IDs}

\item{lower_bounds}{Named vector of lower bounds for any delay(s). The names
have to correspond to the names given to the delay distributions passed.
If \code{NULL} (default) no parameters are given a lower bound.}
}
\value{
A list of variables as expected by the stan model
}
\description{
Create parameters for stan
}
\keyword{internal}
