% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{deconvolution_opts}
\alias{deconvolution_opts}
\title{Deconvolution Model Specification}
\usage{
deconvolution_opts(backcalc = backcalc_opts())
}
\arguments{
\item{backcalc}{A call to \code{\link[=backcalc_opts]{backcalc_opts()}} defining the deconvolution
settings. Defaults to \code{\link[=backcalc_opts]{backcalc_opts()}}.}
}
\value{
A \verb{<deconvolution_opts>} object that can be passed to the \code{model}
argument of \code{\link[=estimate_infections]{estimate_infections()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Specifies that infections should be estimated using the deconvolution
(non-mechanistic) model. This model directly estimates infections on the log
scale without using the renewal equation, and therefore does not require a
generation time distribution.
}
\examples{
# default settings
deconvolution_opts()

# with custom prior window
deconvolution_opts(backcalc = backcalc_opts(prior_window = 21))
}
\seealso{
\code{\link[=renewal_opts]{renewal_opts()}} \code{\link[=estimate_infections]{estimate_infections()}} \code{\link[=backcalc_opts]{backcalc_opts()}}
}
