% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{delay_opts}
\alias{delay_opts}
\title{Delay Distribution Options}
\usage{
delay_opts(
  ...,
  mean = NULL,
  sd = NULL,
  seed,
  type = "uncertain",
  cache = FALSE
)
}
\arguments{
\item{...}{Delay distributions as a list with the following parameters:
"mean", "mean_sd", "sd_mean", "sd_sd", and "max" defining a truncated log
normal (with all parameters except for max defined in logged form).}

\item{mean}{Options defining the variation in the mean of the distribution.
Supply one per delay. Currently only supports \code{gp_opts()}.}

\item{sd}{Options defining the variation in the sd of the distribution.
Supply one per delay. Currently only supports \code{gp_opts()}.}

\item{seed}{Integer, number of timepoints to use to seed the model.
If not set estimated from the combined mean of the specified distributions.}

\item{type}{A character string, defaulting to "uncertain". Defines the type
of distributions to use. Supported options are "uncertain".}

\item{cache}{Logical, defaults to \code{FALSE}. Should a cache be used to only
evaluate PMFs when a new combination of distribution parameters is observed?
This setting may be useful when distributions are time-varying but discrete
in time (for example when a gaussian process is used with a step size
greater than 1).}
}
\value{
A list summarising the input delay distributions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns delay distributions formatted for usage by downstream
functions.
}
\examples{
# no delays
delay_opts()

# a static delay
delay_opts(list(mean = 1, mean_sd = 0.1, sd = 0.1, sd_sd = 0.01, max = 30))
# a delay assumed to vary over time using as gaussian process
delay_opts(list(mean = 1, mean_sd = 0.1, sd = 0.1, sd_sd = 0.01, max = 30),
           mean = list(gp_opts()), sd = list(gp_opts()))
}
\seealso{
convert_to_logmean convert_to_logsd bootstrapped_dist_fit
}
