% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{delay_opts}
\alias{delay_opts}
\title{Delay Distribution Options}
\usage{
delay_opts(dist = fixed(0), ..., fixed = FALSE)
}
\arguments{
\item{dist}{A delay distribution or series of delay distributions. Default is
an fixed distribution with of 0, \code{\link[=dist_spec]{dist_spec()}}, i.e. no delay.}

\item{...}{deprecated; use \code{dist} instead}

\item{fixed}{deprecated; use \code{dist} instead}
}
\value{
A \verb{<delay_opts>} object summarising the input delay distributions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns delay distributions formatted for usage by downstream
functions.
}
\examples{
# no delays
delay_opts()

# A single delay that has uncertainty
delay <- dist_spec(mean = 1, mean_sd = 0.2, sd = 0.5, sd_sd = 0.1, max = 14)
delay_opts(delay)

# A single delay without uncertainty
delay <- dist_spec(mean = 1, sd = 0.5, max = 14)
delay_opts(delay)

# Multiple delays (in this case twice the same)
delay_opts(delay + delay)
}
\seealso{
\code{\link[=convert_to_logmean]{convert_to_logmean()}} \code{\link[=convert_to_logsd]{convert_to_logsd()}}
\code{\link[=bootstrapped_dist_fit]{bootstrapped_dist_fit()}} \code{\link[=dist_spec]{dist_spec()}}
}
\author{
Sam Abbott

Sebastian Funk
}
