% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{delay_opts}
\alias{delay_opts}
\title{Delay Distribution Options}
\usage{
delay_opts(..., fixed = FALSE)
}
\arguments{
\item{...}{Delay distributions as a list with the following parameters:
"mean", "mean_sd", "sd_mean", "sd_sd", and "max" defining a truncated log
normal (with all parameters except for max defined in logged form).}

\item{fixed}{Logical, defaults to \code{FALSE}. Should reporting delays be treated
as coming from fixed (vs uncertain) distributions. Making this simplification
drastically reduces compute requirements.}
}
\value{
A list summarising the input delay distributions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns delay distributions formatted for usage by downstream
functions.
}
\examples{
# no delays
delay_opts()

# A single delay that has uncertainty
delay <- list(mean = 1, mean_sd = 0.2, sd = 0.5, sd_sd = 0.1, max = 15)
delay_opts(delay)

# A single delay where we override the uncertainty assumption
delay_opts(delay, fixed = TRUE)

# A delay where uncertainty is implict
delay_opts(list(mean = 1, mean_sd = 0, sd = 0.5, sd_sd = 0, max = 15))
}
\seealso{
convert_to_logmean convert_to_logsd bootstrapped_dist_fit
}
