% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{delay_opts}
\alias{delay_opts}
\title{Delay Distribution Options}
\usage{
delay_opts(...)
}
\arguments{
\item{...}{Delay distributions as a list with the following parameters:
"mean", "mean_sd", "sd_mean", "sd_sd", and "max" defining a truncated log
normal (with all parameters except for max defined in logged form).}
}
\value{
A list summarising the input delay distributions.
}
\description{
\lifecycle{stable}
Returns delay distributions formatted for usage by downstream
functions.
}
\examples{
# no delays
delay_opts()
}
\seealso{
convert_to_logmean convert_to_logsd bootstrapped_dist_fit
}
