% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{discrete_pmf}
\alias{discrete_pmf}
\title{Discretised probability mass function}
\usage{
discrete_pmf(
  distribution = c("exp", "gamma", "lognormal", "normal", "fixed"),
  params,
  max_value,
  tolerance,
  width
)
}
\arguments{
\item{distribution}{A character string representing the distribution to be
used (one of "exp", "gamma", "lognormal", "normal" or "fixed")}

\item{params}{A list of parameters values (by name) required for each model.
For the exponential model this is a rate parameter and for the gamma model
this is alpha and beta.}

\item{max_value}{Numeric, the maximum value to allow. Defaults to 120.
Samples outside of this range are resampled.}

\item{tolerance}{Numeric; the desired tolerance level. Any part of the
cumulative distribution function beyond 1 minus this tolerance level is
removed.}

\item{width}{Numeric, the width of each discrete bin.}
}
\value{
A vector representing a probability distribution.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}
This function returns the probability mass function of a discretsied and
truncated distribution defined by distribution type, maximum value and model
parameters.

Each entry of the probability mass function corresponds to the 2-length
interval ending at the entry except for the first interval that covers (0,
1).  That is, the probability mass function is a vector where the first
entry corresponds to the integral over the (0,1] interval of the continuous
distribution, the second entry corresponds to the (0,2] interval, the third
entry corresponds to the (1, 3] interval etc.
}
\keyword{internal}
