% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{discretise}
\alias{discretise}
\title{Discretise a <dist_spec>}
\usage{
discretise(x, silent = TRUE)
}
\arguments{
\item{x}{A \verb{<dist_spec>}}

\item{silent}{Logical; if \code{TRUE} then any distribution that can't be
discretised will be returned as is. If \code{FALSE} then an error will be
thrown.}
}
\value{
A \verb{<dist_spec>} where all distributions with constant parameters are
nonparametric.
}
\description{
Discretise a <dist_spec>
}
\details{
By default it will discretise all the distributions it can discretise
(i.e. those with finite support and constant parameters).
}
\author{
Sebastian Funk
}
