% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_delay.R
\name{dist_fit}
\alias{dist_fit}
\title{Fit an Integer Adjusted Exponential, Gamma or Lognormal distributions}
\usage{
dist_fit(
  values = NULL,
  samples = 1000,
  cores = 1,
  chains = 2,
  dist = "exp",
  verbose = FALSE,
  backend = "rstan"
)
}
\arguments{
\item{values}{Numeric vector of values}

\item{samples}{Numeric, number of samples to take. Must be >= 1000.
Defaults to 1000.}

\item{cores}{Numeric, defaults to 1. Number of CPU cores to use (no effect
if greater than the number of chains).}

\item{chains}{Numeric, defaults to 2. Number of MCMC chains to use. More is
better with the minimum being two.}

\item{dist}{Character string, which distribution to fit. Defaults to
exponential (\code{"exp"}) but gamma (\code{"gamma"}) and lognormal (\code{"lognormal"}) are
also supported.}

\item{verbose}{Logical, defaults to FALSE. Should verbose progress messages
be printed.}

\item{backend}{Character string indicating the backend to use for fitting
stan models. Supported arguments are "rstan" (default) or "cmdstanr".}
}
\value{
A stan fit of an interval censored distribution
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Fits an integer adjusted exponential, gamma or lognormal distribution using
stan.

\strong{This function is deprecated.} Please use \code{\link[=estimate_dist]{estimate_dist()}} instead,
which provides better handling of censoring and truncation.
}
\examples{
\donttest{
# integer adjusted exponential model
dist_fit(rexp(1:100, 2),
  samples = 1000, dist = "exp",
  cores = ifelse(interactive(), 4, 1), verbose = TRUE
)


# integer adjusted gamma model
dist_fit(rgamma(1:100, 5, 5),
  samples = 1000, dist = "gamma",
  cores = ifelse(interactive(), 4, 1), verbose = TRUE
)

# integer adjusted lognormal model
dist_fit(rlnorm(1:100, log(5), 0.2),
  samples = 1000, dist = "lognormal",
  cores = ifelse(interactive(), 4, 1), verbose = TRUE
)
}
}
