% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_fit}
\alias{dist_fit}
\title{Fit an integer adjusted exponential or gamma distribution}
\usage{
dist_fit(values = NULL, samples = NULL, dist = "exp")
}
\arguments{
\item{values}{Numeric vector of values}

\item{samples}{Numeric, number of samples to take}

\item{dist}{Character string, which distribution to fit. Defaults to exponential (\code{"exp"}) but
gamma is also supported (\code{"gamma"}).}
}
\value{
A \code{stan} fit of an interval censored distribution
}
\description{
Fit an integer adjusted exponential or gamma distribution
}
\examples{

dist_fit(rexp(1:100, 2), samples = 1000, dist = "exp")

}
