% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist_prototype.R
\name{.convert_to_pcd_data}
\alias{.convert_to_pcd_data}
\title{Convert data to primarycensored format (internal helper)}
\usage{
.convert_to_pcd_data(data, dist, primary_dist, verbose)
}
\arguments{
\item{data}{Input data (vector or data.frame)}

\item{dist}{Distribution type}

\item{primary_dist}{Primary distribution}

\item{verbose}{Verbose output?}
}
\value{
A list of data for primarycensored Stan model
}
\description{
Convert data to primarycensored format (internal helper)
}
\keyword{internal}
