% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist.R
\name{.extract_to_dist_spec}
\alias{.extract_to_dist_spec}
\title{Extract parameters and convert to dist_spec}
\usage{
.extract_to_dist_spec(fit, dist, max_value, backend)
}
\description{
Extract parameters and convert to dist_spec
}
\keyword{internal}
