% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist_prototype.R
\name{.linelist_to_pcd_data}
\alias{.linelist_to_pcd_data}
\title{Convert linelist to primarycensored data (internal)}
\usage{
.linelist_to_pcd_data(data, primary_dist, verbose)
}
\arguments{
\item{data}{Input data (vector or data.frame)}

\item{primary_dist}{Primary distribution}

\item{verbose}{Verbose output?}
}
\value{
List for primarycensored
}
\description{
Convert linelist to primarycensored data (internal)
}
\keyword{internal}
