% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_dist_prototype.R
\name{.vector_to_pcd_data}
\alias{.vector_to_pcd_data}
\title{Convert vector to primarycensored data (internal)}
\usage{
.vector_to_pcd_data(values, primary_dist, verbose)
}
\arguments{
\item{primary_dist}{Primary distribution}

\item{verbose}{Verbose output?}
}
\value{
List for primarycensored
}
\description{
Convert vector to primarycensored data (internal)
}
\keyword{internal}
