% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiNow2-package.R
\docType{package}
\name{EpiNow2-package}
\alias{EpiNow2}
\alias{EpiNow2-package}
\title{EpiNow2: Estimate Real-Time Case Counts and Time-Varying Epidemiological Parameters}
\description{
Estimates the time-varying reproduction number, rate of spread, and doubling time using a range of open-source tools (Abbott et al. (2020) \doi{10.12688/wellcomeopenres.16006.1}), and current best practices (Gostic et al. (2020) \doi{10.1101/2020.06.18.20134858}). It aims to help users avoid some of the limitations of naive implementations in a framework that is informed by community feedback and is actively supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiforecasts.io/EpiNow2/}
  \item \url{https://epiforecasts.io/EpiNow2/dev/}
  \item \url{https://github.com/epiforecasts/EpiNow2}
  \item Report bugs at \url{https://github.com/epiforecasts/EpiNow2/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID})

Authors:
\itemize{
  \item Sam Abbott \email{sam.abbott@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
  \item Joel Hellewell \email{joel.hellewell@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-2683-0849}{ORCID})
  \item Katharine Sherratt \email{katharine.sherratt@lshtm.ac.uk}
  \item Katelyn Gostic \email{kgostic@uchicago.edu}
  \item Joe Hickson \email{joseph.hickson@metoffice.gov.uk}
  \item Hamada S. Badr \email{badr@jhu.edu} (\href{https://orcid.org/0000-0002-9808-2344}{ORCID})
  \item Michael DeWitt \email{me.dewitt.jr@gmail.com} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID})
  \item James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID})
  \item EpiForecasts
}

Other contributors:
\itemize{
  \item Robin Thompson \email{robin.thompson@lshtm.ac.uk} [contributor]
  \item Sophie Meakin \email{sophie.meaking@lshtm.ac.uk} [contributor]
  \item James Munday \email{james.munday@lshtm.ac.uk} [contributor]
  \item Nikos Bosse [contributor]
  \item Paul Mee \email{paul.mee@lshtm.ac.uk} [contributor]
  \item Peter Ellis \email{peter.ellis2013nz@gmail.com} [contributor]
  \item Pietro Monticone \email{pietro.monticone@edu.unito.it} [contributor]
  \item Lloyd Chapman \email{lloyd.chapman1@lshtm.ac.uk } [contributor]
  \item Andrew Johnson \email{andrew.johnson@arjohnsonau.com} [contributor]
}

}
\keyword{internal}
