% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{epinow2_stan_header}
\alias{epinow2_stan_header}
\title{Path to the C++ header used by the EpiNow2 Stan models}
\usage{
epinow2_stan_header()
}
\value{
A character string with the path to the header.
}
\description{
Models that include \code{inst/stan/functions} need this header, which
implements Stan functions declared there without a body.
}
\examples{
epinow2_stan_header()
}
