% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_infections.R
\name{estimate_infections}
\alias{estimate_infections}
\title{Estimate Infections, the Time-Varying Reproduction Number and the Rate of Growth}
\usage{
estimate_infections(
  reported_cases,
  family = "negbin",
  incubation_period,
  reporting_delay,
  generation_time,
  rt_prior,
  prior_smoothing_window = 7,
  horizon = 14,
  model,
  cores = 1,
  chains = 2,
  samples = 1000,
  warmup = 1000,
  estimate_rt = TRUE,
  adapt_delta = 0.99,
  max_treedepth = 15,
  return_fit = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date).}

\item{family}{A character string indicating the reporting model to use. Defaults to negative
binomial ("negbin") with poisson ("poisson") also supported.}

\item{incubation_period}{A list containing the mean, standard deviation of the mean (mean_sd),
standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
incubation period (assuming a lognormal distribution with all parameters excepting the max allowed value
on the log scale).}

\item{reporting_delay}{A list containing the mean, standard deviation of the mean (mean_sd),
standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
reporting delay (assuming a lognormal distribution with all parameters excepting the max allowed value
on the log scale).}

\item{generation_time}{A list containing the mean, standard deviation of the mean (mean_sd),
standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
generation time (assuming a gamma distribution).}

\item{rt_prior}{A list contain the mean and standard deviation (sd) of the gamma distributed prior for
Rt. By default this is assumed to be mean 1 with a standard deviation of 1.}

\item{prior_smoothing_window}{Numeric defaults to 7. The number of days over which to take a rolling average
for the prior based on reported cases.}

\item{horizon}{Numeric, defaults to 14. Number of days into the future to forecast.}

\item{model}{A compiled stan model. By default uses the internal package model.}

\item{cores}{Numeric, defaults to 2. The number of cores to use when fitting the stan model.}

\item{chains}{Numeric, defaults to 2. The number of MCMC chains to use.}

\item{estimate_rt}{Logical, defaults TRUE. Should Rt be estimated when imputing infections.}

\item{adapt_delta}{Numeric, defaults to 0.99. See ?rstan::sampling.}

\item{max_treedepth}{Numeric, defaults to 15. See ?rstan::sampling.}

\item{return_fit}{Logical, defaults to FALSE. Should the fitted stan model be returned.}

\item{verbose}{Logical, defaults to FALSE. Should verbose progress messages be printed.}
}
\description{
This function uses a non-parametric approach to reconstruct cases by date of infection from reported
cases. It can optionally then estimate the time-varying reproduction number and the rate of growth.
}
\examples{
\dontrun{
## Get example case counts and format
reported_cases <- NCoVUtils::get_ecdc_cases(countries = "Russia")
reported_cases <- NCoVUtils::format_ecdc_data(reported_cases)
reported_cases <- data.table::as.data.table(reported_cases)[, confirm := cases][, cases := NULL][1:90]
 
## Set up example generation time
generation_time <- list(mean = EpiNow2::covid_generation_times[1, ]$mean,
                        mean_sd = EpiNow2::covid_generation_times[1, ]$mean_sd,
                        sd = EpiNow2::covid_generation_times[1, ]$sd,
                        sd_sd = EpiNow2::covid_generation_times[1, ]$sd_sd,
                        max = 30)
## Set                   
incubation_period <- list(mean = EpiNow2::covid_incubation_period[1, ]$mean,
                          mean_sd = EpiNow2::covid_incubation_period[1, ]$mean_sd,
                          sd = EpiNow2::covid_incubation_period[1, ]$sd,
                          sd_sd = EpiNow2::covid_incubation_period[1, ]$sd_sd,
                          max = 30)
                   
reporting_delay <- list(mean = log(5),
                        mean_sd = log(2),
                        sd = log(2),
                        sd_sd = log(1.5),
                        max = 30)
                        
## Run model
out <- EpiNow2::estimate_infections(reported_cases, family = "negbin",
                                    generation_time = generation_time,
                                    incubation_period = incubation_period,
                                    reporting_delay = reporting_delay,
                                    samples = 1000, warmup = 500,
                                    rt_prior = list(mean = 1, sd = 1),
                                    cores = 4, chains = 4,
                                    estimate_rt = TRUE,
                                    verbose = TRUE, return_fit = TRUE)

out   
}                                
}
