% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_infections.R
\name{estimate_infections}
\alias{estimate_infections}
\title{Estimate Infections, the Time-Varying Reproduction Number and the Rate of Growth}
\usage{
estimate_infections(
  reported_cases,
  family = "negbin",
  incubation_period,
  reporting_delay,
  generation_time,
  rt_prior,
  prior_smoothing_window = 7,
  model,
  cores = 1,
  chains = 2,
  samples = 1000,
  warmup = 1000,
  estimate_rt = TRUE,
  adapt_delta = 0.99,
  max_treedepth = 15,
  return_fit = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{chains}{}

\item{adapt_delta}{}
}
\description{
This function uses a non-parametric approach to reconstruct cases by date of infection from reported
cases. It can optionally then estimate the time-varying reproduction number and the rate of growth.
}
\examples{
\dontrun{
## Get example case counts and format
reported_cases <- NCoVUtils::get_ecdc_cases(countries = "Russia")
reported_cases <- NCoVUtils::format_ecdc_data(reported_cases)
reported_cases <- data.table::as.data.table(reported_cases)[, confirm := cases][, cases := NULL][1:90]
 
## Set up example generation time
generation_time <- list(mean = EpiNow2::covid_generation_times[1, ]$mean,
                        mean_sd = EpiNow2::covid_generation_times[1, ]$mean_sd,
                        sd = EpiNow2::covid_generation_times[1, ]$sd,
                        sd_sd = EpiNow2::covid_generation_times[1, ]$sd_sd,
                        max = 30)
## Set                   
incubation_period <- list(mean = EpiNow2::covid_incubation_period[1, ]$mean,
                          mean_sd = EpiNow2::covid_incubation_period[1, ]$mean_sd,
                          sd = EpiNow2::covid_incubation_period[1, ]$sd,
                          sd_sd = EpiNow2::covid_incubation_period[1, ]$sd_sd,
                          max = 30)
                   
reporting_delay <- list(mean = log(5),
                        mean_sd = log(2),
                        sd = log(2),
                        sd_sd = log(1.5),
                        max = 30)
                        
## Run model
out <- EpiNow2::estimate_infections(reported_cases, family = "negbin",
                                    generation_time = generation_time,
                                    incubation_period = incubation_period,
                                    reporting_delay = reporting_delay,
                                    samples = 1000, warmup = 500,
                                    rt_prior = list(mean = 1, sd = 1),
                                    cores = 4, chains = 4,
                                    estimate_rt = TRUE,
                                    verbose = TRUE, return_fit = TRUE)

out   
}                                
}
