% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_confirmed}
\alias{example_confirmed}
\title{Example Confirmed Case Data Set}
\format{
A data frame containing cases reported on each date.
}
\usage{
example_confirmed
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
An example data frame of observed cases
}
\keyword{datasets}
