% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_incubation_period}
\alias{example_incubation_period}
\title{Example incubation period}
\format{
A \code{dist_spec} object summarising the distribution
}
\usage{
example_incubation_period
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
An example of an incubation period estimate. See here for details:
https://github.com/epiforecasts/EpiNow2/blob/main/data-raw/incubation-period.R # nolint
}
\keyword{datasets}
