% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_reporting_delay}
\alias{example_reporting_delay}
\title{Example reporting delay}
\format{
A \code{dist_spec} object summarising the distribution
}
\usage{
example_reporting_delay
}
\description{
An example of an reporting delay estimate. See here for details:
https://github.com/epiforecasts/EpiNow2/blob/main/data-raw/reporting-delay # nolint
}
\keyword{datasets}
