% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_truncated}
\alias{example_truncated}
\title{Example Case Data Set with Truncation}
\format{
A list of \code{data.table}s containing cases reported on each date until
a point of truncation.
Each element of the list is a \code{data.table} with the following columns:
\describe{
\item{date}{Date of case report.}
\item{confirm}{Number of confirmed cases.}
}
}
\usage{
example_truncated
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
An example dataset of observed cases with truncation applied.
This data is generated internally for use in the example of
\code{estimate_truncation()}. For details on how the data is generated, see
\url{https://github.com/epiforecasts/EpiNow2/blob/main/data-raw/estimate-truncation.R} #nolint
}
\keyword{datasets}
