% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expose_stan_fns}
\alias{expose_stan_fns}
\title{Expose internal package stan functions in R}
\usage{
expose_stan_fns(files, target_dir, ...)
}
\arguments{
\item{files}{A character vector indicating the target files.}

\item{target_dir}{A character string indicating the target directory for the
file.}

\item{...}{Additional arguments passed to \code{\link[rstan:expose_stan_functions]{rstan::expose_stan_functions()}}.}
}
\value{
No return value, called for side effects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
his function exposes internal stan functions in R from a user
supplied list of target files. Allows for testing of stan functions in R and
potentially user use in R code.
}
