% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expose_stan_fns}
\alias{expose_stan_fns}
\title{Expose internal package stan functions in R}
\usage{
expose_stan_fns(files, target_dir, ...)
}
\arguments{
\item{files}{A character vector indicating the target files}

\item{target_dir}{A character string indicating the target directory for the file}

\item{...}{Additional arguments passed to \code{rstan::expose_stan_functions}.}
}
\description{
This function exposes internal stan functions in R from a user
supplid list of target files.
}
\examples{
\donttest{
expose_stan_fns("rt.stan", target_dir = system.file("stan/functions", package = "EpiNow2"))

# test by updating Rt
update_Rt(rep(1, 10), log(1.2), rep(0.1, 9), rep(10, 0), numeric(0), 0)
}
}
