% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_parameter_samples}
\alias{extract_parameter_samples}
\title{Extract Parameter Samples from a Stan Model}
\usage{
extract_parameter_samples(stan_fit, data, reported_dates, reported_inf_dates)
}
\arguments{
\item{stan_fit}{A fit Stan model as returned by \code{rstan:sampling}}

\item{data}{A list of the data supplied to the \code{rstan::sampling} call.}

\item{reported_dates}{A vector of dates to report estimates for.}

\item{reported_inf_dates}{A vector of dates to report infection estimates for.}
}
\value{
A list of dataframes each containing the posterior of a parameter
}
\description{
Extract Parameter Samples from a Stan Model
}
