% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract Samples from All Parameters}
\usage{
extract_parameters(samples)
}
\arguments{
\item{samples}{Extracted stan model (using \code{\link[rstan:stanfit-method-extract]{rstan::extract()}})}
}
\value{
A \verb{<data.table>} containing the parameter name, sample id and sample
value, or NULL if parameters don't exist in the samples
}
\description{
Extract Samples from All Parameters
}
\keyword{internal}
