% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract all samples from a (cmdstan) stan model}
\usage{
extract_samples(stan_fit, variables = NULL, drop_length_1 = FALSE)
}
\arguments{
\item{drop_length_1}{Logical; whether the first dimension should be dropped
if it is off length 1; this is necessary when processing simulation results.}
}
\value{
List of data.tables with samples
}
\description{
Extract all samples from a (cmdstan) stan model
}
