% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{extract_single_dist}
\alias{extract_single_dist}
\title{Extract a single element of a composite \verb{<dist_spec>}}
\usage{
extract_single_dist(x, i)
}
\arguments{
\item{x}{A composite \code{dist_spec} object}

\item{i}{The index to extract}
}
\value{
A single \code{dist_spec} object
}
\description{
Extract a single element of a composite \verb{<dist_spec>}
}
\author{
Sebastian Funk
}
\keyword{internal}
