% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fill_missing}
\alias{fill_missing}
\title{Fills missing data in a data set to prepare it for use within the package}
\usage{
fill_missing(
  data,
  missing = c("missing", "accumulate", "zero"),
  na = c("missing", "accumulate", "zero"),
  initial_accumulate,
  obs_column
)
}
\arguments{
\item{data}{Data frame with a \code{date} column. The other columns depend on the
model that the data are to be used, e.g. \code{\link[=estimate_infections]{estimate_infections()}} or
\code{\link[=estimate_secondary]{estimate_secondary()}}. See the documentation there for the expected
format.}

\item{missing}{Character. Options are "missing" (the default), "accumulate"
and "zero". This determines how missing dates the data are interpreted.
If set to "missing", any missing dates in the observation data will be
interpreted as missing and skipped in the likelihood. If set to
"accumulate", modelled observations will be accumulated and added to the
next non-missing data point. This can be used to model incidence data that
is reported at less than daily intervals. If set to "accumulate", the
first data point is not included in the likelihood (unless \code{initial} is
set to a non-zero value) but used only to reset modelled observations to
zero. If "zero" then all observations on missing dates will be assumed to
be of value 0.}

\item{na}{Character. How to process dates that have NA values. The options
available are the same ones as for the \code{missing} argument.}

\item{initial_accumulate}{Integer. The number of initial dates to accumulate
if \code{missing} or a column name is set to \code{"accumulate"}. This number of
dates is added to the beginning of the data set to be accumulated onto the
first data point. This is useful, for example, for modelling weekly
incidence data, in which case this should be set to 7. If accumulating and
the first data point is one to accumulate and this is not set, then that
data point will be removed with a warning.}

\item{obs_column}{Character. If given, only the column specified here will
be used for checking missingness. This is useful if using a data set that
has NA values in multiple columns, but only one of them corresponds to
observations that are to be processed here. By default, all columns will
be processed and, e.g. \code{accumulate} will be set if any column is NA in any
given row.}
}
\value{
a data.table with an \code{accumulate} column that indicates whether
values are accumulated (see the documnetation of the \code{data} arugment in
\code{\link[=estiamte_infections]{estiamte_infections()}})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
cases <- data.table::copy(reported_cases)
## calculate weekly sum
cases[, confirm := frollsum(confirm, 7)]
## limit to dates once a week
cases <- cases[seq(7, nrow(reported_cases_weekly), 7)]
## set the second observation to missing
cases[2, confirm := NA]
## fill missing data
fill_missing(cases, missing = "accumulate", initial_accumulate = 7)
}
