% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{filter_opts}
\alias{filter_opts}
\title{Filter Options for a Target Region}
\usage{
filter_opts(opts, region)
}
\arguments{
\item{opts}{Either a list of calls to an \verb{_opts()} function or a single
call to an \verb{_opts()} function.}

\item{region}{A character string indicating a region of interest.}
}
\value{
A list of options
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
A helper function that allows the selection of region specific settings if
present and otherwise applies the overarching settings.
}
\author{
Sam Abbott
}
