% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_infections.R
\name{fit_model_with_nuts}
\alias{fit_model_with_nuts}
\title{Fit a Stan Model using the NUTs sampler}
\usage{
fit_model_with_nuts(
  args,
  future = FALSE,
  max_execution_time = Inf,
  verbose = FALSE
)
}
\arguments{
\item{args}{List of stan arguments}

\item{future}{Logical, defaults to \code{FALSE}. Should \code{future} be used to run stan chains in parallel.}

\item{max_execution_time}{Numeric, defaults to Inf. What is the maximum execution time per chain in seconds.
Results will still be returned as long as at least 2 chains complete successfully within the timelimit.}

\item{verbose}{Logical, defaults to \code{FALSE}. Should verbose progress information be returned.}
}
\value{
A stan model object
}
\description{
Fit a Stan Model using the NUTs sampler
}
