% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{fix_dist}
\alias{fix_dist}
\title{Fix the parameters of a \code{dist_spec}}
\usage{
fix_dist(x, strategy = c("center", "sample"))
}
\arguments{
\item{x}{A \link{dist_spec} object}

\item{strategy}{Character; either "center" (use the central estimate) or
"sample" (randomly sample mean and standard deviation from
uncertainty given in the \code{dist_spec})}
}
\value{
A \code{dist_spec} object without uncertainty
}
\description{
If the given \code{dist_spec} has any uncertainty, it is removed and the
corresponding distribution converted into a fixed one.
}
\author{
Sebastian Funk
}
