% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_fit}
\alias{format_fit}
\title{Format Posterior Samples}
\usage{
format_fit(posterior_samples, horizon, shift, CrIs)
}
\arguments{
\item{posterior_samples}{A list of posterior samples as returned by
\code{\link[=format_simulation_output]{format_simulation_output()}}.}

\item{horizon}{Numeric, forecast horizon.}

\item{shift}{Numeric, the shift to apply to estimates.}

\item{CrIs}{Numeric vector of credible intervals to calculate.}
}
\value{
A list of samples and summarised posterior parameter estimates.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Summaries posterior samples and adds additional custom variables.
}
\keyword{internal}
