% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{generation_time_opts}
\alias{generation_time_opts}
\title{Generation Time Distribution Options}
\usage{
generation_time_opts(
  mean = 1,
  mean_sd = 0,
  sd = 0,
  sd_sd = 0,
  max = 15,
  fixed = FALSE,
  disease,
  source
)
}
\arguments{
\item{mean}{Numeric, defaults to 1. If the only non-zero summary parameter
then  this is the fixed interval of the generation time. If the \code{sd} is
non-zero then this is the mean of a gamma distribution.}

\item{mean_sd}{Numeric, defaults to 0. The prior uncertainty for the mean
of the generation time.}

\item{sd}{Numeric, defaults to 0. Sets the standard deviation for a gamma
distribution generation time.}

\item{sd_sd}{Numeric, defaults to 0. The prior uncertainty for the standard
deviation of the generation time.}

\item{max}{Numeric, defaults to 15. The maximum number of days to include in
in the generation time.}

\item{fixed}{Logical, defaults to \code{FALSE}. Should reporting delays be treated
as coming from fixed (vs uncertain) distributions. Making this simplification
drastically reduces compute requirements.}

\item{disease}{A character string indicating the disease of interest.}

\item{source}{A character string indicating the source of interest.}
}
\value{
A list summarising the discretised (upper-)truncated gamma delay
distribution with the following parameters:
\itemize{
\item "gt_mean", the mean generation time.
\item "gt_mean_sd", the standard deviation in the estimate of "gt_mean" parameter
(assumed normally distributed).
\item "gt_sd", the standard deviation of the generation time.
\item "gt_sd_sd", the standard deviation of the estimate of the "gt_sd"
parameter (assumed normally distributed).
\item "gt_max", the maximum generation time.
\item "gt_fixed", Binary value indicating if the generation time has uncertainty.
\item "gt_pmf" A vector describing the PMF of the distribution implied by the
mean of the summary parameters.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns generation time parameters in a format for lower level model use.
}
\seealso{
convert_to_logmean convert_to_logsd bootstrapped_dist_fit
}
