% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{generation_time_opts}
\alias{generation_time_opts}
\title{Generation Time Distribution Options}
\usage{
generation_time_opts(
  dist,
  prior_weight = NULL,
  zeroes = c("truncate", "shift", "add_to_1")
)
}
\arguments{
\item{dist}{A delay distribution or series of delay distributions generated
using \code{\link[=dist_spec]{dist_spec()}} or \code{\link[=get_generation_time]{get_generation_time()}}. If no distribution is given
a fixed generation time of 1 will be assumed.}

\item{prior_weight}{numeric, weight given to the generation time prior.
By default the generation time prior will be weighted by the number of
observation data points, usually preventing the posteriors from shifting
much from the given distribution. Another sensible option would be 1,
i.e. treating the generation time distribution as a single parameter.}

\item{zeroes}{character, how to deal with the first element of the
generation time distribution, corresponding to generation time 0 which
is not allowed in the renewal equation framework used here.
One of "truncate" (default; truncate the distribution at 0), "shift" i
(shift the generation time by 1, so that the first entry corresponds to
generation time 1) or "add_to_1" ' (add the probability mass to the
existing probability mass for generation time 1).}
}
\value{
A list summarising the input delay distributions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns generation time parameters in a format for lower level model use. The
generation time can either be given as a \code{disease} and \code{source} to
be passed to \link{get_generation_time}, or as parameters of a distribution to be
passed to \link{dist_spec}.
}
\examples{
# default settings with a fixed generation time of 1
generation_time_opts()

# A fixed gamma distributed generation time
generation_time_opts(dist_spec(mean = 3, sd = 2, max = 15))

# An uncertain gamma distributed generation time
generation_time_opts(dist_spec(mean = 3, sd = 2, mean_sd = 1, sd_sd = 0.5, max = 15))
# A generation time sourced from the ligerature
dist <- get_generation_time(disease = "SARS-CoV-2", source = "ganyani")
generation_time_opts(dist)
}
\seealso{
convert_to_logmean convert_to_logsd bootstrapped_dist_fit dist_spec
}
\author{
Sebastian Funk

Sam Abbott
}
