% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{generation_time_opts}
\alias{generation_time_opts}
\title{Generation Time Distribution Options}
\usage{
generation_time_opts(
  mean = 1,
  mean_sd = 0,
  sd = 0,
  sd_sd = 0,
  max = 15,
  fixed = FALSE,
  disease,
  source
)
}
\arguments{
\item{mean}{Numeric, defaults to 1. If the only non-zero summary parameter
then  this is the fixed interval of the generation time. If the \code{sd} is
non-zero then this is the mean of a gamma distribution.}

\item{mean_sd}{Numeric, defaults to 0. The prior uncertainty for the mean
of the generation time.}

\item{sd}{Numeric, defaults to 0. Sets the standard deviation for a gamma
distribution generation time.}

\item{sd_sd}{Numeric, defaults to 0. The prior uncertainty for the standard
deviation of the generation time.}

\item{...}{Delay distributions as a list with the following parameters:
"mean", "mean_sd", "sd_mean", "sd_sd", and "max" defining a truncated log
normal (with all parameters except for max defined in logged form).}
}
\value{
A list summarising the input delay distributions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns generation time parameters in a format for lower level model use.
}
\examples{
# default settings with a fixed generation time
generation_time_opts()

# A fixed gamma distributed generation time
generation_time_opts(mean = 3, sd = 2)

# An uncertain gamma distributed generation time
generation_time_opts(mean = 3, sd = 2, mean_sd = 1, sd_sd = 0.5)
}
\seealso{
convert_to_logmean convert_to_logsd bootstrapped_dist_fit
}
