% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generation_times}
\alias{generation_times}
\title{Literature Estimates of Generation Times}
\format{
A \code{data.table} of summarising the distribution
}
\usage{
generation_times
}
\description{
Generation time estimates. See
here for details: https://github.com/epiforecasts/EpiNow2/blob/master/data-raw/generation-time.R
}
\keyword{datasets}
