% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_dist}
\alias{get_dist}
\title{Get a Literature Distribution}
\usage{
get_dist(data, disease, source, max_value = 14, fixed = FALSE)
}
\arguments{
\item{data}{A \verb{<data.table>} in the format of \code{generation_times}.}

\item{disease}{A character string indicating the disease of interest.}

\item{source}{A character string indicating the source of interest.}

\item{max_value}{Numeric, the maximum value to allow. Defaults to 14 days.}

\item{fixed}{Logical, defaults to \code{FALSE}. Should distributions be supplied
as fixed values (vs with uncertainty)?}
}
\value{
A list defining a distribution
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been deprecated. Please specify a distribution
using functions such as \code{\link[=Gamma]{Gamma()}} or \code{\link[=LogNormal]{LogNormal()}} instead.
}
\seealso{
\code{\link[=dist_spec]{dist_spec()}}
}
\keyword{internal}
