% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_spec.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters of a parametric distribution}
\usage{
get_parameters(x, id = NULL)
}
\value{
A list of parameters of the distribution.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
dist <- Gamma(shape = 3, rate = 2)
get_parameters(dist)
}
