% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_parameters.epinowfit}
\alias{get_parameters.epinowfit}
\alias{get_parameters.estimate_dist}
\title{Extract parameters from EpiNow2 model fits}
\usage{
\method{get_parameters}{epinowfit}(x, ...)

\method{get_parameters}{estimate_dist}(x, ...)
}
\arguments{
\item{x}{A fitted EpiNow2 model object.}

\item{...}{Not used.}
}
\value{
A named list of parameters.
}
\description{
S3 methods for \code{\link[distspec:get_parameters]{distspec::get_parameters()}} that extract the estimated delay
distribution and scalar parameters from fitted EpiNow2 model objects.
}
