% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_regional_results}
\alias{get_regional_results}
\title{Get Combined Regional Results}
\usage{
get_regional_results(regional_output, results_dir, date, forecast = FALSE)
}
\arguments{
\item{results_dir}{A character string indicating the folder containing the \code{EpiNow2}
results to extract.}

\item{date}{A Character string (in the format "yyyy-mm-dd") indicating the date to extract
data for. Defaults to "latest" which finds the latest results available.}

\item{forecast}{Logical, defaults to \code{FALSE}. Should forecast results be returned.}
}
\value{
A list of estimates, forecasts and estimated cases by date of report.
}
\description{
Get Combined Regional Results
}
\examples{


\dontrun{
# see ?regional_epinow for code to generate regional results to use with this example.

out$summary <- NULL

get_regional_results(out, forecast = TRUE)

}
}
