% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_regions}
\alias{get_regions}
\title{Get Folders with Nowcast Results}
\usage{
get_regions(results_dir)
}
\arguments{
\item{results_dir}{A character string giving the directory in which results
are stored (as produced by \code{regional_rt_pipeline}).}
}
\value{
A named character vector containing the results to plot.
}
\description{
Get Folders with Nowcast Results
}
