% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_seeding_time}
\alias{get_seeding_time}
\title{Estimate seeding time from delays and generation time}
\usage{
get_seeding_time(delays, generation_time)
}
\arguments{
\item{delays}{Delays as specified using \code{\link[=dist_spec]{dist_spec()}}}

\item{generation_time}{Generation time as specified using \code{\link[=dist_spec]{dist_spec()}}}
}
\value{
An integer seeding time
}
\description{
The seeding time is set to the mean of the specified delays, constrained
to be at least the maximum generation time
}
\author{
Sebastian Funk
}
