% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_seeding_time}
\alias{get_seeding_time}
\title{Estimate seeding time from delays and generation time}
\usage{
get_seeding_time(delays, generation_time)
}
\arguments{
\item{delays}{Delays specified using distribution functions such as
\code{\link[=gamma]{gamma()}} or \code{\link[=lognormal]{lognormal()}}}

\item{generation_time}{Generation specified using distribution functions
such as \code{\link[=gamma]{gamma()}} or \code{\link[=lognormal]{lognormal()}}}
}
\value{
An integer seeding time
}
\description{
The seeding time is set to the mean of the specified delays, constrained
to be at least the maximum generation time
}
\author{
Sebastian Funk
}
