% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{gp_ls_range}
\alias{gp_ls_range}
\title{Range of lengthscales the approximate GP represents accurately}
\usage{
gp_ls_range(stan_data)
}
\arguments{
\item{stan_data}{A list of Stan data as returned by \code{\link[=create_gp_data]{create_gp_data()}}.}
}
\value{
A numeric vector giving the shortest and longest lengthscale (in
days) that the boundary factor and number of basis functions can
approximate accurately, or \code{NULL} for the periodic kernel.
}
\description{
Range of lengthscales the approximate GP represents accurately
}
\keyword{internal}
