% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incubation_periods}
\alias{incubation_periods}
\title{Literature Estimates of Incubation Periods}
\format{
A \code{data.table} of summarising the distribution
}
\usage{
incubation_periods
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Incubation period estimates. See here for details:
https://github.com/epiforecasts/EpiNow2/blob/main/data-raw/incubation-period.R # nolint
}
\keyword{datasets}
