% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{is_param_spec}
\alias{is_param_spec}
\title{Test whether an object is a parameter specification}
\usage{
is_param_spec(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical, \code{TRUE} if \code{x} is a \verb{<param_spec>}.
}
\description{
A \verb{<param_spec>} is the common superclass of \verb{<dist_spec>} (a constant or
uncertain value) and \verb{<state_spec>} (a time-varying value created by \code{\link[=GP]{GP()}} or
\code{\link[=RW]{RW()}}). It is the type accepted wherever a parameter's value may be either
constant or time-varying.
}
\keyword{internal}
