% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lapply_func}
\alias{lapply_func}
\title{Choose a parallel or sequential apply function}
\usage{
lapply_func(..., backend = "rstan", future.opts = list())
}
\arguments{
\item{...}{Additional parameters to pass to underlying option functions,
\code{\link[=stan_sampling_opts]{stan_sampling_opts()}} or \code{\link[=stan_vb_opts]{stan_vb_opts()}}, depending on the method}

\item{backend}{Character string indicating the backend to use for fitting
stan models. Supported arguments are "rstan" (default) or "cmdstanr".}
}
\value{
A function that can be used to apply a function to a list
}
\description{
Internal function that chooses an appropriate "apply"-type function (either
\code{\link[=lapply]{lapply()}} or \code{\link[future.apply:future_lapply]{future.apply::future_lapply()}})
}
\keyword{internal}
