% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_conf}
\alias{make_conf}
\title{Format Credible Intervals}
\usage{
make_conf(value, CrI = 90, reverse = FALSE)
}
\arguments{
\item{value}{List of value to map into a string. Requires,
\code{point}, \code{lower}, and \code{upper.}}

\item{CrI}{Numeric, credible interval to report. Defaults to 90.}

\item{reverse}{Logical, defaults to FALSE. Should the reported
credible interval be switched.}
}
\value{
A character vector formatted for reporting
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Combines a list of values into formatted credible intervals.
}
\examples{
value <- list(median = 2, lower_90 = 1, upper_90 = 3)
make_conf(value)
}
\keyword{internal}
