% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_conf}
\alias{make_conf}
\title{Format Credible Intervals}
\usage{
make_conf(value, round_type = NULL, digits = 0)
}
\arguments{
\item{value}{List of value to map into a string. Requires,
\code{point}, \code{lower}, and \code{upper.}}

\item{round_type}{Function, type of rounding to apply. Defaults to \code{round}.}

\item{digits}{Numeric, defaults to 0. Amount of rounding to apply}
}
\value{
A character vector formatted for reporting
}
\description{
Format Credible Intervals
}
\examples{

value <- list(list(point = 1, lower = 0, upper = 3))

make_conf(value, round_type = round, digits = 0)
}
