% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_output_arguments}
\alias{match_output_arguments}
\title{Match User Supplied Arguments with Supported Options}
\usage{
match_output_arguments(
  input_args = NULL,
  supported_args = NULL,
  logger = NULL,
  level = "info"
)
}
\arguments{
\item{input_args}{A character vector of input arguments (can be partial).}

\item{supported_args}{A character vector of supported output arguments.}

\item{logger}{A character vector indicating the logger to target messages
at. Defaults to no logging.}

\item{level}{Character string defaulting to "info". Logging level see
documentation of futile.logger for details. Supported options are "info" and
"debug".}
}
\value{
A logical vector of named output arguments
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Match user supplied arguments with supported options and return a logical
list for internal usage.
}
\keyword{internal}
