% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{mean.dist_spec}
\alias{mean.dist_spec}
\title{Returns the mean of one or more delay distribution}
\usage{
\method{mean}{dist_spec}(x, ...)
}
\arguments{
\item{x}{The \code{\link[=dist_spec]{dist_spec()}} to use}

\item{...}{Not used}
}
\value{
A vector of means.
}
\description{
This works out the mean of all the (parametric / nonparametric) delay distributions
combined in the passed \code{\link[=dist_spec]{dist_spec()}}.
}
\author{
Sebastian Funk
}
