% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{new_state_spec}
\alias{new_state_spec}
\title{Construct a state specification}
\usage{
new_state_spec(type, mean, init, settings = list())
}
\arguments{
\item{type}{Character, the state type (\code{"gp"} or \code{"rw"}).}

\item{mean, init}{A \verb{<dist_spec>}; exactly one must be supplied.}

\item{settings}{A list of additional state settings (e.g. a \verb{<gp_opts>}
object for \code{"gp"}, or the step standard deviation prior for \code{"rw"}).}
}
\value{
A \verb{<state_spec>} object.
}
\description{
Construct a state specification
}
\keyword{internal}
