% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{opts_list}
\alias{opts_list}
\title{Forecast optiong}
\usage{
opts_list(opts, reported_cases, ...)
}
\arguments{
\item{opts}{An \verb{_opts()} function call such as \code{\link[=rt_opts]{rt_opts()}}.}

\item{reported_cases}{A data frame containing a \code{region} variable
indicating the target regions.}

\item{...}{Optional override for region defaults. See the examples
for use case.}
}
\value{
A named list of options per region which can be passed to the \verb{_opt}
accepting arguments of \code{regional_epinow}.
}
\description{
Define a list of \verb{_opts()} to pass to \code{\link[=regional_epinow]{regional_epinow()}} \verb{_opts()} accepting
arguments. This is useful when different settings are needed between regions
within a single \code{\link[=regional_epinow]{regional_epinow()}} call. Using \code{\link[=opts_list]{opts_list()}} the defaults
can be applied to all regions present with an override passed to regions as
necessary (either within \code{\link[=opts_list]{opts_list()}} or externally).
}
\examples{
# uses example case vector
cases <- example_confirmed[1:40]
cases <- rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]
))

# default settings
opts_list(rt_opts(), cases)

# add a weekly random walk in realland
opts_list(rt_opts(), cases, realland = rt_opts(rw = 7))

# add a weekly random walk externally
rt <- opts_list(rt_opts(), cases)
rt$realland$rw <- 7
rt
}
\seealso{
\code{\link[=regional_epinow]{regional_epinow()}} \code{\link[=rt_opts]{rt_opts()}}
}
