% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.epinow}
\alias{plot.epinow}
\title{Plot the Output from epinow}
\usage{
\method{plot}{epinow}(x, type = "summary", ...)
}
\arguments{
\item{x}{A list of output as produced by \code{epinow}}

\item{type}{A character vector indicating the name of plots to return. Defaults
to  "summary" with supported options being "infections", "reports", "R", "growth_rate",
"summary", "all".}

\item{...}{Pass additional arguments to report_plots}
}
\value{
List of plots as produced by \code{report_plots}
}
\description{
\code{plot} method for class "epinow".
}
\seealso{
plot plot.estimate_infections report_plots estimate_infections
}
