% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.estimate_infections}
\alias{plot.estimate_infections}
\alias{plot}
\title{Plot the Output from estimate_infections}
\usage{
\method{plot}{estimate_infections}(x, plots = "summary", ...)
}
\arguments{
\item{x}{A list of output as produced by \code{estimate_infections}}

\item{plots}{A character vector indicating the name of plots to return. Defaults
to  "summary" with supported options being infections, reports, R, growth_rate, summary}

\item{...}{Pass additional arguments to report_plots}
}
\value{
List of plots as produced by \code{report_plots}
}
\description{
\code{plot} method for class "estimate_infections".
}
\seealso{
plot report_plots estimate_infections
}
