% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_truncation.R
\name{plot.estimate_truncation}
\alias{plot.estimate_truncation}
\title{Plot method for estimate_truncation}
\usage{
\method{plot}{estimate_truncation}(x, ...)
}
\arguments{
\item{x}{A list of output as produced by \code{estimate_truncation}}

\item{...}{Pass additional arguments to plot function. Not currently in use.}
}
\value{
\code{ggplot2} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{plot} method for class "estimate_truncation". Returns
a plot faceted over each dataset used in fitting with the latest
observations as columns, the data observed at the time (and so truncated)
as dots and the truncation adjusted estimates as a ribbon.
}
\seealso{
plot estimate_truncation
}
\author{
Sam Abbott
}
