% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_CrIs}
\alias{plot_CrIs}
\title{Plot EpiNow2 Credible Intervals}
\usage{
plot_CrIs(plot, CrIs, alpha, linewidth)
}
\arguments{
\item{plot}{A \code{ggplot2} plot}

\item{CrIs}{Numeric list of credible intervals present in the data. As produced
by \code{extract_CrIs}}

\item{alpha}{Numeric, overall alpha of the target line range}

\item{size}{linewidth, linewidth of the default line range.}
}
\value{
A \code{ggplot2} plot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Adds lineranges for user specified credible intervals
}
