% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Plot Estimates}
\usage{
plot_estimates(
  estimate,
  reported,
  ylab = "Cases",
  hline,
  bar_width = 1.5,
  obs_as_col = TRUE
)
}
\arguments{
\item{estimate}{A data.table of estimates containing the following variables: date, type
(must contain "estimate", "estimate based on partial data" and optionally "forecast"),}

\item{reported}{}

\item{ylab}{}

\item{hline}{}

\item{bar_width}{}

\item{obs_as_col}{}
}
\value{

}
\description{
Plot Estimates
}
\examples{
\dontrun{
# first run the example in ?epinow

plot_estimates(estimate = out$estimates$summarised[variable == "reported_cases_rt"],
               reported <- cases[, confirm := cases], 
               ylab = "Cases")



}


}
