% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot a Summary of the Latest Results}
\usage{
plot_summary(summary_results, x_lab = "Region", log_cases = FALSE, max_cases)
}
\arguments{
\item{summary_results}{A data.table as returned by \code{summarise_results} (the
\code{data} object).}

\item{x_lab}{A character string giving the label for the x axis, defaults to
region.}

\item{log_cases}{Logical, should cases be shown on a logged scale. Defaults
to \code{FALSE}.}

\item{max_cases}{Numeric, no default. The maximum number of cases to plot.}
}
\value{
A \code{ggplot2} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}
Used to return a summary plot across regions (using results generated by
\code{summarise_results}).

May be depreciated in later releases in favour of enhanced S3 methods.
}
