% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{+.dist_spec}
\alias{+.dist_spec}
\title{Combines multiple delay distributions into a new delay distribution}
\usage{
\method{`+`}{dist_spec}(e1, e2)
}
\arguments{
\item{...}{The delay distributions (from calls to \code{\link[=dist_spec]{dist_spec()}}) to combine}
}
\value{
Combined delay distributions (with class \code{\link[=dist_spec]{dist_spec()}}`)
}
\description{
This combines the parameters so that they can
be fed as multiple delay distributions to \code{\link[=epinow]{epinow()}} or \code{\link[=estimate_infections]{estimate_infections()}}.
}
\author{
Sebastian Funk
}
