% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{+.dist_spec}
\alias{+.dist_spec}
\title{Creates a delay distribution as the sum of two other delay distributions}
\usage{
\method{+}{dist_spec}(e1, e2)
}
\arguments{
\item{e1}{The first delay distribution (of type \code{\link[=dist_spec]{dist_spec()}}) to
combine.}

\item{e2}{The second delay distribution (of type \code{\link[=dist_spec]{dist_spec()}}) to
combine.}
}
\value{
A delay distribution representing the sum of the two delays
(with class \code{\link[=dist_spec]{dist_spec()}})
}
\description{
This is done via convolution with \code{\link[stats:convolve]{stats::convolve()}}. Nonparametric delays
that can be combined are processed together, and their cumulative
distribution function is truncated at a specified tolerance level, ensuring
numeric stability.
}
\examples{
# A fixed lognormal distribution with mean 5 and sd 1.
dist1 <- lognormal(
  meanlog = 1.6, sdlog = 1, max = 20
)
dist1 + dist1

# An uncertain gamma distribution with mean 3 and sd 2
dist2 <- gamma(
  mean = normal(3, 0.5), sd = normal(2, 0.5), max = 20
)
dist1 + dist2

# Using tolerance parameter
EpiNow2:::dist_spec_plus(dist1, dist1, tolerance = 0.5)
}
\author{
Sebastian Funk
}
