% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{+.dist_spec}
\alias{+.dist_spec}
\title{Creates a delay distribution as the sum of two other delay distributions}
\usage{
\method{+}{dist_spec}(e1, e2)
}
\arguments{
\item{e1}{The fist delay distributions (from a calls to \code{\link[=dist_spec]{dist_spec()}}) to
combine}

\item{e2}{The second delay distributions (from a calls to \code{\link[=dist_spec]{dist_spec()}})
to combine}
}
\value{
Delay distributions representing the sum of the two delays
(with class \code{\link[=dist_spec]{dist_spec()}}`)
}
\description{
This is done via convolution with \code{stats::convolve()}.
}
\examples{
# A fixed lognormal distribution with mean 5 and sd 1.
lognormal <- dist_spec(
 mean = 5, sd = 1, max = 20, distribution = "lognormal"
)

# An uncertain gamma distribution with mean 3 and sd 2
gamma <- dist_spec(
 mean = 3, sd = 2, mean_sd = 0.5, sd_sd = 0.5, max = 20,
 distribution = "gamma"
)
lognormal + gamma
}
\author{
Sebastian Funk
}
