% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow.R
\name{process_region}
\alias{process_region}
\title{Process regional estimate}
\usage{
process_region(out, return_estimates, target_region, timing)
}
\arguments{
\item{out}{List of output returned by \code{epinow}}

\item{return_estimates}{Logical, defaults to TRUE. Should estimates be returned.}

\item{target_region}{Character string indicating the region being evaluated}

\item{timing}{Output from \code{Sys.time}}
}
\value{
A list of processed output
}
\description{
Process regional estimate
}
