% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow.R
\name{process_regions}
\alias{process_regions}
\title{Process all Region Estimates}
\usage{
process_regions(regional_out, regions)
}
\arguments{
\item{regional_out}{A list of output from multiple runs of \code{regional_epinow}}

\item{regions}{A character vector identifying the regions that have been run}
}
\value{
A list of all regional estimates and successful regional estimates
}
\description{
Process all Region Estimates
}
