% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{reconstruct_nonparametric}
\alias{reconstruct_nonparametric}
\title{Reconstruct a nonparametric delay distribution}
\usage{
reconstruct_nonparametric(stan_data, np_id)
}
\description{
Reconstruct a nonparametric delay distribution
}
\keyword{internal}
