% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{reconstruct_parametric}
\alias{reconstruct_parametric}
\title{Reconstruct a parametric delay distribution}
\usage{
reconstruct_parametric(stan_data, param_id, posterior)
}
\description{
Reconstruct a parametric delay distribution
}
\keyword{internal}
