% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_epinow.R
\name{regional_epinow}
\alias{regional_epinow}
\title{Real-time Rt Estimation, Forecasting and Reporting by Region}
\usage{
regional_epinow(
  reported_cases,
  target_folder = NULL,
  target_date,
  non_zero_points = 2,
  return_output = TRUE,
  output = c("regions", "summary", "samples", "plots", "timings"),
  summary_args = list(),
  ...
)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date), and region (\code{region}).}

\item{target_date}{A character string giving the target date for which to extract results
(in the format "yyyy-mm-dd"). Defaults to latest available estimates.}

\item{non_zero_points}{Numeric, the minimum number of time points with non-zero cases in a region required for
that region to be evaluated. Defaults to 2.}

\item{return_output}{Logical, defaults to TRUE. Should output be returned. This must either be true or a
\code{target_folder} must be specified in order to enable output saving to disk.}

\item{output}{A character vector of optional output to return. Supported options are the individual regional estimates
("regions"),  samples ("samples"),
plots ("plots"), and the stan fit of the underlying model ("fit"). The default is to return samples and plots alongside summarised estimates
and summary statistics. This arugment uses partial matching so for example passing "sam" will lead to samples
being reported.}

\item{summary_args}{A list of arguments passed to \code{regional_summary}. See the \code{regional_summary} documentation for details.}

\item{...}{Pass additional arguments to \code{epinow}. See the documentation for \code{epinow} for details.}
}
\value{
A list of output stratified at the top level into regional output and across region output summary output
}
\description{
Estimates Rt by region. See the documentation for \code{epinow} for further information. The progress of producing
estimates across multiple regions is tracked using the \code{progressr} package. Modify this behaviour using progressr::handlers
and enable it in batch by setting \code{R_PROGRESSR_ENABLE=TRUE} as an environment variable.
}
\examples{
 \donttest{
# construct example distributions
generation_time <- get_generation_time(disease = "SARS-CoV-2", source = "ganyani")
incubation_period <- get_incubation_period(disease = "SARS-CoV-2", source = "lauer")
reporting_delay <- list(mean = log(10), mean_sd = log(2),
                        sd = log(2), sd_sd = log(1.1), max = 30)
                        
# uses example case vector
cases <- EpiNow2::example_confirmed[1:40]
cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))

# setup logging
setup_logging(file = paste0(tempdir(), "/info.log"), 
              name = "EpiNow2.epinow")
                
# run epinow across multiple regions and generate summaries
# samples and warmup have been reduced for this example
out <- regional_epinow(reported_cases = cases, 
                       samples = 100,
                       generation_time = generation_time,
                       delays = list(incubation_period, reporting_delay),
                       stan_args = list(warmup = 100, 
                                        cores = ifelse(interactive(), 4, 1)))
}
}
