% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{regional_summary}
\alias{regional_summary}
\title{Generate Regional Summary Output}
\usage{
regional_summary(
  regional_output,
  results_dir,
  summary_dir,
  target_date,
  region_scale = "Region",
  csv_region_label = "region",
  log_cases = FALSE,
  return_summary = TRUE
)
}
\arguments{
\item{results_dir}{An optional character string indicating the location of the results directory to extract results
from.}

\item{summary_dir}{A character string giving the directory
in which to store summary of results.}

\item{target_date}{A character string giving the target date for which to extract results
(in the format "yyyy-mm-dd"). Defaults to latest available estimates.}

\item{region_scale}{A character string indicating the name to give the regions being summarised.}

\item{csv_region_label}{Character string indicating the label to assign to a region when saving to .csv.}

\item{log_cases}{Logical, should cases be shown on a logged scale. Defaults to \code{FALSE}}

\item{return_summary}{Logical, defaults to \code{TRUE}. Should summary measures be returned.}
}
\value{
A list of summary measures and plots
}
\description{
Generate Regional Summary Output
}
\examples{

\dontrun{
## Requires additional packages:
library(EpiSoon)
library(forecastHybrid)

## Construct example distributions
generation_time <- list(mean = EpiNow2::covid_generation_times[1, ]$mean,
                        mean_sd = EpiNow2::covid_generation_times[1, ]$mean_sd,
                        sd = EpiNow2::covid_generation_times[1, ]$sd,
                        sd_sd = EpiNow2::covid_generation_times[1, ]$sd_sd,
                        max = 30)
                          
incubation_period <- list(mean = EpiNow2::covid_incubation_period[1, ]$mean,
                          mean_sd = EpiNow2::covid_incubation_period[1, ]$mean_sd,
                          sd = EpiNow2::covid_incubation_period[1, ]$sd,
                          sd_sd = EpiNow2::covid_incubation_period[1, ]$sd_sd,
                          max = 30)
                   
reporting_delay <- list(mean = log(10),
                        mean_sd = 0.8,
                        sd = log(2),
                        sd_sd = 0.1,
                        max = 30)
                        
## Uses example case vector from EpiSoon
cases <- EpiNow2::example_confirmed[1:30]

cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))
  
## Run basic nowcasting pipeline
out <- regional_epinow(reported_cases = cases,
                       generation_time = generation_time,
                       incubation_period = incubation_period,
                       reporting_delay = reporting_delay,
                       samples = 1000, warmup = 200, cores = 4, chains = 4,
                       verbose = TRUE)
                       
regional_summary(regional_out,
                 region_scale = "Country")

}

}
