% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{regional_summary}
\alias{regional_summary}
\title{Generate Regional Summary Output}
\usage{
regional_summary(
  results_dir = NULL,
  summary_dir = NULL,
  target_date = NULL,
  region_scale = "Region",
  csv_region_label = "region",
  log_cases = FALSE
)
}
\arguments{
\item{results_dir}{A character string indicating the location of the results directory to extract results
from.}

\item{summary_dir}{A character string giving the directory
in which to store summary of results.}

\item{target_date}{A character string giving the target date for which to extract results
(in the format "yyyy-mm-dd").}

\item{region_scale}{A character string indicating the name to give the regions being summarised.}

\item{csv_region_label}{Character string indicating the label to assign to a region when saving to .csv.}

\item{log_cases}{Logical, should cases be shown on a logged scale. Defaults to \code{FALSE}}
}
\description{
Generate Regional Summary Output
}
\examples{

\dontrun{

## Example asssumes that CovidGlobalNow (github.com/epiforecasts/covid-global) is  
## in the directory above the root.
regional_summary(results_dir = "../covid-global/national",
                 summary_dir = "../covid-global/national-summary",
                 target_date = "2020-03-19",
                 region_scale = "Country")

}

}
