% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_regional_rt.R
\name{report_regional_rt}
\alias{report_regional_rt}
\title{Real-time Rt Estimation, Forecasting and Reporting by Region}
\usage{
report_regional_rt(
  reported_cases,
  target_folder,
  target_date,
  case_limit = 20,
  cores = 1,
  return_estimates = TRUE,
  ...
)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date), and region (\code{region}).}

\item{case_limit}{Numeric, the minimum number of cases in a region required for that region to be evaluated. Defaults to 20.}

\item{cores}{Numeric, defaults to 2. The number of cores to use when fitting the stan model.}

\item{...}{Pass additional arguments to \code{report_rt}}
}
\value{

}
\description{
Estimates Rt by region. See the documentation for \code{report_rt} for further information.
}
\examples{
 \dontrun{
## Requires additional packages:
library(EpiSoon)
library(forecastHybrid)

## Construct example distributions
generation_time <- list(mean = EpiNow2::covid_generation_times[1, ]$mean,
                        mean_sd = EpiNow2::covid_generation_times[1, ]$mean_sd,
                        sd = EpiNow2::covid_generation_times[1, ]$sd,
                        sd_sd = EpiNow2::covid_generation_times[1, ]$sd_sd,
                        max = 30)
                          
incubation_period <- list(mean = EpiNow2::covid_incubation_period[1, ]$mean,
                          mean_sd = EpiNow2::covid_incubation_period[1, ]$mean_sd,
                          sd = EpiNow2::covid_incubation_period[1, ]$sd,
                          sd_sd = EpiNow2::covid_incubation_period[1, ]$sd_sd,
                          max = 30)
                   
reporting_delay <- list(mean = log(10),
                        mean_sd = 0.8,
                        sd = log(2),
                        sd_sd = 0.1,
                        max = 30)
                        
## Uses example case vector from EpiSoon
cases <- data.table::setDT(EpiSoon::example_obs_cases)
cases <- cases[, `:=`(confirm = as.integer(cases))][,
                  cases := NULL]

cases <- data.table::rbindlist(list(
  data.table::copy(cases)[, region := "testland"],
  cases[, region := "realland"]))
  
## Run basic nowcasting pipeline
out <- report_regional_rt(reported_cases = cases,
                          generation_time = generation_time,
                          incubation_period = incubation_period,
                          reporting_delay = reporting_delay,
                          forecast_model = function(y, ...){
                              EpiSoon::forecastHybrid_model(
                                 y = y[max(1, length(y) - 21):length(y)],
                                 model_params = list(models = "aefz", weights = "equal"),
                                 forecast_params = list(PI.combination = "mean"), ...)},
                                 samples = 1000, warmup = 500, cores = 2, chains = 2,
                          verbose = TRUE)
}
}
