% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_summary}
\alias{report_summary}
\title{Provide Summary Statistics for Estimated Infections and Rt}
\usage{
report_summary(
  summarised_estimates,
  rt_samples,
  target_folder = NULL,
  return_numeric = FALSE
)
}
\arguments{
\item{summarised_estimates}{A data.table of summarised estimates containing the following variables:
variable, median, bottom, and top. It should contain the following estimates: R, infections, and r
(rate of growth).}

\item{rt_samples}{A data.table containing Rt samples with the following variables: sample and value.}

\item{target_folder}{Character string specifying where to save results (will create if not present).}

\item{return_numeric}{Should numeric summary information be returned.}
}
\value{
A data.table containing formatted and numeric summary measures
}
\description{
Provide Summary Statistics for Estimated Infections and Rt
}
