% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_summary}
\alias{report_summary}
\title{Provide Summary Statistics on an Rt Pipeline}
\usage{
report_summary(target_folder)
}
\arguments{
\item{target_folder}{Character string indicating the folder into which to save results.
Also used to extract previously generated results.}
}
\description{
Provide Summary Statistics on an Rt Pipeline
}
