% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{rstan_vb_opts}
\alias{rstan_vb_opts}
\title{Rstan Variational Bayes Options}
\usage{
rstan_vb_opts(samples = 1000, trials = 10, iter = 10000, ...)
}
\arguments{
\item{samples}{Numeric, default 1000. Overall number of approximate posterior
samples.}

\item{trials}{Numeric, defaults to 10. Number of attempts to use \code{rstan::vb}
before failing.}

\item{iter}{Numeric, defaulting to 10000. Number of iterations to use in
\code{rtan::vb}.}

\item{...}{Additional parameters to pass to \code{rstan::vb}.}
}
\value{
A list of arguments to pass to \code{rstan::vb}
}
\description{
Defines a list specifying the arguments passed to
\code{rstan::vb}. Custom settings can be supplied which override the defaults.
}
\examples{
rstan_vb_opts(samples = 2000)
}
