% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow.R
\name{run_region}
\alias{run_region}
\title{Run epinow with Regional Processing Code}
\usage{
run_region(
  target_region,
  reported_cases,
  target_folder,
  target_date,
  return_estimates,
  return_partial_estimates,
  complete_logger,
  progress_fn,
  ...
)
}
\arguments{
\item{target_region}{Character string indicating the region being evaluated}

\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date), and region (\code{region}).}

\item{target_date}{A character string giving the target date for which to extract results
(in the format "yyyy-mm-dd"). Defaults to latest available estimates.}

\item{return_partial_estimates}{Logical, default to \code{FALSE}. Should estimates required for
\code{regional_summary} be returned.}

\item{complete_logger}{Character string indicating the logger to output
the completion of estimation to.}

\item{progress_fn}{Function as returned by \code{progressr::progressor}. Allows the use of a
progress bar.}

\item{...}{Pass additional arguments to \code{epinow}}
}
\value{
A list of processed output as produced by \code{process_region}
}
\description{
Run epinow with Regional Processing Code
}
