% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow.R
\name{run_region}
\alias{run_region}
\title{Run epinow with Regional Processing Code}
\usage{
run_region(
  target_region,
  reported_cases,
  target_folder,
  target_date,
  return_estimates,
  ...
)
}
\arguments{
\item{target_region}{Character string indicating the region being evaluated}

\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date), and region (\code{region}).}

\item{target_folder}{Character string specifying where to save results (will create if not present).}

\item{target_date}{Date, defaults to maximum found in the data if not specified.}

\item{return_estimates}{Logical, defaults to TRUE. Should estimates be returned.}

\item{...}{Pass additional arguments to \code{epinow}}
}
\value{
A list of processed output as produced by \code{process_region}
}
\description{
Run epinow with Regional Processing Code
}
