% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_epinow.R
\name{run_region}
\alias{run_region}
\title{Run epinow with Regional Processing Code}
\usage{
run_region(
  target_region,
  generation_time,
  delays,
  truncation,
  rt,
  backcalc,
  gp,
  obs,
  stan,
  horizon,
  CrIs,
  reported_cases,
  target_folder,
  target_date,
  return_output,
  output,
  complete_logger,
  verbose,
  progress_fn,
  ...
)
}
\arguments{
\item{target_region}{Character string indicating the region being evaluated}

\item{generation_time}{A call to \code{\link[=generation_time_opts]{generation_time_opts()}} defining the
generation time distribution used. For backwards compatibility a list of
summary parameters can also be passed.}

\item{delays}{A call to \code{\link[=delay_opts]{delay_opts()}} defining delay distributions and
options. See the documentation of \code{\link[=delay_opts]{delay_opts()}} and the examples below for
details.}

\item{truncation}{A call to \code{\link[=trunc_opts]{trunc_opts()}} defining the truncation of
the observed data. Defaults to \code{\link[=trunc_opts]{trunc_opts()}}. Truncation can also be
estimated from data. See \code{\link[=estimate_truncation]{estimate_truncation()}} for an approach. If using
\code{\link[=estimate_truncation]{estimate_truncation()}}, the \code{dist} list element in the output should be
passed to this \code{truncation} argument here. This will apply the reporting
delay to your data using the estimated posterior (\code{dist}) as a prior, thereby
propagating the uncertainty associated with data truncation into the
estimates. See an example workflow in the \code{\link[=estimate_truncation]{estimate_truncation()}} help file.}

\item{rt}{A list of options as generated by \code{\link[=rt_opts]{rt_opts()}} defining Rt
estimation. Defaults to \code{\link[=rt_opts]{rt_opts()}}. Set to \code{NULL} to switch to using back
calculation rather than generating infections using Rt.}

\item{backcalc}{A list of options as generated by \code{\link[=backcalc_opts]{backcalc_opts()}} to
define the back calculation. Defaults to \code{\link[=backcalc_opts]{backcalc_opts()}}.}

\item{gp}{A list of options as generated by \code{\link[=gp_opts]{gp_opts()}} to define the
Gaussian process. Defaults to \code{\link[=gp_opts]{gp_opts()}}. Set to \code{NULL} to disable the
Gaussian process.}

\item{obs}{A list of options as generated by \code{\link[=obs_opts]{obs_opts()}} defining the
observation model. Defaults to \code{\link[=obs_opts]{obs_opts()}}.}

\item{stan}{A list of stan options as generated by \code{\link[=stan_opts]{stan_opts()}}. Defaults
to \code{\link[=stan_opts]{stan_opts()}}. Can be used to override \code{data}, \code{init}, and \code{verbose}
settings if desired.}

\item{horizon}{Numeric, defaults to 7. Number of days into the future to
forecast.}

\item{CrIs}{Numeric vector of credible intervals to calculate.}

\item{reported_cases}{A \verb{<data.frame>} of confirmed cases (confirm) by date
(date), and region (\code{region}).}

\item{target_folder}{Character string specifying where to save results (will
create if not present).}

\item{target_date}{Date, defaults to maximum found in the data if not
specified.}

\item{return_output}{Logical, defaults to FALSE. Should output be returned,
this automatically updates to TRUE if no directory for saving is specified.}

\item{output}{A character vector of optional output to return. Supported
options are the individual regional estimates ("regions"),  samples
("samples"), plots ("plots"), copying the individual region dated folder into
a latest folder (if \code{target_folder} is not null, set using "latest"), the
stan fit of the underlying model ("fit"), and an overall summary across
regions ("summary"). The default is to return samples and plots alongside
summarised estimates and summary statistics. If \code{target_folder} is not NULL
then the default is also to copy all results into a latest folder.}

\item{complete_logger}{Character string indicating the logger to output
the completion of estimation to.}

\item{verbose}{Logical defaults to FALSE. Outputs verbose progress messages
to the console from \code{\link[=epinow]{epinow()}}.}

\item{progress_fn}{Function as returned by \code{\link[progressr:progressor]{progressr::progressor()}}. Allows
the use of a  progress bar.}

\item{...}{Pass additional arguments to \code{\link[=epinow]{epinow()}}. See the documentation for
\code{\link[=epinow]{epinow()}} for details.}
}
\value{
A list of processed output as produced by \code{\link[=process_region]{process_region()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
Internal function that handles calling \code{\link[=epinow]{epinow()}}. Future work will extend
this function to better handle stan logs and allow the user to modify
settings between regions.
}
\seealso{
\code{\link[=regional_epinow]{regional_epinow()}}
}
