% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{save_estimate_infections}
\alias{save_estimate_infections}
\title{Save Estimated Infections}
\usage{
save_estimate_infections(
  estimates,
  target_folder = NULL,
  samples = TRUE,
  return_fit = TRUE
)
}
\arguments{
\item{estimates}{List of data frames as output by \code{estimate_infections}}

\item{target_folder}{Character string specifying where to save results (will
create if not present).}

\item{samples}{Logical, defaults to TRUE. Should samples be saved}

\item{return_fit}{Logical, defaults to TRUE. Should the fit stan object
be returned.}
}
\value{
No return value, called for side effects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Saves output from \code{estimate_infections} to a target directory.
}
\seealso{
estimate_infections
}
\keyword{internal}
