% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{save_input}
\alias{save_input}
\title{Save Observed Data}
\usage{
save_input(data, target_folder)
}
\arguments{
\item{data}{A \verb{<data.frame>} of confirmed cases (confirm) by date (date).
\code{confirm} must be numeric and \code{date} must be in date format. Optionally
this can also have a logical \code{accumulate} column which indicates whether
data should be added to the next data point. This is useful when modelling
e.g. weekly incidence data. See also the \code{\link[=fill_missing]{fill_missing()}} function which
helps add the \code{accumulate} column with the desired properties when dealing
with non-daily data. If any accumulation is done this happens after
truncation as specified by the \code{truncation} argument.}

\item{target_folder}{Character string specifying where to save results (will
create if not present).}
}
\value{
No return value, called for side effects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Saves observed data to a target location if given.
}
\keyword{internal}
