% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{save_input}
\alias{save_input}
\title{Save Observed Data}
\usage{
save_input(reported_cases, target_folder)
}
\arguments{
\item{reported_cases}{}

\item{target_folder}{Character string specifying where to save results (will create if not present).}
}
\description{
Save Observed Data
}
