% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{sd.dist_spec}
\alias{sd.dist_spec}
\title{Returns the standard deviation of one or more delay distribution}
\usage{
\method{sd}{dist_spec}(x, ...)
}
\arguments{
\item{x}{The \code{\link[=dist_spec]{dist_spec()}} to use}

\item{...}{Not used}
}
\value{
A vector of standard deviations.
}
\description{
This works out the standard deviation of all the (parametric /
nonparametric) delay distributions combined in the passed \code{\link[=dist_spec]{dist_spec()}}.
}
\examples{
# A fixed lognormal distribution with sd 5 and sd 1.
dist1 <- lognormal(mean = 5, sd = 1, max = 20)
sd(dist1)

# A gamma distribution with mean 3 and sd 2
dist2 <- gamma(mean = 3, sd = 2)
sd(dist2)

# The sd of the sum of two distributions
sd(dist1 + dist2)
}
\author{
Sebastian Funk
}
