% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_default_logging}
\alias{setup_default_logging}
\title{Setup Default Logging}
\usage{
setup_default_logging(
  logs = tempdir(),
  mirror_epinow = FALSE,
  target_date = NULL
)
}
\arguments{
\item{logs}{Character path indicating the target folder in which to store log
information. Defaults to the temporary directory if not specified. Default logging
can be disabled if \code{logs} is set to NULL. If specifying a custom logging setup then
the code for \code{setup_default_logging} and the \code{setup_logging} function are a sensible
place to start.}

\item{mirror_epinow}{Logical, defaults to FALSE. Should internal logging be
returned from \code{epinow} to the console.}

\item{target_date}{Date, defaults to maximum found in the data if not specified.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#questioning'><img src='figures/lifecycle-questioning.svg' alt='Questioning lifecycle'></a>}}{\strong{Questioning}}
Sets up default logging. Usage of logging is currently being explored as the current setup
cannot log stan errors or progress.
}
\examples{
setup_default_logging()
}
