% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_dt}
\alias{setup_dt}
\title{Convert to Data Table}
\usage{
setup_dt(data)
}
\arguments{
\item{data}{A \verb{<data.frame>} of disease reports (confirm) by date (date).
\code{confirm} must be numeric and \code{date} must be in date format. Optionally
this can also have a logical \code{accumulate} column which indicates whether
data should be added to the next data point. This is useful when modelling
e.g. weekly incidence data. See also the \code{\link[=fill_missing]{fill_missing()}} function which
helps add the \code{accumulate} column with the desired properties when dealing
with non-daily data. If any accumulation is done this happens after
truncation as specified by the \code{truncation} argument.}
}
\value{
A data table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Convenience function that sets the number of \code{{data.table}} cores to 1 and
maps input to be a \code{{data.table}}
}
\keyword{internal}
