% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{setup_dt}
\alias{setup_dt}
\title{Convert to Data Table}
\usage{
setup_dt(reported_cases)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date). confirm must be integer and date must be
in date format.}
}
\value{
A data table
}
\description{
Convert to Data Table
}
