% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_dt}
\alias{setup_dt}
\title{Convert to Data Table}
\usage{
setup_dt(reported_cases)
}
\arguments{
\item{reported_cases}{A data frame of confirmed cases (confirm) by date (date). confirm must be integer and date must be
in date format.}
}
\value{
A data table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Convenience function that sets the number of \code{data.table} cores to 1 and
maps input to be a \code{data.table}
}
