% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setup_forecast}
\alias{setup_forecast}
\title{Ensure forecast is properly set up}
\usage{
setup_forecast(forecast, horizon = NULL)
}
\arguments{
\item{forecast}{A list of options as generated by \code{\link[=forecast_opts]{forecast_opts()}} defining
the forecast opitions. Defaults to \code{\link[=forecast_opts]{forecast_opts()}}. If NULL then no
forecasting will be done.}

\item{horizon}{Deprecated; use \code{forecast} instead to specify the predictive
horizon}
}
\description{
Ensure forecast is properly set up
}
\keyword{internal}
