% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_infections.R
\name{simulate_infections}
\alias{simulate_infections}
\title{Deprecated; use \code{\link[=forecast_infections]{forecast_infections()}} instead}
\usage{
simulate_infections(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link[=forecast_infections]{forecast_infections()}}}
}
\value{
the result of \code{\link[=forecast_infections]{forecast_infections()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Calling this function passes all arguments to \code{\link[=forecast_infections]{forecast_infections()}}
}
