% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_secondary.R
\name{simulate_secondary}
\alias{simulate_secondary}
\title{Simulate infections using the renewal equation}
\usage{
simulate_secondary(
  primary,
  day_of_week_effect = NULL,
  secondary = secondary_opts(),
  delays = delay_opts(),
  truncation = trunc_opts(),
  obs = obs_opts(),
  CrIs = c(0.2, 0.5, 0.9),
  backend = "rstan",
  ...
)
}
\arguments{
\item{primary}{a data frame of primary reports (column \code{primary}) by date
(column \code{date}). Column \code{primary} must be numeric and \code{date} must be in
date format.  it will be assumed that \code{primary} is zero on the missing
days.}

\item{day_of_week_effect}{either \code{NULL} (no day of the week effect) or a
numerical vector of length specified in \code{\link[=obs_opts]{obs_opts()}} as \code{week_length}
(default: 7) if \code{week_effect} is set to TRUE. Each element of the vector
gives the weight given to reporting on this day (normalised to 1).
The default is \code{NULL}.}

\item{secondary}{A call to \code{\link[=secondary_opts]{secondary_opts()}} or a list containing the
following  binary variables: cumulative, historic, primary_hist_additive,
current, primary_current_additive. These parameters control the structure of
the secondary model, see \code{\link[=secondary_opts]{secondary_opts()}} for details.}

\item{delays}{A call to \code{\link[=delay_opts]{delay_opts()}} defining delay distributions and
options. See the documentation of \code{\link[=delay_opts]{delay_opts()}} and the examples below for
details.}

\item{truncation}{A call to \code{\link[=trunc_opts]{trunc_opts()}} defining the truncation of
observed data.  Defaults to \code{\link[=trunc_opts]{trunc_opts()}}. See \code{\link[=estimate_truncation]{estimate_truncation()}} for
an approach to estimating truncation from data.}

\item{obs}{A list of options as generated by \code{\link[=obs_opts]{obs_opts()}} defining the
observation model. Defaults to \code{\link[=obs_opts]{obs_opts()}}.}

\item{CrIs}{Numeric vector of credible intervals to calculate.}

\item{backend}{Character string indicating the backend to use for fitting
stan models. Supported arguments are "rstan" (default) or "cmdstanr".}

\item{...}{deprecated; only included for backward compatibility}
}
\value{
A data.table of simulated secondary observations (column \code{secondary})
by date.
}
\description{
Simulations are done from given initial infections and, potentially
time-varying, reproduction numbers. Delays and parameters of the observation
model can be specified using the same options as in \code{\link[=estimate_infections]{estimate_infections()}}.
}
\details{
In order to simulate, all parameters that are specified such as the mean and
standard deviation of delays or observation scaling, must be fixed.
Uncertain parameters are not allowed.

A previous function called \code{\link[=simulate_infections]{simulate_infections()}} that simulates from a
given model fit has been renamed \code{\link[=forecast_infections]{forecast_infections()}}. Using
\code{\link[=simulate_infections]{simulate_infections()}} with existing estimates is now deprecated. This
option will be removed in version 2.1.0.
}
\examples{
\donttest{
  cases <- as.data.table(example_confirmed)[, primary := confirm]
  sim <- simulate_secondary(
    cases,
    delays = delay_opts(fix_dist(example_reporting_delay)),
    obs = obs_opts(family = "poisson")
  )
}
}
\author{
Sebastian Funk
}
