% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{stan_laplace_opts}
\alias{stan_laplace_opts}
\title{Stan Laplace algorithm Options}
\usage{
stan_laplace_opts(backend = "cmdstanr", ...)
}
\arguments{
\item{backend}{Character string indicating the backend to use for fitting
stan models. Supported arguments are "rstan" (default) or "cmdstanr".}

\item{...}{Additional parameters to pass to \code{\link[cmdstanr:model-method-laplace]{cmdstanr::laplace()}}.}
}
\value{
A list of arguments to pass to \code{\link[cmdstanr:model-method-laplace]{cmdstanr::laplace()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} Defines a list specifying the
arguments passed to \code{\link[cmdstanr:model-method-laplace]{cmdstanr::laplace()}}.
}
\examples{
stan_laplace_opts()
}
