% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{stan_opts}
\alias{stan_opts}
\title{Stan Options}
\usage{
stan_opts(samples = 1000, backend = "rstan", ...)
}
\arguments{
\item{samples}{Numeric, default 1000. Overall number of posterior samples.
When using multiple chains iterations per chain is samples / chains.}

\item{...}{Additional parameters to pass  underlying option functions.}

\item{object}{Stan model object. By default uses the compiled package default.}

\item{method}{A character string, defaulting to sampling. Currently supports
\code{rstan::sampling} ("sampling") or \code{rstan:vb} ("vb").}
}
\value{
A list of arguments to pass to the appropriate rstan functions.
}
\description{
Defines a list specifying the arguments passed to underlying stan
backend functions via \code{rstan_sampling_opts} and \code{rstan_vb_opts}. Custom settings
can be supplied which override the defaults.
}
\examples{
# using default of rstan::sampling
stan_opts(samples = 2000)

# using vb
stan_opts(method = "vb")
}
\seealso{
rstan_opts
}
