% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{stan_opts}
\alias{stan_opts}
\title{Stan Options}
\usage{
stan_opts(samples = 2000, backend = "rstan", return_fit = TRUE, ...)
}
\arguments{
\item{samples}{Numeric, default 2000. Overall number of posterior samples.
When using multiple chains iterations per chain is samples / chains.}

\item{backend}{Character string indicating the backend to use for fitting stan models.
Currently only "rstan" is supported.}

\item{return_fit}{Logical, defaults to TRUE. Should the fit stan model be returned.}

\item{...}{Additional parameters to pass  underlying option functions.}
}
\value{
A list of arguments to pass to the appropriate rstan functions.
}
\description{
\lifecycle{stable}
Defines a list specifying the arguments passed to underlying stan
backend functions via \code{rstan_sampling_opts} and \code{rstan_vb_opts}. Custom settings
can be supplied which override the defaults.
}
\examples{
# using default of rstan::sampling
stan_opts(samples = 1000)

# using vb
stan_opts(method = "vb")
}
\seealso{
rstan_opts
}
