% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{stan_pathfinder_opts}
\alias{stan_pathfinder_opts}
\title{Stan pathfinder algorithm Options}
\usage{
stan_pathfinder_opts(backend = "cmdstanr", samples = 2000, trials = 10, ...)
}
\arguments{
\item{backend}{Character string indicating the backend to use for fitting
stan models. Supported arguments are "rstan" (default) or "cmdstanr".}

\item{samples}{Numeric, default 2000. Overall number of posterior samples.
When using multiple chains iterations per chain is samples / chains.}

\item{trials}{Numeric, defaults to 10. Number of attempts to use
rstan::vb()] before failing.}

\item{...}{Additional parameters to pass to \code{\link[cmdstanr:model-method-laplace]{cmdstanr::laplace()}}.}
}
\value{
A list of arguments to pass to \code{\link[cmdstanr:model-method-laplace]{cmdstanr::laplace()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Defines a list specifying the arguments passed to \code{\link[cmdstanr:model-method-laplace]{cmdstanr::laplace()}}.
}
\examples{
stan_laplace_opts()
}
