% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_key_measures}
\alias{summarise_key_measures}
\title{Summarise rt and cases}
\usage{
summarise_key_measures(results_dir, summary_dir, type = "country", date)
}
\arguments{
\item{results_dir}{Character string indicating the directory from which to extract results.}

\item{summary_dir}{Character string the directory into which to save results as a csv.}

\item{type}{Character string, the region identifier to apply (defaults to region).}

\item{date}{A Character string (in the format "yyyy-mm-dd") indicating the date to extract
data for. Defaults to "latest" which finds the latest results available.}
}
\value{
A list of summarised Rt and cases by date of infection
}
\description{
Summarise rt and cases
}
