% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise Realtime Results}
\usage{
summarise_results(
  regions = NULL,
  results_dir = "results",
  target_date = NULL,
  region_scale = "Region"
)
}
\arguments{
\item{regions}{A character string containing the list of regions to extract results for
(must all have results for the same target date).}

\item{results_dir}{A character string indicating the location of the results directory to extract results
from.}

\item{target_date}{A character string indicating the target date to extract results for. All regions must have results
for this date.}

\item{region_scale}{A character string indicating the name to give the regions being summarised.}
}
\value{
A list of summary data
}
\description{
Summarise Realtime Results
}
