% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary.epinow}
\alias{summary.epinow}
\alias{summary}
\title{Summary output from epinow}
\usage{
\method{summary}{epinow}(
  object,
  output = c("estimates", "forecast", "estimated_reported_cases"),
  type = c("snapshot", "parameters"),
  target_date = NULL,
  params = NULL,
  CrIs = c(0.2, 0.5, 0.9),
  ...
)
}
\arguments{
\item{object}{A list of output as produced by "epinow".}

\item{output}{A character string of output to summarise. Defaults to
"estimates" but also supports "forecast", and "estimated_reported_cases".}

\item{type}{A character vector of data types to return. Defaults to
"snapshot" but also supports "parameters". "snapshot" returns
a summary at a given date (by default the latest date informed by data).
"parameters" returns summarised parameter estimates that can be further
filtered using \code{params} to show just the parameters of interest and date.
This argument is only used when \code{output = "estimates"}.}

\item{target_date}{Date, defaults to maximum found in the data if not
specified.}

\item{params}{A character vector of parameters to filter for.}

\item{CrIs}{Numeric vector of credible intervals to calculate.}

\item{...}{Pass additional summary arguments to lower level methods}
}
\value{
Returns a \verb{<data.frame>} of summary output
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
\code{summary} method for class "epinow".
}
\seealso{
\code{\link[=summary.estimate_infections]{summary.estimate_infections()}} \code{\link[=epinow]{epinow()}}
}
