% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary.epinow}
\alias{summary.epinow}
\alias{summary}
\title{Summarised Output from Epinow}
\usage{
\method{summary}{epinow}(object, type = "summary")
}
\arguments{
\item{object}{A list of output as produced by "epinow".}

\item{type}{A character vector of data types to return. Defaults to "summary"
but also supports "estimates", "forecast", "estimated_reported_cases", and "all".}
}
\value{
Returns a list of summary output
}
\description{
\code{summary} method for class "epinow".
}
\seealso{
summary epinow report_summary
}
