% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary.estimate_infections}
\alias{summary.estimate_infections}
\title{Summary output from estimate_infections}
\usage{
\method{summary}{estimate_infections}(
  object,
  type = c("snapshot", "parameters", "samples"),
  date = NULL,
  params = NULL,
  ...
)
}
\arguments{
\item{object}{A list of output as produced by "estimate_infections".}

\item{type}{A character vector of data types to return. Defaults to
"snapshot" but also supports "parameters", and "samples". "snapshot" return
a summary at a given date (by default the latest date informed by data).
"parameters" returns summarised parameter estimates that can be further
filtered using \code{params} to show just the parameters of interest and date.
"samples" similarly returns posterior
samples.}

\item{date}{A date in the form "yyyy-mm-dd" to inspect estimates for.}

\item{params}{A character vector of parameters to filter for.}

\item{...}{Pass additional arguments to \code{report_summary}}
}
\value{
Returns a \verb{<data.frame>} of summary output
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
\code{summary} method for class "estimate_infections".
}
\seealso{
summary estimate_infections report_summary
}
