% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary.estimate_truncation}
\alias{summary.estimate_truncation}
\title{Summary method for estimate_truncation}
\usage{
\method{summary}{estimate_truncation}(object, ...)
}
\arguments{
\item{object}{An \verb{<estimate_truncation>} object as produced by
\code{\link[=estimate_truncation]{estimate_truncation()}}.}

\item{...}{Ignored}
}
\value{
A \verb{<dist_spec>} object containing the estimated truncation
distribution, which can be passed to \code{\link[=trunc_opts]{trunc_opts()}} and used in
\code{\link[=estimate_infections]{estimate_infections()}}, \code{\link[=epinow]{epinow()}}, or \code{\link[=regional_epinow]{regional_epinow()}}.
}
\description{
Returns the estimated truncation distribution from an
\code{\link[=estimate_truncation]{estimate_truncation()}} fit.
}
