% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{trunc_opts}
\alias{trunc_opts}
\title{Truncation Distribution Options}
\usage{
trunc_opts(dist = NULL)
}
\arguments{
\item{dist}{A list defining the truncation distribution as parameters of a
discretised (upper-)truncated lognormal density distribution; defaults to
\code{NULL} in which case no truncation is used. Otherwise it defines the
truncation distributions as a list with the following parameters:
"mean", the mu parameter or mean of the natural logarithm of the truncation;
"mean_sd", the standard deviation in the estimate of "mean" parameter
(assumed normally distributed); "sd", the sigma parameter or standard
deviation of the natural logarithm of the truncation; "sd_sd", the standard
deviation of the estimate of the "sd" parameter (assumed normally
distributed) sd_sd"; and "max", the maximum truncation.
The "mean" and "sd" and "max" parameters are mandatory;
the "mean_sd" and "sd_sd"
parameters should be provided to represent
uncertainty in the parameter values of the delay but are optional.}
}
\value{
A list summarising the input truncation distribution.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns a truncation distribution formatted for usage by downstream functions. See
\code{estimate_truncation} for an approach to estimate this distribution.
}
\examples{
# no truncation
trunc_opts()
}
\seealso{
convert_to_logmean convert_to_logsd bootstrapped_dist_fit
}
