% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{trunc_opts}
\alias{trunc_opts}
\title{Truncation Distribution Options}
\usage{
trunc_opts(dist = list())
}
\arguments{
\item{dist}{Parameters of a discretised (upper-)truncated lognormal
truncation distribution as a list with parameters that are all passed to
\link{delay_dist}.}
}
\value{
A list summarising the input truncation distribution.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Returns a truncation distribution formatted for usage by
downstream functions. See \code{estimate_truncation()} for an approach to
estimate this distribution.
}
\examples{
# no truncation
trunc_opts()

# truncation dist
trunc_opts(dist = list(mean = 3, sd = 2))
}
\seealso{
convert_to_logmean convert_to_logsd bootstrapped_dist_fit delay_dist
}
