% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{trunc_opts}
\alias{trunc_opts}
\title{Truncation Distribution Options}
\usage{
trunc_opts(dist = NULL)
}
\arguments{
\item{dist}{A list defining the truncation distribution, defaults to \code{NULL} in which
case no truncation is used. Must have the following elements if defined: "mean", "mean_sd",
"sd_mean", "sd_sd", and "max" defining a truncated log normal (with all parameters except
for max defined in logged form).}
}
\value{
A list summarising the input truncation distribution.
}
\description{
\lifecycle{stable}
Returns a truncation distribution formatted for usage by downstream functions. See
\code{estimate_truncation} for an approach to estimate this distribution.
}
\examples{
# no truncation
trunc_opts()
}
\seealso{
convert_to_logmean convert_to_logsd bootstrapped_dist_fit
}
