% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{tune_inv_gamma}
\alias{tune_inv_gamma}
\title{Tune an Inverse Gamma to Achieve the Target Truncation}
\usage{
tune_inv_gamma(lower = 2, upper = 21)
}
\arguments{
\item{lower}{Numeric, defaults to 2. Lower truncation bound.}

\item{upper}{Numeric, defaults to 21. Upper truncation bound.}
}
\value{
A list of alpha and beta values that describe a inverse gamma
distribution that achieves the target truncation.
}
\description{
Tune an Inverse Gamma to Achieve the Target Truncation
}
\examples{

tune_inv_gamma(lower = 2, upper = 21)
}
