% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinow-internal.R
\name{update_horizon}
\alias{update_horizon}
\title{Updates Forecast Horizon Based on Input Data and Target}
\usage{
update_horizon(horizon, target_date, data)
}
\arguments{
\item{horizon}{Numeric, defaults to 7. Number of days into the future to
forecast.}

\item{target_date}{Date, defaults to maximum found in the data if not
specified.}

\item{data}{A \verb{<data.frame>} of confirmed cases (confirm) by date
(date). \code{confirm} must be numeric and \code{date} must be in date format.}
}
\value{
Numeric forecast horizon adjusted for the users intention
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Makes sure that a forecast is returned for the user specified time period
beyond the target date.
}
