% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{update_list}
\alias{update_list}
\title{Update a List}
\usage{
update_list(defaults = list(), optional = list())
}
\arguments{
\item{defaults}{A list of default settings}

\item{optional}{A list of optional settings to override defaults}
}
\value{
A list
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
Used to handle updating settings in a list. For example when making
changes to \code{opts_list} output.
}
